% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acres.R
\docType{data}
\name{acres}
\alias{acres}
\title{Tampa Bay intertidal and supratidal land use and cover}
\format{
A data frame
}
\usage{
acres
}
\description{
Tampa Bay intertidal and supratidal land use and cover for Habitat Master Plan reporting
}
\examples{
\dontrun{
load(url("https://github.com/tbep-tech/hmpu-workflow/raw/master/data/acres.RData"))

save(acres, file = 'data/acres.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
