% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_attain.R
\name{anlz_attain}
\alias{anlz_attain}
\title{Get attainment categories}
\usage{
anlz_attain(avedat)
}
\arguments{
\item{avedat}{result returned from \code{\link{anlz_avedat}}}
}
\value{
a \code{data.frame} for each year and bay segment showing the attainment category
}
\description{
Get attainment categories for each year and bay segment using chlorophyll and light attenuation
}
\examples{
\dontrun{
avedat <- anlz_avedat(epcdata)
anlz_attain(avedat)
}
}
