% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_attainsite.R
\name{anlz_attainsite}
\alias{anlz_attainsite}
\title{Get site attainments}
\usage{
anlz_attainsite(
  avedatsite,
  thr = c("chla", "la"),
  trgs = NULL,
  yrrng = NULL,
  thrs = FALSE
)
}
\arguments{
\item{avedatsite}{result returned from \code{\link{anlz_avedatsite}}}

\item{thr}{chr string indicating with water quality value and appropriate threshold to to plot, one of "chl" for chlorophyll and "la" for light availability}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{yrrng}{optional numeric value for year to return, defaults to all}

\item{thrs}{logical indicating if attainment category is relative to targets (default) or thresholds}
}
\value{
a \code{data.frame} for each year and site showing the attainment category
}
\description{
Get site attainment categories for chlorophyll or light attenuation
}
\details{
This function is a simplication of the attainment categories returned by \code{\link{anlz_attain}}.  Sites are only compared to the targets/thresholds that apply separately for chlorophyll or light attenuation.
}
\examples{
avedatsite <- anlz_avedatsite(epcdata)
anlz_attainsite(avedatsite)
}
\concept{analyze}
