% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_avedatsite.R
\name{anlz_avedatsite}
\alias{anlz_avedatsite}
\title{Estimate annual means by site}
\usage{
anlz_avedatsite(epcdata, partialyr = FALSE)
}
\arguments{
\item{epcdata}{\code{data.frame} formatted from \code{read_importwq}}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter}
}
\value{
Mean estimates for chlorophyll and secchi
}
\description{
Estimate annual means by site for chlorophyll and secchi data
}
\examples{
# view average estimates
anlz_avedatsite(epcdata)
}
\seealso{
Other analyze: 
\code{\link{anlz_attain}()},
\code{\link{anlz_avedat}()},
\code{\link{anlz_iwrraw}()},
\code{\link{anlz_refs}()},
\code{\link{anlz_tbbimed}()},
\code{\link{anlz_tdlcrkindic}()},
\code{\link{anlz_tdlcrk}()},
\code{\link{anlz_transectave}()},
\code{\link{anlz_transectocc}()},
\code{\link{anlz_yrattain}()}
}
\concept{analyze}
