% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_enteromap.R
\name{anlz_enteromap}
\alias{anlz_enteromap}
\title{Assign threshold categories to Enterococcus data}
\usage{
anlz_enteromap(fibdata, yrsel = NULL, mosel = NULL, wetdry = FALSE)
}
\arguments{
\item{fibdata}{data frame of Enterococcus sample data as returned by \code{\link{enterodata}} or \code{\link{anlz_fibwetdry}}}

\item{yrsel}{optional numeric to filter data by year}

\item{mosel}{optional numeric to filter data by month}

\item{wetdry}{logical; if \code{TRUE}, incorporate wet/dry differences. if \code{FALSE} (default), do not differentiate between wet and dry samples.}
}
\value{
A \code{data.frame} similar to \code{fibdata} with additional columns describing station categories and optionally filtered by arguments passed to the function
}
\description{
Assign threshold categories to Enterococcus data
}
\details{
This function is based on \code{\link{anlz_fibmap}}, but is specific to Enterococcus data downloaded via \code{\link{read_importentero}}. It creates categories for mapping using \code{\link{show_enteromap}}. Optionally, if samples have been defined as 'wet' or not via \code{\link{anlz_fibwetdry}}, this can be represented via symbols on the map.  Categories based on relevant thresholds are assigned to each observation.  The categories are specific to Enterococcus in marine waters (\code{class} of 2 or 3M). A station is categorized into one of four ranges defined by the thresholds as noted in the \code{cat} column of the output, with corresponding colors appropriate for each range as noted in the \code{col} column of the output.
}
\examples{
\dontrun{
anlz_enteromap(enterodata, yrsel = 2020, mosel = 9)

# wet/dry samples
entero_wetdry <- anlz_fibwetdry(enterodata, catch_precip)
anlz_enteromap(entero_wetdry, yrsel = 2020, mosel = 9, wetdry = TRUE)

# this will give the same output as the first example
anlz_enteromap(entero_wetdry, yrsel = 2020, mosel = 9, wetdry = FALSE)
}
}
