% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_fibmap.R
\name{anlz_fibmap}
\alias{anlz_fibmap}
\title{Assign threshold categories to Fecal Indicator Bacteria (FIB) data}
\usage{
anlz_fibmap(fibdata, yrsel = NULL, mosel = NULL, areasel = NULL)
}
\arguments{
\item{fibdata}{input FIB \code{data.frame} as returned by \code{\link{read_importfib}}}

\item{yrsel}{optional numeric value to filter output by years in \code{fibdata}}

\item{mosel}{optional numeric value to filter output by month in \code{fibdata}}

\item{areasel}{optional character string to filter output by stations in the \code{"Alafia"} or \code{"Hillsborough"} River basins, see details}
}
\value{
A \code{data.frame} similar to \code{fibdata} with additional columns describing station categories and optionally filtered by arguments passed to the function
}
\description{
Assign threshold categories to Fecal Indicator Bacteria (FIB) data
}
\details{
This function is used to create FIB categories for mapping using \code{\link{show_fibmap}}.  Categories based on relevant thresholds are assigned to each observation.  The categories are specific to E. coli or Enterococcus and are assigned based on the station class as freshwater (\code{class} as 1 or 3F) or marine (\code{class} as 2 or 3M), respectively.  A station is categorized into one of four ranges defined by the thresholds as noted in the \code{cat} column of the output, with corresponding colors appropriate for each range as noted in the \code{col} column of the output.

The \code{areasel} argument can indicate either \code{"Alafia"} or \code{"Hillsborough"} for the corresponding river basins, where rows in  \code{fibdata} are filtered based on the the selection.  The Alafia River basin includes values in the \code{area} column of \code{fibdata} as \code{"Alafia River"} and \code{"Alafia River Tributary"}.  The Hillsborough River basin includes values in the \code{area} column of \code{fibdat} as \code{"Hillsborough River"}, \code{"Hillsborough River Tributary"}, \code{"Lake Thonotosassa"}, \code{"Lake Thonotosassa Tributary"}, and \code{"Lake Roberta"}.  Not all areas may be present based on the selection.
}
\examples{
# assign categories to all
anlz_fibmap(fibdata)

# filter by year, month, and area
anlz_fibmap(fibdata, yrsel = 2020, mosel = 7, areasel = 'Alafia')
}
\concept{anlz}
