% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_fibwetdry.R
\name{anlz_fibwetdry}
\alias{anlz_fibwetdry}
\title{Identify Fecal Indicator Bacteria samples as coming from a 'wet' or 'dry' time period}
\usage{
anlz_fibwetdry(fibdata, precipdata, temporal_window = 2, wet_threshold = 0.5)
}
\arguments{
\item{fibdata}{input data frame}

\item{precipdata}{input data frame as returned by \code{\link{read_importrain}}. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object \code{\link{catchprecip}} has this data from 1995-2023 for select stations.}

\item{temporal_window}{numeric, number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)}

\item{wet_threshold}{numeric, inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period}
}
\value{
a data frame; the original fibdata data frame with three additional columns. \code{rain_sampleDay} is the total rain (inches) on the day of sampling, \code{rain_total} is the total rain (inches) for the period of time defined by \code{temporal_window}, and \code{wet_sample} is logical, indicating whether the rainfall for that station's catchment exceeded the amount over the time period specified in args.
}
\description{
Identify Fecal Indicator Bacteria samples as coming from a 'wet' or 'dry' time period
}
\details{
This function allows the user to specify a threshold for declaring a sample to be taken after an important amount of rain over an important amount of days, and declaring it to be 'wet'. This is of interest because samples taken after significant precipitation (definitions of this vary, which is why the user can specify desired thresholds) are more likely to exceed relevant bacterial thresholds. Identifying samples as 'wet' or not allows for calculation of further indices for wet and dry subsets of samples.
}
\examples{
entero_wetdry <- anlz_fibwetdry(enterodata, catchprecip)
head(entero_wetdry)
}
