% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_hmpreport.R
\name{anlz_hmpreport}
\alias{anlz_hmpreport}
\title{Evaluate Habitat Master Plan progress for report card}
\usage{
anlz_hmpreport(acres, subtacres, hmptrgs)
}
\arguments{
\item{acres}{\code{data.frame} for intertidal and supratidal land use and cover of habitat types for each year of data}

\item{subtacres}{\code{data.frame} for subtidal cover of habitat types for each year of data}

\item{hmptrgs}{\code{data.frame} of Habitat Master Plan targets and goals}
}
\value{
A summarized \code{data.frame} appropriate for creating a report card
}
\description{
Evaluate Habitat Master Plan progress for report card
}
\details{

}
\examples{
# view summarized data for report card
anlz_hmpreport(acres, subtacres, hmptrgs)
}
\concept{analyze}
