% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_refs.R
\name{anlz_refs}
\alias{anlz_refs}
\title{Convert references csv to bib}
\usage{
anlz_refs(path)
}
\arguments{
\item{path}{chr string of path to reference csv file or data frame object}
}
\value{
A data frame with references formatted as bib entries
}
\description{
Convert references csv to bib
}
\examples{

# input and format
path <- 'https://raw.githubusercontent.com/tbep-tech/tbep-refs/master/tbep-refs.csv'
bibs <- anlz_refs(path)

\dontrun{
# save output
 writeLines(bibs, 'formatted.bib')
}
}
\seealso{
Other analyze: \code{\link{anlz_attain}},
  \code{\link{anlz_avedatsite}}, \code{\link{anlz_avedat}},
  \code{\link{anlz_iwrraw}},
  \code{\link{anlz_tdlcrkindic}},
  \code{\link{anlz_tdlcrk}}, \code{\link{anlz_yrattain}}
}
\concept{analyze}
