% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_sedimentaddtot.R
\name{anlz_sedimentaddtot}
\alias{anlz_sedimentaddtot}
\title{Add contaminant totals to sediment data}
\usage{
anlz_sedimentaddtot(
  sedimentdata,
  yrrng = c(1993, 2021),
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"),
  funding_proj = "TBEP",
  param = NULL,
  pelave = TRUE
)
}
\arguments{
\item{sedimentdata}{input sediment \code{data.frame} as returned by \code{\link{read_importsediment}}}

\item{yrrng}{numeric vector indicating min, max years to include, use single year for one year of data}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{funding_proj}{chr string for the funding project, one to many of "TBEP" (default), "TBEP-Special", "Apollo Beach", "Janicki Contract", "Rivers", "Tidal Streams"}

\item{param}{optional character string of a parameter to filter the results}

\item{pelave}{logical indicating if output is used for \code{\link{anlz_sedimentpel}}}
}
\value{
A \code{data.frame} object similar to the input, but filtered by the arguments and contaminant totals added. Replicate samples are also removed.
}
\description{
Add contaminant totals to sediment data
}
\details{
This function adds totals to the \code{sedimentdata} input for total PCBs, total DDT, total LMW PAH, total HMW PAH, and total PAH.  Appropriate TEL/PEL values for the totals are also added.
}
\examples{
anlz_sedimentaddtot(sedimentdata)
}
\concept{anlz}
