% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_sedimentpelave.R
\name{anlz_sedimentpelave}
\alias{anlz_sedimentpelave}
\title{Get average concentrations for a sediment parameter by bay segment}
\usage{
anlz_sedimentpelave(
  sedimentdata,
  yrrng = c(1993, 2022),
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"),
  funding_proj = "TBEP"
)
}
\arguments{
\item{sedimentdata}{input sediment \code{data.frame} as returned by \code{\link{read_importsediment}}}

\item{yrrng}{numeric vector indicating min, max years to include, use single year for one year of data}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{funding_proj}{chr string for the funding project, one to many of "TBEP" (default), "TBEP-Special", "Apollo Beach", "Janicki Contract", "Rivers", "Tidal Streams"}
}
\value{
A \code{data.frame} of the average of the average PEL ratios for each bay segment
}
\description{
Get average concentrations for a sediment parameter by bay segment
}
\details{
Summaries for all bay segments are returned by default. The averages and confidence intervals are specific to the year ranges in \code{yrrng}.
}
\examples{
anlz_sedimentpelave(sedimentdata)
}
\seealso{
\code{\link{show_sedimentpelave}}
}
\concept{anlz}
