% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbbimed.R
\name{anlz_tbbimed}
\alias{anlz_tbbimed}
\title{Get annual medians of Tampa Bay Benthic Index scores by bay segment}
\usage{
anlz_tbbimed(
  tbbiscr,
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"),
  rev = FALSE,
  yrrng = c(1993, 2018)
)
}
\arguments{
\item{tbbiscr}{input data frame as returned by \code{\link{anlz_tbbiscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{yrrng}{numeric indicating year ranges to evaluate}
}
\value{
A data frame of annual medians by bay segment
}
\description{
Get annual medians of Tampa Bay Benthic Index scores by bay segment
}
\examples{
tbbiscr <- anlz_tbbiscr(benthicdata)
anlz_tbbimed(tbbiscr)
}
\seealso{
Other anlz: 
\code{\link{anlz_tbbiscr}()},
\code{\link{anlz_tbniave}()},
\code{\link{anlz_tbnimet}()},
\code{\link{anlz_tbniscr}()}
}
\concept{anlz}
