% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbniave.R
\name{anlz_tbniave}
\alias{anlz_tbniave}
\title{Get annual averages of Tampa Bay Nekton Index scores by bay segment}
\usage{
anlz_tbniave(
  tbniscr,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  rev = FALSE,
  perc = c(32, 46)
)
}
\arguments{
\item{tbniscr}{input dat frame as returned by \code{\link{anlz_tbniscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB"}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{perc}{numeric values indicating break points for score categories}
}
\value{
A data frame of annual averages by bay segment
}
\description{
Get annual averages of Tampa Bay Nekton Index scores by bay segment
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
anlz_tbniave(tbniscr)
}
\seealso{
Other anlz: 
\code{\link{anlz_tbnimet}()},
\code{\link{anlz_tbniscr}()}
}
\concept{anlz}
