% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbnimet.R
\name{anlz_tbnimet}
\alias{anlz_tbnimet}
\title{Get all raw metrics for Tampa Bay Benthic Index}
\usage{
anlz_tbnimet(fimdata, all = TRUE)
}
\arguments{
\item{fimdata}{\code{data.frame} formatted from \code{read_importfim}}

\item{all}{logical indicating if only TBNI metrics are returned, otherwise all are calcualted}
}
\value{
A data frame of raw metrics in wide fomat.  If \code{all = TRUE}, all metrics are returned, otherwise only \code{NumTaxa}, \code{BenthicTaxa}, \code{TaxaSelect}, \code{NumGuilds}, and \code{Shannon} are returned.
}
\description{
Get all raw metrics for Tampa Bay Benthic Index
}
\details{
All raw metrics are returned in addition to those required for the TBNI.  Each row shows metric values for a station, year, and month where fish catch data were available.
}
\examples{
anlz_tbnimet(fimdata)
}
\concept{anlz}
