% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tdlcrk.R
\name{anlz_tdlcrk}
\alias{anlz_tdlcrk}
\title{Estimate tidal creek report card scores}
\usage{
anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2018)
}
\arguments{
\item{tidalcreeks}{\code{\link[sf]{sf}} object for population of tidal creeks}

\item{iwrraw}{FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}}

\item{yr}{numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date}
}
\value{
A \code{\link{data.frame}} with the report card scores for each creek, as red, orange, yellow, green, or no data
}
\description{
Estimate tidal creek report card scores
}
\examples{
anlz_tdlcrk(tidalcreeks, iwrraw, yr)
}
\seealso{
Other analyze: \code{\link{anlz_attain}},
  \code{\link{anlz_avedatsite}}, \code{\link{anlz_avedat}},
  \code{\link{anlz_yrattain}}
}
\concept{analyze}
