% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tdlcrkindic.R
\name{anlz_tdlcrkindic}
\alias{anlz_tdlcrkindic}
\title{Analyze tidal creek water quality indicators}
\usage{
anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
}
\arguments{
\item{tidalcreeks}{\code{\link[sf]{sf}} object for population of tidal creeks}

\item{iwrraw}{FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}}

\item{yr}{numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date}
}
\value{
A \code{\link{data.frame}} with the indicator values for each tidal creek
}
\description{
Estimate tidal creek water quality indicators to suppport report card scores
}
\details{
Annual geometric means for additional water quality data available at each wbid, JEI combination.  Florida trophic state index values are also estimated where data are available.
}
\examples{
anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
}
\seealso{
Other analyze: \code{\link{anlz_attain}},
  \code{\link{anlz_avedatsite}}, \code{\link{anlz_avedat}},
  \code{\link{anlz_iwrraw}}, \code{\link{anlz_refs}},
  \code{\link{anlz_tdlcrk}}, \code{\link{anlz_yrattain}}
}
\concept{analyze}
