% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_transectocc.R
\name{anlz_transectocc}
\alias{anlz_transectocc}
\title{Get seagrass average abundance and occurrence across transects}
\usage{
anlz_transectocc(transect)
}
\arguments{
\item{transect}{data frame returned by \code{\link{read_transect}}}
}
\value{
A data frame with abundance and frequency occurrence estimates aggregated by species, transect, and date
}
\description{
Get seagrass average abundance and occurrence across transects
}
\details{
Abundance and frequency occurrence are estimated as in methods described in Sherwood et al. 2017, equations 1 and 2. In short, frequency occurrence is estimated as the number of instances a species was observed along a transect divided by the number of placements along a transect and average abundance was estimated as the sum of species-specific Braun-Blanquet scores divided by the number of placements along a transect.  All attached and drift algae species are aggregated, except Caulerpa spp. which are grouped separately.
}
\examples{
#' \dontrun{
transect <- read_transect()
}
anlz_transectocc(transect)
}
\references{
Sherwood, E.T., Greening, H.S., Johansson, J.O.R., Kaufman, K., Raulerson, G.E. 2017. Tampa Bay (Florida, USA): Documenting seagrass recovery since the 1980's and reviewing the benefits. Southeastern Geographer. 57(3):294-319.
}
