% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_trnjsn.R
\name{anlz_trnjsn}
\alias{anlz_trnjsn}
\title{Format transect data from Water Atlas}
\usage{
anlz_trnjsn(trnjsn, training = FALSE)
}
\arguments{
\item{trnjsn}{output from \code{\link{read_trnjsn}}}

\item{training}{logical if input are transect training data or complete database}
}
\value{
data frame in long format
}
\description{
Format transect data from Water Atlas
}
\examples{
trnjsn <- read_trnjsn(training = TRUE)
anlz_trnjsn(trnjsn, training = TRUE)
}
