% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthicdata.R
\docType{data}
\name{benthicdata}
\alias{benthicdata}
\title{Benthic data for the Tampa Bay Benthic Index current as of 08012021}
\format{
A nested \code{\link[tibble]{tibble}} with 3 rows and 2 variables:
\describe{
  \item{name}{chr}
  \item{value}{list}
  }
}
\usage{
benthicdata
}
\description{
Benthic data for the Tampa Bay Benthic Index current as of 08012021
}
\examples{
\dontrun{
# location to download data
path <- '~/Desktop/benthic.zip'

# load and assign to object
benthicdata <- read_importbenthic(path, download_latest = TRUE, remove = TRUE)

save(benthicdata, file = 'data/benthicdata.RData', compress = 'xz')

}
}
\concept{data}
\keyword{datasets}
