% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmap.R
\docType{data}
\name{bsmap}
\alias{bsmap}
\title{Terrain basemap}
\format{A \code{\link[ggmap]{ggmap}} object}
\usage{
bsmap
}
\description{
Terrain basemap
}
\seealso{
Other utilities: \code{\link{epcdata}},
  \code{\link{stations}}, \code{\link{targets}},
  \code{\link{tbseg}}
}
\concept{utilities}
\keyword{datasets}
