% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_pixels.R
\docType{data}
\name{catch_pixels}
\alias{catch_pixels}
\title{Catchments and radar pixels (for precip) of selected Enterococcus stations}
\format{
A data frame with two columns:
\describe{
 \item{station}{a character string of the Water Quality Portal station name}
 \item{pixel}{a number; each row is a single pixel from the Southwest Florida Water Management District's radar-estimated rainfall shapefile}
}
}
\usage{
catch_pixels
}
\description{
Catchments and radar pixels (for precip) of selected Enterococcus stations
}
\keyword{datasets}
