% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catch_precip.R
\docType{data}
\name{catch_precip}
\alias{catch_precip}
\title{Daily precip by catchment for selected Enterococcus stations}
\format{
A data frame with three columns:
\describe{
 \item{station}{a character string of the Water Quality Portal station name}
 \item{date}{a date}
 \item{rain}{a number; inches of rain for that date, averaged across all pixels in the station's catchment}
}
}
\source{
Southwest Florida Management District radar-estimated daily rainfall data, \url{ftp://ftp.swfwmd.state.fl.us/pub/radar_rainfall/Daily_Data/}
}
\usage{
catch_precip
}
\description{
Daily precip by catchment for selected Enterococcus stations
}
\details{
Daily precipitation data from 1995-2023, provided by the Southwest Florida Water Management District, were downloaded at the pixel level and averaged to the catchment level for key Enterococcus sampling stations. Generated by \code{data-raw/catch_precip-raw.R} (view on github: \url{https://github.com/tbep-tech/tbeptools/blob/master/data-raw/catch_precip-raw.R})
}
\keyword{datasets}
