% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epcdata.R
\docType{methods}
\name{chlplot}
\alias{chlplot}
\alias{chlplot,epcdata-method}
\title{Plot annual chlorophyll for a segment}
\usage{
chlplot(object, bay_segment = c("OTB", "HB", "MTB", "LTB"))

\S4method{chlplot}{epcdata}(object, bay_segment = c("OTB", "HB", "MTB",
  "LTB"))
}
\arguments{
\item{object}{\code{\link{epcdata}} object}

\item{bay_segment}{chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot chlorophyll annual values by standard for a bay segment
}
\examples{
\dontrun{
chlplot(epcdata, bay_segment = 'OTB')
}
}
