% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbasin.R
\docType{data}
\name{dbasin}
\alias{dbasin}
\title{Spatial data object of minor subbasins (dbasins) in the Tampa Bay watershed}
\format{
A simple features \code{\link[sf]{sf}} object (POLYGON) with 436 features and 26 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
}
\usage{
dbasin
}
\description{
Spatial data object of minor subbasins (dbasins) in the Tampa Bay watershed
}
\examples{
library(sf)
plot(st_geometry(dbasin))
}
\concept{data}
\keyword{datasets}
