% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-epcdata.R, R/methods-epcdata.R
\docType{class}
\name{epcdata-class}
\alias{epcdata-class}
\alias{epcdata}
\alias{show,epcdata-method}
\alias{rawdat}
\alias{rawdat,epcdata-method}
\alias{frmdat}
\alias{frmdat,epcdata-method}
\alias{avedat}
\alias{avedat,epcdata-method}
\title{An S4 class to represent an \code{epcdata} object}
\usage{
\S4method{show}{epcdata}(object)

rawdat(object)

\S4method{rawdat}{epcdata}(object)

frmdat(object)

\S4method{frmdat}{epcdata}(object)

avedat(object)

\S4method{avedat}{epcdata}(object)
}
\arguments{
\item{object}{\code{epcdata} object created with \code{\link{load_epchc_wq}}}
}
\description{
The S4 \code{epcdata} object includes several methods shown below.
}
\section{Slots}{

\describe{
\item{\code{rawdat}}{\code{data.frame} of raw imported EPCHC data}

\item{\code{frmdat}}{\code{data.frame} of lightly formatted EPCHC data for secchi and chlorophyll data}

\item{\code{avedat}}{\code{list} of annual (\code{ann}) and monthly (\code{mos}) segment means for secchi, light attenuation, and chlorophyll data}
}}

\examples{
\dontrun{
# file path
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'

# load and assign to object
epcdata <- load_epchc_wq(xlsx)

rawdat(epcdata)
frmdat(epcdata)
avedat(epcdata)
}
}
