% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/form_epchc_wq.R
\name{form_epchc_wq}
\alias{form_epchc_wq}
\title{Format water quality data}
\usage{
form_epchc_wq(epcdata)
}
\arguments{
\item{epcdata}{input \code{data.frame} returned from \code{\link{load_epchc_wq}}}
}
\value{
A lightly formatted \code{data.frame} with chloropyll and secchi observations
}
\description{
Format water quality data
}
\details{
Secchi data VOB depths or secchis < 0.5 ft from bottom are assigned \code{NA}
}
\examples{
\dontrun{
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'
epcdata <- load_epchc_wq(xlsx) \%>\%
  form_epch_wq
}
}
