% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_epchc_wq}
\alias{get_epchc_wq}
\title{Download latest water quality file from epchc.org}
\usage{
get_epchc_wq(xlsx, download_latest_epchc = TRUE)
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite it not current}

\item{download_latest_epchc}{logical to download latest file regardless of local copy}
}
\value{
The local copy specified in the path by \code{xlsx} is overwritten by the new file is not current or \code{download_latest_epchc = TRUE}
}
\description{
Download latest water quality file from epchc.org
}
\details{
The local copy is checked against a temporary file downloade from \url{ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx}.  The local file is replaced with the downloaded file if the MD5 hashes are different.
}
\examples{
\dontrun{
test_xlsx <- file.path(dir_gdrive, "data/wq/2018_Results_Updated.xls")
get_epch_wq(test_xlsx)
}
}
