% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmptrgs.R
\docType{data}
\name{hmptrgs}
\alias{hmptrgs}
\title{Habitat Master Plan targets and goals}
\format{
A data frame showing 2030 targets and 2050 goals
}
\usage{
hmptrgs
}
\description{
Habitat Master Plan targets and goals
}
\examples{
\dontrun{
load(url("https://github.com/tbep-tech/hmpu-workflow/raw/master/data/trgs.RData"))

hmptrgs <- trgs
save(hmptrgs, file = 'data/hmptrgs.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
