% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_epchc_wq.R
\name{mean_epchc_wq}
\alias{mean_epchc_wq}
\title{Estimate annual means}
\usage{
mean_epchc_wq(datin)
}
\arguments{
\item{datin}{\code{data.frame} formatted from \code{load_epchc_wq}}
}
\value{
Mean estimates for chlorophyll and secchi
}
\description{
Estimate annual means for chlorophyll and secchi data
}
\examples{
\dontrun{
# file path
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'

# load and assign to object
epcdata <- load_epchc_wq(xlsx)

# view average estimates
avedat(epcdata)
}
}
