% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polcofibdata.R
\docType{data}
\name{polcofibdata}
\alias{polcofibdata}
\title{Polk County FIB data as of 20250304}
\format{
A data frame with 833 rows and 13 variables:
\describe{
  \item{polco_station}{chr, Station name}
  \item{SampleTime}{POSIXct, Date/time of sampling}
  \item{class}{chr, Waterbody class (\code{Fresh}, \code{Marine})}
  \item{yr}{num, Year of sampling}
  \item{mo}{num, Month of sampling}
  \item{Latitude}{num, Latitude, WGS84}
  \item{Longitude}{num, Latitude, WGS84}
  \item{Sample_Depth_m}{num, Depth of sample, meters}
  \item{var}{chr, Variable name (\code{ecoli}, \code{entero}}
  \item{val}{num, Value of variable}
  \item{uni}{num, Units of variable}
  \item{qual}{num, Qualifier code}
  \item{area}{chr, Location name}
  }
}
\usage{
polcofibdata
}
\description{
Polk County FIB data as of 20250304
}
\details{
This dataset includes FIB data from Polk County (21FLPOLK_WQX) where station class is marine (estuary) and Enterococcus data is present or the station class is freshwater (stream/river, reservoir) and E. coli data is present. The data is formatted from the raw data loaded from \code{\link{read_importwqp}}.
}
\examples{
\dontrun{
polcofibdata <- read_importwqp('21FLPOLK_WQX', type = 'fib')

nrow(polcofibdata)
ncol(polcofibdata)

save(polcofibdata, file = 'data/polcofibdata.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
