% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chkdate.R
\name{read_chkdate}
\alias{read_chkdate}
\title{Compare date of local xlsx file with the same file on the server}
\usage{
read_chkdate(urlin, xlsx)
}
\arguments{
\item{urlin}{chr string of full path to file on the server}

\item{xlsx}{chr string of full path to local file}
}
\value{
A logical vector indicating if the local file is current
}
\description{
Compare date of local xlsx file with the same file on the server
}
\examples{
\dontrun{
urlin <- "ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx"
xlsx <- '~/Desktop/2018_Results_Updated.xls'
read_chkdate(urlin, xlsx)
}
}
\seealso{
Other read: \code{\link{read_dlcurrent}},
  \code{\link{read_formfim}}, \code{\link{read_formphyto}},
  \code{\link{read_formwq}}, \code{\link{read_importfim}},
  \code{\link{read_importphyto}},
  \code{\link{read_importwq}}
}
\concept{read}
