% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlcurrent.R
\name{read_dlcurrent}
\alias{read_dlcurrent}
\title{Download latest file from epchc.org}
\usage{
read_dlcurrent(locin, download_latest = TRUE, urlin)
}
\arguments{
\item{locin}{chr string path for local file, to overwrite it not current}

\item{download_latest}{logical to download latest file regardless of local copy}

\item{urlin}{url for file location}
}
\value{
The local copy specified in the path by \code{locin} is overwritten by the new file is not current or \code{download_latest = TRUE}.  The function does nothing if \code{download_latest = FALSE}.
}
\description{
Download latest file from epchc.org
}
\details{
The local copy is checked against a temporary file downloaded from the location specified by \code{urlin}.  The local file is replaced with the downloaded file if the MD5 hashes are different.
}
\examples{
\dontrun{
locin <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'
urlin1 <- 'https://epcbocc.sharepoint.com/:x:/s/PublishedShare/'
urlin2 <- 'IQAtFpJjaVHzS6ENmWdNf4csAXO5fC5QFwI8U4b69Yyc8FA?e=tMe7Yy'
urlin <- paste0(urlin1, urlin2, '&download=1')
read_dlcurrent(locin = locin, urlin = urlin)
}
}
\concept{read}
