% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formbenthic.R
\name{read_formbenthic}
\alias{read_formbenthic}
\title{Format benthic data for the Tampa Bay Benthic Index}
\usage{
read_formbenthic(pathin)
}
\arguments{
\item{pathin}{A path to unzipped csv files with base tables used to calculate benthic index}
}
\value{
A nested \code{\link[tibble]{tibble}} of station, field sample, and taxa data
}
\description{
Format benthic data for the Tampa Bay Benthic Index
}
\details{
Function is used internally within \code{\link{read_importbenthic}}
}
\examples{
\dontrun{

# location to download data
path <- '~/Desktop/benthic.zip'

# download
urlin1 <- 'https://epcbocc.sharepoint.com/:u:/s/PublishedShare/'
urlin2 <- 'IQDODa-pbW8pTaZMg-93L9sTAZW1Y6hjj1TD_8RY0mFGiQ8?e=FIphoq'
urlin <- paste0(urlin1, urlin2, '&download=1')
read_dlcurrent(path, download_latest = TRUE, urlin = urlin)

# unzip
tmppth <- tempfile()
utils::unzip(path, exdir = tmppth, overwrite = TRUE)

# format benthic data
read_formbenthic(pathin = tmppth)

# remove temporary path
unlink(tmppth, recursive = TRUE)

}
}
\concept{read}
