% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formfib.R
\name{read_formfib}
\alias{read_formfib}
\title{Format Fecal Indicator Bacteria (FIB) data}
\usage{
read_formfib(datin)
}
\arguments{
\item{datin}{input \code{data.frame} loaded from \code{\link{read_importwq}}}
}
\value{
A lightly formatted \code{data.frame} with FIB data
}
\description{
Format Fecal Indicator Bacteria (FIB) data
}
\details{
Returns FIB data including E. coli and Enterococcus cell concentrations and waterbody class (freshwater as 1 or 3F, marine as 2 or 3M) for staions and sample dates, function is used internally within \code{\link{read_importfib}}

 Values are returned for E. coli (\code{ecoli}), Enterococcus (\code{ecocci}), Fecal Coliform (\code{fcolif}), and Total Coliform (\code{totcol}).  Values shown are # of colonies per 100 mL of water (\code{#/100mL}). Qualifiers columns for each are also returned with the \code{_q} suffix. Qualifier codes can be interpreted from the source spreadsheet.

 Concentrations noted with \code{<} or \code{>} in the raw data are reported as is, with only the numeric value shown.  Samples with this notation can be determined from the qualifier columns.

 Only stations with AreaName in the source data as Hillsborough River, Hillsborough River Tributary, Alafia River, Alafia River Tributary, Lake Thonotosassa, Lake Thonotosassa Tributary, Lake Roberta are returned.
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'

# load raw data and assign to object
epcall <- read_importepc(xlsx, download_latest = T)

# final formatting
fibdata <- read_formfib(epcall)
}
}
\seealso{
\code{\link{read_importfib}}
}
\concept{read}
