% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formphyto.R
\name{read_formphyto}
\alias{read_formphyto}
\title{Format phytoplankton data}
\usage{
read_formphyto(datin)
}
\arguments{
\item{datin}{input \code{data.frame} loaded from \code{\link{read_importphyto}}}
}
\value{
A formatted \code{data.frame} with phytoplankton count data
}
\description{
Format phytoplankton data
}
\details{
Only seven taxonomic groups are summarized. Pyrodinium bahamense, Karenia brevis, Tripos hircus, Pseudo-nitzschia sp., and Pseudo-nitzschia pungens are retained at the species level.  Bacillariophyta and Cyanobacteria are retained at the phylum level.  All other taxa are grouped into an "other" category.
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/phyto_data.xlsx'

# load and assign to object
phytodata <- read_importphyto(xlsx)
}
}
\seealso{
\code{\link{read_importphyto}}

Other read: 
\code{\link{read_chkdate}()},
\code{\link{read_dlcurrent}()},
\code{\link{read_formfim}()},
\code{\link{read_formwq}()},
\code{\link{read_importfim}()},
\code{\link{read_importphyto}()},
\code{\link{read_importwq}()}
}
\concept{read}
