% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formwq.R
\name{read_formwq}
\alias{read_formwq}
\title{Format water quality data}
\usage{
read_formwq(datin)
}
\arguments{
\item{datin}{input \code{data.frame} loaded from \code{\link{read_importwq}}}
}
\value{
A lightly formatted \code{data.frame} with chloropyll and secchi observations
}
\description{
Format water quality data
}
\details{
Secchi data VOB depths or secchis < 0.5 ft from bottom are assigned \code{NA}, function is used internally within \code{\link{read_importwq}}
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/2018_Results_Updated.xls'

# load and assign to object
epcdata <- read_importwq(xlsx, download_latest = T)
}
}
\seealso{
\code{\link{read_importwq}}

Other read: 
\code{\link{read_chkdate}()},
\code{\link{read_dlcurrent}()},
\code{\link{read_formbenthic}()},
\code{\link{read_formfim}()},
\code{\link{read_formphyto}()},
\code{\link{read_importbenthic}()},
\code{\link{read_importfim}()},
\code{\link{read_importphyto}()},
\code{\link{read_importwq}()}
}
\concept{read}
