% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formwqmanatee.R
\name{read_formwqmanatee}
\alias{read_formwqmanatee}
\title{Format water quality and station metadata for Manatee County}
\usage{
read_formwqmanatee(res, sta, trace = F)
}
\arguments{
\item{res}{A data frame containing water quality results obtained from the API}

\item{sta}{A data frame containing station metadata obtained from the API}

\item{trace}{Logical indicating whether to display progress messages, default is \code{FALSE}}
}
\value{
A data frame containing formatted water quality and station metadata for Manatee County.
}
\description{
Format water quality and station metadata for Manatee County
}
\details{
This function is used by \code{\link{read_importwqmanatee}} to combine, format, and process water quality data (\code{res}) and station metadata (\code{sta}) obtained from the Water Quality Data Exchange API for Manatee County. The resulting data frame includes the date, time, station identifier, latitude, longitude, variable name, value, unit, and quality flag.
}
\examples{
\dontrun{
url <- list(
  Result = "https://www.waterqualitydata.us/data/Result/search?mimeType=csv&zip=no",
  Station = "https://www.waterqualitydata.us/data/Station/search?mimeType=csv&zip=no"
)

headers <- c(
  "Content-Type" = "application/json",
  "Accept" = "application/zip"
)

body <- list(
  organization = c("21FLMANA_WQX"),
  sampleMedia = c("Water"),
  characteristicType = c("Information", "Nutrient", "Biological, Algae, Phytoplankton,
    Photosynthetic Pigments"),
  providers = c("STORET"),
  siteType = c("Estuary")#,
  # ActivityTypeCode = c("Field Msr/Obs", "Sample-Routine")
)

res <- url[['Result']] \%>\%
  httr::POST(httr::add_headers(headers), body = jsonlite::toJSON(body)) \%>\%
  httr::content('text') \%>\%
  read.csv(text = .)

sta <- url[['Station']] \%>\%
  httr::POST(httr::add_headers(headers), body = jsonlite::toJSON(body)) \%>\%
  httr::content('text') \%>\%
  read.csv(text = .)

# combine and format
read_formwqmanatee(res, sta)
}
}
\seealso{
\code{\link{read_importwqmanatee}}
}
\concept{read}
