% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importfib.R
\name{read_importfib}
\alias{read_importfib}
\title{Load local water quality file for Fecal Indicator Bacteria (FIB)}
\usage{
read_importfib(xlsx, download_latest = FALSE, na = c("", "NULL"), all = TRUE)
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{xlsx} if available}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}

\item{all}{logical indicating if all stations with FIB data are returned, default is \code{TRUE}, see details}
}
\value{
A \code{data.frame} of formatted FIB data.
}
\description{
Load local water quality file for Fecal Indicator Bacteria (FIB)
}
\details{
Loads the "RWMDataSpreadsheet" worksheet from the file located at \code{xlsx}.  The file is downloaded from \url{https://epcbocc.sharepoint.com/:x:/s/PublishedShare/IQAtFpJjaVHzS6ENmWdNf4csAXO5fC5QFwI8U4b69Yyc8FA?e=tMe7Yy&download=1}.  The files can be viewed at \url{https://epcbocc.sharepoint.com/:f:/s/Share/EiypSSYdsEFCi84Sv_6-t7kBUYaXiIqN0B1n2w57Z_V3kQ?e=NdZQcU}.

Returns FIB data including E. coli, Enterococcus, Fecal Coliform, and Total Coliform concentrations and waterbody class (freshwater as 1 or 3F, marine as 2 or 3M) for stations and sample dates.

Values are returned for E. coli (\code{ecoli}), Enterococcus (\code{entero}), Fecal Coliform (\code{fcolif}), and Total Coliform (\code{totcol}).  Values shown are # of colonies per 100 mL of water (\code{#/100mL}). Qualifier columns for each are also returned with the \code{_q} suffix. Qualifier codes can be interpreted from the source spreadsheet.

Concentrations noted with \code{<} or \code{>} in the raw data are reported as is, with only the numeric value shown.  Samples with this notation can be determined from the qualifier columns.

If \code{all = FALSE}, only stations with AreaName in the source data as Hillsborough River, Hillsborough River Tributary, Alafia River, Alafia River Tributary, Lake Thonotosassa, Lake Thonotosassa Tributary, and Lake Roberta.
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'

# load and assign to object
fibdata <- read_importfib(xlsx, download_latest = T)

}
}
\seealso{
\code{\link{read_formfib}}
}
\concept{read}
