% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importfim.R
\name{read_importfim}
\alias{read_importfim}
\title{Load local FIM data for the Tampa Bay Nekton Index}
\usage{
read_importfim(csv, download_latest = FALSE, connecttimeout = 10,
  tryurl = FALSE)
}
\arguments{
\item{csv}{chr string path for local csv file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{csv} if available}

\item{connecttimeout}{numeric for maximum number of seconds to wait until connection timeout for \code{\link[RCurl]{getURL}}}

\item{tryurl}{logical indicating if \code{\link[RCurl]{getURL}} is repeatedly called in a \code{while} loop if first connection is unsuccessful}
}
\value{
A \code{data.frame} of formatted FIM data.
}
\description{
Load local FIM data for the Tampa Bay Nekton Index
}
\details{
Data downloaded from \url{'ftp://ftp.floridamarine.org/users/fim/tmac/NektonIndex/TampaBay_NektonIndexData_20200406.csv'}.
}
\examples{
\dontrun{
# file path
csv <- 'C:/Users/Owner/Desktop/fimraw.csv'

# load and assign to object
fimdata <- read_importfim(csv)
}
}
\seealso{
\code{\link{read_formwq}}, \code{\link{read_importphyto}}

Other read: \code{\link{read_chkdate}},
  \code{\link{read_dlcurrent}},
  \code{\link{read_formphyto}}, \code{\link{read_formwq}},
  \code{\link{read_importphyto}},
  \code{\link{read_importwq}}
}
\concept{read}
