% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importprism.R
\name{read_importprism}
\alias{read_importprism}
\title{Import PRISM daily weather data}
\usage{
read_importprism(
  sf_zones,
  fld_zones,
  zonal_csv,
  zonal_fun = "mean",
  bbox = sf::st_bbox(sf_zones),
  dir_tif = fs::path_ext_remove(zonal_csv),
  vars = c("tmin", "tmax", "tdmean", "ppt"),
  date_beg = as.Date("1981-01-01"),
  date_end = Sys.Date(),
  redo = F,
  verbose = F
)
}
\arguments{
\item{sf_zones}{spatial feature object ([`sf`]) with zones to extract zonal
statistics from PRISM daily weather data}

\item{fld_zones}{unique field name in `sf_zones` to include in extracted
zonal statistics}

\item{zonal_csv}{path to output of zonal statistics (`terra::zonal()`) as
table in CSV format}

\item{zonal_fun}{function to apply to the `terra::zonal()` function. Defaults
to "mean".}

\item{bbox}{bounding box of the spatial extent to crop PRISM daily rasters,
which could be bigger than `st_bbox(sf_zones)`.}

\item{dir_tif}{directory path to store downloaded and cropped PRISM daily
rasters (as GeoTIFF) across years with file name of format
`prism_daily_%m-%d.tif`. The default path is
`fs::path_ext_remove(zonal_csv)`.  Date and version in the raster layers
will be compared with the [update
schedule](https://prism.oregonstate.edu/calendar/list.php) to determine if
it should be updated, otherwise will be skipped.}

\item{vars}{character vector of PRISM variables to download and crop. The
default is `c("tmin", "tmax", "tdmean", "ppt")`. See Details for others and
details.}

\item{date_beg}{defaults to the start of PRISM daily availability 1981-01-01.}

\item{date_end}{defaults to today's date, but will be adjusted to the most
recently available data, per [update
schedule](https://prism.oregonstate.edu/calendar/list.php)}

\item{redo}{logical whether to recalculate the zonal statistics. Defaults to
False.}

\item{verbose}{logical whether to show informative messages on processing}
}
\value{
data frame with the following columns:
- ...
}
\description{
Download and crop PRISM daily weather data.
}
\details{
The [Parameter-elevation Relationship on Independent Slopes Model
(PRISM)](https://prism.oregonstate.edu/) is a combined dataset consisting of
ground gauge station and RADAR products, at a 4km grid resolution covering
the contiguous United States from 1981 to present.

### Variables

The available variables (`vars`) are:
- `tmin`: minimum temperature
- `tmax`: maximum temperature
- `tmean`: mean temperature (`tmean = mean(tmin, tmax)`)
- `tdmean`: mean dew point temperature
- `ppt`: total precipitation (rain and snow)
- `vpdmin`: daily minimum vapor pressure deficit
- `vpdmax`: daily maximum vapor pressure deficit

### Updates

The daily PRISM data is available over three stability levels, or 8 versions
given by the [update
schedule](https://prism.oregonstate.edu/calendar/list.php):
- `stable`: data is considered final and will not change (before 6 months)
- `provisional`: data is considered final but may be updated (previous 6 months)
- `early`: data is preliminary and may be updated (this month)

### More on PRISM

For more on [Parameter-elevation Relationship on Independent Slopes Model
(PRISM)](https://prism.oregonstate.edu/), see:
 -   [PRISM model methods (Daly et al., 2008)](https://prism.oregonstate.edu/documents/pubs/2008intjclim_physiographicMapping_daly.pdf)
 -   [PRISM dataset descriptions](https://prism.oregonstate.edu/documents/PRISM_datasets.pdf)
 -   [PRISM download methods](https://prism.oregonstate.edu/downloads) information (FTP, web services)
}
\examples{
dir_tmp <- tempdir()
tmp_csv <- file.path(dir_tmp, "prism.csv")

d <- read_importprism(
  sf_zones  = tbsegshed,
  fld_zones = "bay_segment",
  zonal_csv = tmp_csv,
  date_beg  = as.Date("1981-01-01"),
  date_end  = as.Date("1981-01-03"),
  verbose   = T)

d
}
