% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importrainmany.R
\name{read_importrainmany}
\alias{read_importrainmany}
\title{Run \code{read_importrain} for multiple years}
\usage{
read_importrainmany(yrs, quiet = FALSE, sleep = 5)
}
\arguments{
\item{yrs}{numeric vector of years to download (do not need to be in order)}

\item{quiet}{logical passed to \code{\link{read_importrain}} to suppress messages (default)}

\item{sleep}{numeric number of seconds to pause between years}
}
\value{
A data frame identical to that returned by \code{\link{read_importrain}} with the years requested
}
\description{
Run \code{read_importrain} for multiple years
}
\details{
This function is a simple wrapper for \code{\link{read_importrain}} to download multiple years of rainfall data from the Southwest Florida Water Management District's (SWFWMD) ftp site: ftp://ftp.swfwmd.state.fl.us/pub/radar_rainfall/Daily_Data/. The function will pause for \code{sleep} seconds between years to avoid overloading the server.
}
\examples{
\dontrun{
read_importrainmany(c(2021, 2022), quiet = F)
}
}
