% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importsediment.R
\name{read_importsediment}
\alias{read_importsediment}
\title{Download and import sediment data for Tampa Bay}
\usage{
read_importsediment(path, download_latest = FALSE, remove = FALSE)
}
\arguments{
\item{path}{chr string for local path where the zipped folder will be downloaded, must include .zip extension}

\item{download_latest}{logical to download latest if a more recent dataset is available}

\item{remove}{logical if the downloaded folder is removed after unzipping}
}
\value{
A \code{data.frame} of sediment data for Tampa Bay
}
\description{
Download and import sediment data for Tampa Bay
}
\details{
This function downloads and unzips a folder of results tables from \url{https://epcbocc.sharepoint.com/:u:/s/PublishedShare/IQAr2aDfEyegTrJLR61qt1CxAQYxfnVYYWAwkbtArczrJS4?e=EQSGO3&download=1} (viewable at \url{https://epcbocc.sharepoint.com/:f:/s/Share/EtOJfziTTa9FliL1oROb9OsBRZU-nO60fu_0NRC162hHjQ?e=4gUXgJ}).

The row entries for columns \code{"BetweenTELPEL"} and \code{"ExceedsPEL"} for rows where the \code{"Qualifier"} column is \code{"U"} or \code{"T"} (below detection, not detected) are assigned \code{NA}, regardless of the entry in the source data.
}
\examples{
\dontrun{
# location to download data
path <- '~/Desktop/sediment.zip'

# load and assign to object
sedimentdata <- read_importsediment(path, download_latest = TRUE)

}
}
\concept{read}
