% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwqmanatee.R
\name{read_importwqmanatee}
\alias{read_importwqmanatee}
\title{Import water quality data for Manatee County}
\usage{
read_importwqmanatee(trace = F)
}
\arguments{
\item{trace}{logical indicating whether to display progress messages, default \code{FALSE}}
}
\value{
A data frame containing the imported water quality data for Manatee County.
}
\description{
Import water quality data for Manatee County
}
\details{
This function retrieves water quality data from the Water Quality Data Exchange API for Manatee County. It fetches results and station metadata, combines and formats them using the \code{read_formwqmanatee} function, and returns the processed data as a data frame.  Parameters included are nutrients, chlorophyll, and Secchi depth
}
\examples{
\dontrun{
mancodata <- read_importwqmanatee()
}
}
\seealso{
\code{\link{read_formwqmanatee}}, \code{\link{read_importwq}}
}
\concept{read}
