% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwqwin.R
\name{read_importwqwin}
\alias{read_importwqwin}
\title{Retrieve water quality data from the Florida Department of Environmental Protection's Watershed Information Network (WIN)}
\usage{
read_importwqwin(start_date, org_id, verbose = FALSE)
}
\arguments{
\item{start_date}{character string for the start date in the format "YYYY-MM-DD"}

\item{org_id}{character string for the organization ID}

\item{verbose}{logical indicating whether to print verbose output}
}
\value{
A data frame containing the water quality data
}
\description{
Retrieve water quality data from the Florida Department of Environmental Protection's Watershed Information Network (WIN)
}
\details{
This function implements \code{\link{util_importwqwin}} iteratively to retrieve one to many pages of water quality results for the specified organization ID and start date. Data are retrieve using the API at \url{https://prodapps.dep.state.fl.us/dear-watershed/swagger-ui/index.html}.
}
\examples{
\dontrun{
dat <- read_importwqwin("2025-01-01", "21FLMANA", verbose = TRUE)
head(dat)
}
}
