% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_prism_rasters.R
\name{read_prism_rasters}
\alias{read_prism_rasters}
\title{Read PRISM raster layers into data frame}
\usage{
read_prism_rasters(dir_tif)
}
\arguments{
\item{dir_tif}{directory path PRISM daily rasters were downloaded and cropped
using [`read_importprism()`].}
}
\value{
A data frame with columns for:
- `path_tif` path to GeoTIFF
- `lyr` layer name
- `date` date of modeled weather data
- `md` month-day
- `variable` variable of modeled weather data
- `version` version (1-8)
- `date_updated` date the model was updated
}
\description{
Scan directory for rasters (*.tif) and parse PRISM layer names into columns of
a data frame.
}
\details{
Any existing rasters in `dir_tif` are scanned based on a common naming
structure for the raster file name (`prism_daily_{month}-{day}.tif`) and
layer names (`{date}_{variable}_v{version}-{date_updated}`) into a data
frame.

For more on the [Parameter-elevation Relationship on Independent Slopes Model
(PRISM)](https://prism.oregonstate.edu/), see [`read_importprism()`].
}
\examples{
dir_tif <- here::here("inst/prism")
read_prism_rasters(dir_tif)
}
\concept{read}
