% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_transect.R
\name{read_transect}
\alias{read_transect}
\title{Import JSON seagrass transect data from Water Atlas}
\usage{
read_transect(training = FALSE, raw = FALSE, retry = 5)
}
\arguments{
\item{training}{logical if training data are imported or the complete database}

\item{raw}{logical indicating if raw, unformatted data are returned, see details}

\item{retry}{integer indicating maximum number of retries on request failure, default 5}
}
\value{
data frame
}
\description{
Import JSON seagrass transect data from Water Atlas
}
\details{
The function imports a JSON file from the USF Water Atlas.  If \code{training = TRUE}, a dataset from the TBEP training survey is imported from \url{https://tampabay.wateratlas.usf.edu/seagrass-transect-data-portal/api/assessments/training}.  If \code{training = FALSE}, the entire transect survey database is imported from \url{https://tampabay.wateratlas.usf.edu/seagrass-transect-data-portal/api/assessments/all__use-with-care}.

Abundance is reported as a numeric value from 0 -5 for Braun-Blanquet coverage estimates, blade length is in cm, and short shoot density is number of shoots per square meter.  The short density is corrected for quadrat size included in the raw data.

If \code{raw = TRUE}, the unformatted data are returned.  The default is to use formatting that allows the raw data to be used with the downstream functions. The raw data may have extra information that may be of use outside of the plotting functions in this package.

The request is retried with exponential backoff up to \code{retry} times if the JSON request fails, since the Water Atlas API intermittently fails.
}
\examples{
\dontrun{
# get training data
transect <- read_transect(training = TRUE)

# import all transect data
transect <- read_transect()
}
}
\concept{read}
