% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_trnjsn.R
\name{read_trnjsn}
\alias{read_trnjsn}
\title{Import JSON transect data from Water Atlas}
\usage{
read_trnjsn(training = FALSE)
}
\arguments{
\item{training}{logical if training data are imported or the complete database}
}
\value{
data frame
}
\description{
Import JSON transect data from Water Atlas
}
\details{
The function imports a JSON file from the USF Water Atlas.  If \code{training = TRUE}, a dataset from the TBEP training survey is imported from \url{http://dev.seagrass.wateratlas.usf.edu/api/assessments/training}.  If \code{training = FALSE}, the entire transect survey database is imported from \url{http://dev.seagrass.wateratlas.usf.edu/api/assessments/all__use-with-care}.
}
\examples{
\dontrun{
# get training data
trnjsn <- read_trnjsn(training = TRUE)

# import all transect data
trnjsn <- read_trnjsn()
}
}
