% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sedimentdata.R
\docType{data}
\name{sedimentdata}
\alias{sedimentdata}
\title{Sediment data for the Tampa Bay current as of 20241212}
\format{
A \code{data.frame} with 226536 rows and 24 variables:
\describe{
\item{ProgramId}{int}
\item{ProgramName}{chr}
\item{FundingProject}{chr}
\item{yr}{int}
\item{AreaAbbr}{chr}
\item{StationID}{int}
\item{StationNumber}{chr}
\item{Latitude}{num}
\item{Longitude}{num}
\item{Replicate}{chr}
\item{SedResultsType}{chr}
\item{Parameter}{chr}
\item{ValueAdjusted}{num}
\item{Units}{chr}
\item{Qualifier}{chr}
\item{MDLnum}{num}
\item{PQLnum}{num}
\item{TEL}{num}
\item{PEL}{num}
\item{BetweenTELPEL}{chr}
\item{ExceedsPEL}{chr}
\item{PreparationDate}{chr}
\item{AnalysisTimeMerge}{chr}
\item{PELGrade}{Factor}
}
}
\usage{
sedimentdata
}
\description{
Sediment data for the Tampa Bay current as of 20241212
}
\examples{
\dontrun{
# location to download data
path <- '~/Desktop/sediment.zip'

# load and assign to object
sedimentdata <- read_importsediment(path, download_latest = TRUE, remove = TRUE)

save(sedimentdata, file = 'data/sedimentdata.RData', compress = 'xz')

}
}
\concept{data}
\keyword{datasets}
