% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_annualassess.R
\name{show_annualassess}
\alias{show_annualassess}
\title{Create a table for the annual management outcome assessments}
\usage{
show_annualassess(
  epcdata,
  yrsel,
  partialyr = F,
  caption = F,
  family = "Arial",
  txtsz = 12,
  width = NULL
)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrsel}{numeric indicating chosen year}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter}

\item{caption}{logical indicating if a caption is added using \code{\link[flextable]{set_caption}}}

\item{family}{chr string indicating font family for text labels}

\item{txtsz}{numeric indicating font size}

\item{width}{optional numeric value indicating width in inches}
}
\value{
A \code{\link[flextable]{flextable}} object showing the segment-averaged chlorophyll-a and light attenuation for the selected year, with bay segment names colored by the management outcome used in \code{\link{show_matrix}}.
}
\description{
Create a table for the annual management outcome assessments for chlorophyll-a and light attenuation by bay segment
}
\examples{
show_annualassess(epcdata, yrsel = 2024)
show_annualassess(epcdata, yrsel = 2024, caption = TRUE)
}
\concept{show}
