% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_boxplot.R
\name{show_boxplot}
\alias{show_boxplot}
\title{Plot monthly chlorophyll values and thresholds for a segment}
\usage{
show_boxplot(epcdata, bay_segment = c("OTB", "HB", "MTB", "LTB"),
  trgs = NULL, yrrng = c(1975, 2018))
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{bay_segment}{chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{yrcur}{numeric for current year to emphasize, shown as separate red points on the plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot monthly chlorophyll values and thresholds for a bay segment
}
\examples{
show_boxplot(epcdata, bay_segment = 'OTB')
}
\seealso{
Other visualize: \code{\link{show_chlmatrix}},
  \code{\link{show_matrix}}, \code{\link{show_thrplot}}
}
\concept{visualize}
