% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_chlmatrix.R
\name{show_chlmatrix}
\alias{show_chlmatrix}
\title{Create a colorized table for chlorophyll exceedances}
\usage{
show_chlmatrix(epcdata, txtsz = 3, trgs = NULL, yrrng = c(1975,
  2018))
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{tab = FALSE}}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{yrrng}{numeric vector indicating min, max years to include}
}
\value{
A static \code{ggplot} object is returned
}
\description{
Create a colorized table for chlorophyll exceedances
}
\examples{
show_chlmatrix(epcdata)
}
\seealso{
\code{\link{show_matrix}}

Other visualize: \code{\link{show_boxplot}},
  \code{\link{show_matrix}}, \code{\link{show_thrplot}}
}
\concept{visualize}
