% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colorizefld.R
\name{show_colorizefld}
\alias{show_colorizefld}
\title{Colorize a field with gt}
\usage{
show_colorizefld(gtbl, fld, pal = c("green", "yellow", "red"),
  digits = 1)
}
\arguments{
\item{gtbl}{input \code{gt} object}

\item{fld}{which field to colorize}

\item{pal}{chr vector of colors}

\item{digits}{precision of digits to show}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Colorize a field with gt
}
\details{
used internally in \code{\link{show_colorizetbl}}
}
