% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_colorizetbl.R
\name{show_colorizetbl}
\alias{show_colorizetbl}
\title{Colorize multiple fields with gt}
\usage{
show_colorizetbl(gtbl, exclude_cols = "Year", pal = c("green",
  "yellow", "red"))
}
\arguments{
\item{gtbl}{input \code{gt} object}

\item{exclude_cols}{which field(s) not to colorize}

\item{pal}{chr vector of colors}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Colorize multiple fields with gt
}
\details{
used internally in \code{\link{show_matrix}}
}
