% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_enteromap.R
\name{show_enteromap}
\alias{show_enteromap}
\title{Map Enterococcus results by month, year, and location}
\usage{
show_enteromap(fibdata, yrsel, mosel, wetdry = FALSE)
}
\arguments{
\item{fibdata}{data frame of Enterococcus sample data as returned by \code{\link{enterodata}} or \code{\link{anlz_fibwetdry}}}

\item{yrsel}{optional numeric to filter data by year}

\item{mosel}{optional numeric to filter data by month}

\item{wetdry}{logical; if \code{TRUE}, incorporate wet/dry differences. if \code{FALSE} (default), do not differentiate between wet and dry samples.}
}
\value{
A \code{leaflet} map for the selected year, month, and area showing stations and FIB concentration category
}
\description{
Map Enterococcus results by month, year, and location
}
\details{
Placing the mouse cursor over an item on the map will reveal additional information about a station.
}
\examples{
\dontrun{
show_enteromap(enterodata, yrsel = 2020, mosel = 9)

# wet/dry samples
entero_wetdry <- anlz_fibwetdry(enterodata, catch_precip)
show_enteromap(entero_wetdry, yrsel = 2020, mosel = 9, wetdry = TRUE)

# this will give the same output as anlz_enteromap(enterodata)
show_enteromap(entero_wetdry, yrsel = 2020, mosel = 9, wetdry = FALSE)
}
}
\seealso{
\code{\link{anlz_enteromap}} for details on the categories
}
\concept{anlz}
