% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fibmatmap.R
\name{show_fibmatmap}
\alias{show_fibmatmap}
\title{Map Fecal Indicator Bacteria matrix results by year}
\usage{
show_fibmatmap(
  fibdata,
  yrsel,
  areasel,
  lagyr = 3,
  subset_wetdry = c("all", "wet", "dry"),
  precipdata = NULL,
  temporal_window = NULL,
  wet_threshold = NULL,
  addsta = FALSE,
  listout = FALSE,
  warn = TRUE
)
}
\arguments{
\item{fibdata}{input data frame as returned by \code{\link{read_importfib}},  \code{\link{read_importentero}}, or \code{\link{read_importwqp}}, see details}

\item{yrsel}{numeric value indicating the year to map}

\item{areasel}{vector of bay segment or area names to include, see details}

\item{lagyr}{numeric for year lag to calculate categories, see details}

\item{subset_wetdry}{character, subset data frame to only wet or dry samples as defined by \code{wet_threshold} and \code{temporal_window}? Defaults to \code{"all"}, which will not subset. If \code{"wet"} or \code{"dry"} is specified, \code{\link{anlz_fibwetdry}} is called using the further specified parameters, and the data frame is subsetted accordingly.}

\item{precipdata}{input data frame as returned by \code{\link{read_importrain}}. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object \code{\link{catchprecip}} has this data from 1995-2023 for select Enterococcus stations. If \code{NULL}, defaults to \code{\link{catchprecip}}.}

\item{temporal_window}{numeric; required if \code{subset_wetdry} is not \code{"all"}. number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)}

\item{wet_threshold}{numeric; required if \code{subset_wetdry} is not \code{"all"}. inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period}

\item{addsta}{logical to add station labels to the map, default \code{FALSE}}

\item{listout}{logical to return a list of simple feature objects for the data used in the \code{leaflet} map, default \code{FALSE}}

\item{warn}{logical to print warnings about stations with insufficient data, default \code{TRUE}}
}
\value{
A \code{leaflet} map for the selected year and area showing station matrix scores if \code{listout = FALSE} (default).  Bay segment scores are also shown if the input data are not from EPCHC. A list of simple feature objects is returned if \code{listout = TRUE}.
}
\description{
Map Fecal Indicator Bacteria matrix results by year
}
\details{
Placing the mouse cursor over an item on the map will reveal additional information about a segment or station.

If the input is from EPCHC (from \code{\link{read_importfib}}, i.e., \code{\link{fibdata}}), valid entries for \code{areasel} include 'Alafia River', 'Hillsborough River', 'Big Bend', 'Cockroach Bay', 'East Lake Outfall', 'Hillsborough Bay', 'Little Manatee River', 'Lower Tampa Bay', 'McKay Bay', 'Middle Tampa Bay', 'Old Tampa Bay', 'Palm River', 'Tampa Bypass Canal', and 'Valrico Lake'.  If the input data is from \code{\link{read_importentero}} (i.e., \code{\link{enterodata}})), valid entries for \code{areasel} include 'OTB', 'HB', 'MTB', 'LTB', 'BCB', and 'MR'.  If the input is from \code{\link{read_importwqp}} for Manatee County (21FLMANA_WQX, \code{\link{mancofibdata}}), Pasco County (21FLPASC_WQX, \code{\link{pascofibdata}}), Polk County (21FLPOLK_WQX, \code{\link{polcofibdata}}), or Hillsborough County Environmental Services Division (21FLHESD_WQX, \code{\link{hcesdfibdata}}) valid entries for \code{areasel} are any that are present in the \code{area} column for the respective input datasets.

Bay segment matrix categories can be shown if input data are from \code{\link{read_importentero}}).  Stations for these data were chosen specifically as downstream endpoints for each bay segment, whereas the other datasets are not appropriate for estimating bay segment outcomes.
}
\examples{
# non-EPCHC, non Manatee County data
show_fibmatmap(enterodata, yrsel = 2020, areasel = 'OTB')

# EPCHC data
show_fibmatmap(fibdata, yrsel = 2016,
   areasel = c("Hillsborough River", "Alafia River"))

# Manatee County data
show_fibmatmap(mancofibdata, yrsel = 2020,  areasel = 'Manatee River')
}
\seealso{
\code{\link{anlz_fibmatrix}} for details on the categories
}
\concept{show}
