% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ratab.R
\name{show_ratab}
\alias{show_ratab}
\title{Create a bay segment assessment table for the 2022-2026 reasonable assurance period}
\usage{
show_ratab(
  epcdata,
  yrsel,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  partialyr = F,
  outtxt1 = "All years below threshold so far, not necessary for NMC Actions 2-5",
  outtxt2 = "All years met threshold, not necessary for NMC Actions 3-5",
  outtxt3 =
    "Not necessary due to observed water quality and seagrass conditions in the bay segment",
  outtxt45 = "Not necessary when chlorophyll-*a* threshold met",
  txtsz = 13,
  width = NULL
)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrsel}{numeric indicating chosen year}

\item{bay_segment}{chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter}

\item{outtxt1}{optional text for NMC action 1, added to the outcome column}

\item{outtxt2}{optional text for NMC action 2, added to the outcome column}

\item{outtxt3}{optional text for NMC action 3, added to the outcome column}

\item{outtxt45}{optional text for NMC actions 4 and 5, added to the outcome column}

\item{txtsz}{numeric indicating font size}

\item{width}{optional numeric value indicating width in inches}
}
\value{
A \code{\link[flextable]{flextable}} object showing the reasonable assurance compliance of the bay segment for the selected year within the five-year period.
}
\description{
Create a bay segment assessment table for the 2022-2026 reasonable assurance period
}
\examples{
show_ratab(epcdata, yrsel = 2022, bay_segment = 'OTB')
}
\concept{show}
