% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_seagrasscoverage.R
\name{show_seagrasscoverage}
\alias{show_seagrasscoverage}
\title{Create a barplot of seagrass coverage over time in Tampa Bay}
\usage{
show_seagrasscoverage(
  seagrass,
  maxyr = 2024,
  family = "sans",
  extend = F,
  lastlab = T,
  covlab = T,
  axsbrk = c(0.08, 0.1)
)
}
\arguments{
\item{seagrass}{input \code{data.frame} included with the package as \code{\link{seagrass}}}

\item{maxyr}{numeric for maximum year to plot}

\item{family}{optional chr string indicating font family for text labels}

\item{extend}{logical indicating if x-axis should be extended to maximum year in \code{seagrass} even if \code{maxyr} is less than that}

\item{lastlab}{logical indicating if text label on \code{maxyr} should be included}

\item{covlab}{logical indicating if text label for coverage goal should be included}

\item{axsbrk}{numeric vector of length two indicating where the x-axis break occurs}
}
\value{
A grid object showing acres of total seagrass coverage in Tampa Bay over time.
}
\description{
Create a barplot of seagrass coverage over time in Tampa Bay
}
\details{
This function creates the flagship seagrass coverage graphic to report on coverage changes over time.  All data were pre-processed and included in the package as the \code{\link{seagrass}} dataset.  Original data are from the Southwest Florida Water Management District and available online at \link[https://data-swfwmd.opendata.arcgis.com/]{https://data-swfwmd.opendata.arcgis.com/}.  This function and the data used to create the plot are distinct from those used for the transect monitoring program.
}
\examples{
show_seagrasscoverage(seagrass)
}
\concept{show}
