% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sedimentmap.R
\name{show_sedimentmap}
\alias{show_sedimentmap}
\title{Make a map for sediment contaminants at stations in Tampa Bay}
\usage{
show_sedimentmap(
  sedimentdata,
  param,
  yrrng = c(1993, 2024),
  funding_proj = "TBEP",
  weight = 1.5
)
}
\arguments{
\item{sedimentdata}{input sediment \code{data.frame} as returned by \code{\link{read_importsediment}}}

\item{param}{chr string for which parameter to plot}

\item{yrrng}{numeric vector indicating min, max years to include, use single year for one year of data}

\item{funding_proj}{chr string for the funding project, one to many of "TBEP" (default), "TBEP-Special", "Apollo Beach", "Janicki Contract", "Rivers", "Tidal Streams"}

\item{weight}{numeric for outline width of station points on the map}
}
\value{
A \code{\link[leaflet]{leaflet}} object
}
\description{
Make a map for sediment contaminants at stations in Tampa Bay
}
\details{
The map shows sediment contaminant concentrations for the selected parameter relative to Threshold Effects Levels (TEL) and Potential Effects Levels (PEL), if available.  Green points show concentrations below the TEL, yellow points show concentrations between the TEL and PEL, and red points show concentrations above the PEL. The applicable TEL and PEL values for the parameter are indicated in the legend.  If TEL and PEL thresholds are not available, a map of the sediment concentrations is shown and a warning is returned to the console.
}
\examples{
show_sedimentmap(sedimentdata, param = 'Arsenic')
}
\concept{show}
