% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sedimentpelave.R
\name{show_sedimentpelave}
\alias{show_sedimentpelave}
\title{Plot summary of PEL averages by bay segment}
\usage{
show_sedimentpelave(
  sedimentdata,
  yrrng = c(1993, 2021),
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"),
  alph = 1,
  ylim = c(0, 0.4),
  lnsz = 1,
  base_size = 12
)
}
\arguments{
\item{sedimentdata}{input sediment \code{data.frame} as returned by \code{\link{read_importsediment}}}

\item{yrrng}{numeric vector indicating min, max years to include, use single year for one year of data}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{alph}{numeric indicating alpha value for score category colors}

\item{ylim}{numeric for y axis limits}

\item{lnsz}{numeric for line size}

\item{base_size}{numeric indicating text scaling size for plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing PEL averages (average of the averages) and 95% confidence intervals for each bay segment
}
\description{
Plot summary of PEL averages by bay segment
}
\details{
Lines for the grades are shown on the plot. Confidence intervals may not be shown for segments with insufficient data.
}
\examples{
show_sedimentpelave(sedimentdata)
}
\concept{show}
