% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sedimentpelaveplotly.R
\name{show_sedimentpelaveplotly}
\alias{show_sedimentpelaveplotly}
\title{Creates a plotly object for average PEL plots}
\usage{
show_sedimentpelaveplotly(p, width = NULL, height = NULL)
}
\arguments{
\item{p}{\code{\link[ggplot2]{ggplot}} object as output from \code{\link{show_sedimentpelave}}}

\item{width}{numeric for width of the plot in pixels}

\item{height}{numeric for height of the plot in pixels}
}
\value{
A \code{\link[plotly]{plotly}} data object
}
\description{
Creates a plotly object for average PEL plots
}
\examples{
p <- show_sedimentpelave(sedimentdata)
show_sedimentpelaveplotly(p)
}
\concept{show}
