% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sedimentpelmap.R
\name{show_sedimentpelmap}
\alias{show_sedimentpelmap}
\title{Make a map for sediment PEL ratios at stations in Tampa Bay}
\usage{
show_sedimentpelmap(
  sedimentdata,
  yrrng = c(1993, 2024),
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"),
  funding_proj = "TBEP",
  weight = 1.5
)
}
\arguments{
\item{sedimentdata}{input sediment \code{data.frame} as returned by \code{\link{read_importsediment}}}

\item{yrrng}{numeric vector indicating min, max years to include, use single year for one year of data}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{funding_proj}{chr string for the funding project, one to many of "TBEP" (default), "TBEP-Special", "Apollo Beach", "Janicki Contract", "Rivers", "Tidal Streams"}

\item{weight}{numeric for outline width of station points on the map}
}
\value{
A \code{\link[leaflet]{leaflet}} object
}
\description{
Make a map for sediment PEL ratios at stations in Tampa Bay
}
\details{
The map shows average PEL ratios graded from A to F for benthic stations monitored in Tampa Bay. The PEL is a measure of how likely a contaminant is to have a toxic effect on invertebrates that inhabit the sediment. The PEL ratio is the contaminant concentration divided by the Potential Effects Levels (PEL) that applies to a contaminant, if available. Higher ratios and lower grades indicate sediment conditions that are likely unfavorable for invertebrates. The station average combines the PEL ratios across all contaminants measured at a station.
}
\examples{
show_sedimentpelmap(sedimentdata)
}
\concept{show}
