% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_segmatrix.R
\name{show_segmatrix}
\alias{show_segmatrix}
\title{Create a colorized table for water quality outcomes and exceedances by segment}
\usage{
show_segmatrix(
  epcdata,
  txtsz = 3,
  trgs = NULL,
  yrrng = c(1975, 2023),
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  abbrev = FALSE,
  family = NA,
  historic = TRUE,
  plotly = FALSE,
  partialyr = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{tab = FALSE}}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{bay_segment}{chr string for bay segments to include, only one of "OTB", "HB", "MTB", "LTB"}

\item{abbrev}{logical indicating if text labels in the plot are abbreviated as the first letter}

\item{family}{optional chr string indicating font family for text labels}

\item{historic}{logical if historic data are used from 2005 and earlier}

\item{plotly}{logical if matrix is created using plotly}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned
}
\description{
Create a colorized table for water quality outcomes by segment that includes the management action and chlorophyll, and light attenuation exceedances
}
\details{
This function provides a combined output for the \code{\link{show_wqmatrix}} and \code{\link{show_matrix}} functions. Only one bay segment can be plotted for each function call.
}
\examples{
show_segmatrix(epcdata, bay_segment = 'OTB')
}
\seealso{
\code{\link{show_wqmatrix}}, \code{\link{show_matrix}}
}
\concept{show}
