% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sitemap.R
\name{show_sitemap}
\alias{show_sitemap}
\title{Map site attainment categories for a selected year}
\usage{
show_sitemap(epcdata, yrsel, trgs = NULL, thrs = FALSE, partialyr = FALSE)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrsel}{numeric for year to plot}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{thrs}{logical indicating if attainment category is relative to targets (default) or thresholds, passed to \code{\link{anlz_attainsite}}}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter}
}
\value{
A static \code{ggplot} object is returned
}
\description{
Map site attainment categories for a selected year
}
\examples{
show_sitemap(epcdata, yrsel = 2019)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
