% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sitesegmap.R
\name{show_sitesegmap}
\alias{show_sitesegmap}
\title{Map site and bay segment attainment categories for a selected year}
\usage{
show_sitesegmap(
  epcdata,
  yrsel,
  param = c("chla", "la"),
  trgs = NULL,
  thrs = FALSE,
  partialyr = FALSE,
  base_size = 12,
  family = NA
)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrsel}{numeric for year to plot}

\item{param}{chr string for which parameter to plot, one of \code{"chla"} for chlorophyll or \code{"la"} for light attenuation}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{thrs}{logical indicating if attainment category is relative to targets (default) or thresholds, passed to \code{\link{anlz_attainsite}}}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter}

\item{base_size}{numeric indicating text scaling size for plot}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{ggplot} object is returned
}
\description{
Map site and bay segment attainment categories for a selected year
}
\details{
The map is similar to that returned by \code{\link{show_sitemap}} with the addition of polygons for each bay segment colored by the annual attainment category and the site points are sized relative to the selected parameter in \code{param}.
}
\examples{
show_sitesegmap(epcdata, yrsel = 2022)
}
\concept{show}
