% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbbimatrix.R
\name{show_tbbimatrix}
\alias{show_tbbimatrix}
\title{Plot a matrix of Tampa Bay Benthic Index scores over time by bay segment}
\usage{
show_tbbimatrix(
  tbbiscr,
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"),
  yrrng = c(1993, 2024),
  window = TRUE,
  alph = 1,
  txtsz = 3,
  family = "sans",
  rev = FALSE,
  position = "top",
  plotly = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{tbbiscr}{input data frame as returned by \code{\link{anlz_tbbiscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{window}{logical indicating whether to use a rolling 5-year window (default TRUE) or single year values (FALSE) for the bay segment categories, see details}

\item{alph}{numeric indicating alpha value for score category colors}

\item{txtsz}{numeric for size of text in the plot}

\item{family}{optional chr string indicating font family for text labels}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{position}{chr string of location for bay segment labels, default on top, passed to \code{\link[ggplot2]{scale_x_discrete}}}

\item{plotly}{logical if matrix is created using plotly}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBBI scores for each bay segment if \code{plotly = FALSE}, otherwise a \code{\link[plotly]{plotly}} object
}
\description{
Plot a matrix of Tampa Bay Benthic Index scores over time by bay segment
}
\details{
Additional summaries are provided for the entire bay, as a summary across categories ("All") and a summary weighted across the relative sizes of each bay segment ("All (wt)").

The default behavior is to use a rolling five-year window to calculate the percent of sites in each TBBI category by bay segment. This applies only to years 2005 and later, where the counts from the current year and the prior four years are summed to calculate the percentages. This is intended to help smooth out inter-annual variability due to reduced sampling effort from 2005 to present. If \code{window = FALSE}, then only single year values are used.
}
\examples{
tbbiscr <- anlz_tbbiscr(benthicdata)
show_tbbimatrix(tbbiscr)
}
\concept{show}
