% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbniscr.R
\name{show_tbniscr}
\alias{show_tbniscr}
\title{Plot Tampa Bay Nekton Index scores over time by bay segment}
\usage{
show_tbniscr(
  tbniscr,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  perc = c(32, 46),
  alph = 1,
  ylim = c(22, 58),
  rev = FALSE,
  plotly = FALSE,
  family = "sans",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{tbniscr}{input dat frame as returned by \code{\link{anlz_tbniscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB"}

\item{perc}{numeric values indicating break points for score categories}

\item{alph}{numeric indicating alpha value for score category colors}

\item{ylim}{numeric for y axis limits}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{plotly}{logical if matrix is created using plotly}

\item{family}{optional chr string indicating font family for text labels}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBNI scores for each bay segment or a \code{\link[plotly]{plotly}} object if \code{plotly = TRUE}
}
\description{
Plot Tampa Bay Nekton Index scores over time by bay segment
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
show_tbniscr(tbniscr)
}
\concept{show}
