% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbniscrall.R
\name{show_tbniscrall}
\alias{show_tbniscrall}
\title{Plot Tampa Bay Nekton Index scores over time as average across bay segments}
\usage{
show_tbniscrall(
  tbniscr,
  perc = c(32, 46),
  alph = 0.3,
  ylim = c(22, 58),
  rev = FALSE,
  plotly = FALSE
)
}
\arguments{
\item{tbniscr}{input dat frame as returned by \code{\link{anlz_tbniscr}}}

\item{perc}{numeric values indicating break points for score categories}

\item{alph}{numeric indicating alpha value for score category colors}

\item{ylim}{numeric for y axis limits}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{plotly}{logical if matrix is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBNI scores for each bay segment or a \code{\link[plotly]{plotly}} object if \code{plotly = TRUE}
}
\description{
Plot Tampa Bay Nekton Index scores over time as average across bay segments
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
show_tbniscrall(tbniscr)
}
\concept{show}
