% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrk.R
\name{show_tdlcrk}
\alias{show_tdlcrk}
\title{Make a map for tidal creek report card}
\usage{
show_tdlcrk(dat, weight = 1.5)
}
\arguments{
\item{dat}{input creek score data returned from \code{\link{anlz_tdlcrk}}}

\item{weight}{numeric for weight of polylines, passed to \code{\link[leaflet]{addPolylines}}}
}
\value{
A \code{\link[leaflet]{leaflet}} object
}
\description{
Make a map for tidal creek report card
}
\examples{
dat <- anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2018)
show_tdlcrk(dat)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_thrplot}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
