% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkindic.R
\name{show_tdlcrkindic}
\alias{show_tdlcrkindic}
\title{Plotly barplots of tidal creek context indicators}
\usage{
show_tdlcrkindic(id, cntdat, yr = 2018, thrsel = FALSE)
}
\arguments{
\item{id}{numeric indicating the \code{id} number of the tidal creek to plot}

\item{cntdat}{output from \code{\link{anlz_tdlcrkindic}}}

\item{yr}{numeric indicating reference year, almost always 2018}

\item{thrsel}{logical if threshold lines and annotations are shown on the plots}
}
\value{
A plotly object
}
\description{
Plotly barplots of tidal creek context indicators
}
\examples{
cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
show_tdlcrkindic(495, cntdat, thrsel = TRUE)
}
