% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkindiccdf.R
\name{show_tdlcrkindiccdf}
\alias{show_tdlcrkindiccdf}
\title{Plotly empirical CDF plots of tidal creek context indicators}
\usage{
show_tdlcrkindiccdf(
  id,
  cntdat,
  yr = 2022,
  thrsel = FALSE,
  pal = c("#5C4A42", "#427355", "#004F7E")
)
}
\arguments{
\item{id}{numeric indicating the \code{id} number of the tidal creek to plot}

\item{cntdat}{output from \code{\link{anlz_tdlcrkindic}}}

\item{yr}{numeric indicating reference year}

\item{thrsel}{logical if threshold lines and annotations are shown on the plots}

\item{pal}{vector of colors for the palette}
}
\value{
A plotly object
}
\description{
Plotly empirical CDF plots of tidal creek context indicators
}
\details{
This function returns several empirical cumulative distribution plots for the tidal creek context indicators.  Points on the plot indicate the observed values and percentiles for the creek specified by \code{id}. The percentiles and CDF values are defined by the "population" of creeks in \code{cntdat}.  Points in the plots are color-coded by sample year to evaluate temporal trends, if any.
}
\examples{
cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2022)

set.seed(123)
id <- sample(unique(cntdat$id), 1)
show_tdlcrkindiccdf(id, cntdat, thrsel = TRUE)
}
\concept{show}
