% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkline.R
\name{show_tdlcrkline}
\alias{show_tdlcrkline}
\title{Add a line or annotation to a plotly graph}
\usage{
show_tdlcrkline(
  varin = c("CHLAC", "TN", "chla_tn_ratio", "DO", "tsi", "no23_ratio"),
  thrsel = FALSE,
  horiz = TRUE,
  annotate = FALSE
)
}
\arguments{
\item{varin}{chr string for the indicator}

\item{thrsel}{logical if something is returned, otherwise NULL, this is a hack for working with the plotly output}

\item{horiz}{logical indicating if output is horizontal or vertical}

\item{annotate}{logical indicating if output is line or annotation text}
}
\value{
A list object passed to the layout argument of plotly, either shapes or annotate depending on user input
}
\description{
Add a line or annotation to a plotly graph for the tidal creek indicators
}
\details{
This function is used internally within \code{\link{show_tdlcrkindic}} and \code{\link{show_tdlcrkindiccdf}}
}
\examples{
# code for vertical line output, chloropyll
show_tdlcrkline('CHLAC', thrsel = TRUE)
}
\concept{show}
