% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkmatrix.R
\name{show_tdlcrkmatrix}
\alias{show_tdlcrkmatrix}
\title{Plot the tidal creek report card matrix}
\usage{
show_tdlcrkmatrix(
  dat,
  class = c("3M", "2"),
  score = c("Prioritize", "Investigate", "Caution", "Monitor"),
  family = NA
)
}
\arguments{
\item{dat}{input creek score data returned from \code{\link{anlz_tdlcrk}}}

\item{class}{character vector indicating which creek classes to show, one to many of \code{'3M'}, \code{'2'}, \code{'3F'}, and \code{'1'}.  Defaults to marine only (\code{'3M', '2'}).}

\item{score}{character vector of score categories to include, one to many of \code{'Prioritize'}, \code{'Investigate'}, \code{'Caution'}, and \code{'Monitor'}. Defaults to all.}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned.
}
\description{
Plot the tidal creek report card matrix
}
\details{
The plot shows a matrix with rows for individual creeks and columns for overall creek score.  The columns show an overall creek score and the number of years in the prior ten years that nitrogen values at a creek were assigned to each of the four score categories.  Number of years is mapped to cell transparency.
}
\examples{
dat <- anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2018)
show_tdlcrkmatrix(dat)
}
