% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkradar.R
\name{show_tdlcrkradar}
\alias{show_tdlcrkradar}
\title{Radar plots for tidal creek indicators}
\usage{
show_tdlcrkradar(id, cntdat)
}
\arguments{
\item{id}{numeric indicating the \code{id} number of the tidal creek to plot}

\item{cntdat}{output from \code{\link{anlz_tdlcrkindic}}}
}
\value{
A radar plot
}
\description{
Radar plots for tidal creek indicators
}
\details{
See details in \code{\link{anlz_tldcrkindic}} for an explanation of the indicators
}
\examples{
cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018, rader = TRUE)
show_tdlcrkradar(495, cntdat)
}
\concept{show}
