% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\docType{data}
\name{stations}
\alias{stations}
\title{Bay stations by segment}
\format{A data frame with 75 rows and 2 variables:
\describe{
  \item{bay_segment}{chr}
  \item{ecphc_station}{num}
}}
\usage{
stations
}
\description{
Bay stations by segment
}
\keyword{datasets}
