% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtacres.R
\docType{data}
\name{subtacres}
\alias{subtacres}
\title{Tampa Bay subtidal cover}
\format{
A data frame
}
\usage{
subtacres
}
\description{
Tampa Bay subtidal cover for Habitat Master Plan reporting
}
\examples{
\dontrun{
load(url("https://github.com/tbep-tech/hmpu-workflow/raw/master/data/subtacres.RData"))

save(subtacres, file = 'data/subtacres.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
