% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\docType{data}
\name{targets}
\alias{targets}
\title{Bay segment targets}
\format{
A data frame with 4 rows and 8 variables:
\describe{
  \item{bay_segment}{chr}
  \item{name}{chr}
  \item{chla_target}{num}
  \item{chla_smallex}{num}
  \item{chla_thresh}{num}
  \item{la_target}{num}
  \item{la_smallex}{num}
  \item{la_thresh}{num}
}
}
\usage{
targets
}
\description{
Bay segment specific management targets including low and high magnitude exceedance thresholds
}
\examples{
\dontrun{

targets <- structure(list(
   bay_segment = c("OTB", "HB", "MTB", "LTB", "BCBN", "BCBS", "TCB", "MR"),
   name = c("Old Tampa Bay", "Hillsborough Bay", "Middle Tampa Bay", "Lower Tampa Bay",
      "Boca Ciega Bay North", "Boca Ciega Bay South", "Terra Ceia Bay", "Manatee River"),
   chla_target = c(8.5, 13.2, 7.4, 4.6, 7.7, 6.1, 7.5, 7.3),
   chla_smallex = c(8.9, 14.1, 7.9, 4.8, NaN, NaN, NaN, NaN),
   chla_thresh = c(9.3, 15, 8.5, 5.1, 8.3, 6.3, 8.7, 8.8),
   la_target = c(0.83, 1.58, 0.83, 0.63, NaN, NaN, NaN, NaN),
   la_smallex = c(0.86, 1.63, 0.87, 0.66, NaN, NaN, NaN, NaN),
   la_thresh = c(0.88, 1.67, 0.91, 0.68, NaN, NaN, NaN, NaN)),
   class = "data.frame", row.names = c(NA, -8L)
   )

save(targets, file = 'data/targets.RData')
}
}
\keyword{datasets}
