% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbseg.R
\docType{data}
\name{tbseg}
\alias{tbseg}
\title{Spatial data object of Tampa Bay segments}
\format{A simple features \code{\link[sf]{sf}} object (POLYGON) with 4 feaures and 2 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
\describe{
  \item{long_name}{chr}
  \item{bay_segment}{chr}
}}
\usage{
tbseg
}
\description{
Spatial data object of Tampa Bay segments
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{stations}},
  \code{\link{targets}}
}
\concept{utilities}
\keyword{datasets}
