% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbsegshed.R
\docType{data}
\name{tbsegshed}
\alias{tbsegshed}
\title{Spatial data object of Tampa Bay segments plus watersheds}
\format{
A simple features \code{\link[sf]{sf}} object (POLYGON) with 7 features and 2 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
\describe{
  \item{long_name}{chr}
  \item{bay_segment}{chr}
}
}
\usage{
tbsegshed
}
\description{
Spatial data object of Tampa Bay segments plus waterhseds
}
\examples{
library(sf)
plot(st_geometry(tbsegshed))
}
\concept{data}
\keyword{datasets}
