% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-epcdata.R
\docType{methods}
\name{thrplot}
\alias{thrplot}
\alias{thrplot,epcdata-method}
\title{Plot annual water quality values and thresholds for a segment}
\usage{
thrplot(object, bay_segment = c("OTB", "HB", "MTB", "LTB"),
  thr = c("chl", "la"))

\S4method{thrplot}{epcdata}(object, bay_segment = c("OTB", "HB", "MTB",
  "LTB"), thr = c("chla", "la"))
}
\arguments{
\item{object}{\code{\link{epcdata}} object}

\item{bay_segment}{chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"}

\item{thr}{chr string indicating with water quality value and appropriate threshold to to plot, one of "chl" for chlorophyll and "la" for light availability}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object
}
\description{
Plot annual water quality values and thresholds for a bay segment
}
\examples{
\dontrun{
thrplot(epcdata, bay_segment = 'OTB', thr = 'chl')
}
}
