% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidaltargets.R
\docType{data}
\name{tidaltargets}
\alias{tidaltargets}
\title{Tidal creek nitrogen targets}
\format{
A data frame with 2 rows and 4 variables:
\describe{
  \item{region}{chr}
  \item{prioritize}{num}
  \item{investigate}{num}
  \item{caution}{num}
}
}
\usage{
tidaltargets
}
\description{
Tidal creek nitrogen targets
}
\keyword{datasets}
