% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transect.R
\docType{data}
\name{transect}
\alias{transect}
\title{Seagrass transect data for Tampa Bay current as of 11202020}
\format{
A data frame with 132354 rows and 11 variables:
\describe{
  \item{Crew}{chr}
  \item{MonitoringAgency}{chr}
  \item{Date}{Date}
  \item{Transect}{chr}
  \item{Site}{chr}
  \item{Depth}{int}
  \item{Savspecies}{chr}
  \item{SeagrassEdge}{num}
  \item{var}{chr}
  \item{aveval}{num}
  \item{sdval}{num}
  }
}
\usage{
transect
}
\description{
Seagrass transect data for Tampa Bay current as of 11202020
}
\examples{
\dontrun{

transect <- read_transect()

save(transect, file = 'data/transect.RData', compress = 'xz')

}
}
\seealso{
Other utilities: 
\code{\link{benthicdata}},
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{fimstations}},
\code{\link{iwrraw}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
