% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnlns.R
\docType{data}
\name{trnlns}
\alias{trnlns}
\title{Seagrass transect locations}
\format{
A \code{sf} LINESTRING object
}
\usage{
trnlns
}
\description{
Seagrass transect locations
}
\details{
The script used to create this dataset is in \code{data-raw/trnlns-raw.R}.
}
\concept{data}
\keyword{datasets}
