% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trnpts.R
\docType{data}
\name{trnpts}
\alias{trnpts}
\title{Seagrass transect starting locations}
\format{
A \code{sf} POINT object
}
\usage{
trnpts
}
\description{
Seagrass transect starting locations
}
\details{
The script used to create this dataset is in \code{data-raw/trnpts-raw.R}.
}
\concept{data}
\keyword{datasets}
