% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dateseq.R
\name{util_dateseq}
\alias{util_dateseq}
\title{Generate a two-column data frame of dates}
\usage{
util_dateseq(start_date, end_date)
}
\arguments{
\item{start_date}{character for the starting date in "YYYY-MM-DD" format}

\item{end_date}{character for the ending date in "YYYY-MM-DD" format}
}
\value{
A data frame with two columns: \code{start} and \code{end}, containing dates.
}
\description{
Generate a two-column data frame of dates
}
\details{
A sequence of dates is generated from the start to end date that includes monthly breaks, such that the first and last day of each month between the start and end dates is returned.  This function is used within \code{\link{read_importwqwin}} to create separate API requests in approximate monthly breaks.
}
\examples{
util_dateseq('2023-01-15', '2023-05-04')
}
\concept{util}
