% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fibicons.R
\name{util_fibicons}
\alias{util_fibicons}
\title{Return leaflet icon set for FIB maps}
\usage{
util_fibicons(indic)
}
\arguments{
\item{indic}{character indicating \code{"entero"} or \code{"fcolif"} for \emph{Enterococcus} or Fecal Coliform}
}
\value{
A leaflet icon set as returned by \code{\link[leaflet]{iconList}}.
}
\description{
Return leaflet icon set for FIB maps
}
\details{
Used internally with \code{\link{show_enteromap}} and \code{\link{show_fibmap}}, former uses wet/dry icons for \emph{Enterococcus} and latter uses \emph{E. Coli}/\emph{Enterococcus} icons
}
\examples{
util_fibicons(indic = 'entero')
util_fibicons(indic = 'fcolif')
}
