% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_html.R
\name{util_html}
\alias{util_html}
\title{Convert character string to html class}
\usage{
util_html(text)
}
\arguments{
\item{text}{character string input}
}
\value{
The same input character string with html class
}
\description{
Convert character string to html class
}
\details{
Adapted from \code{\link[htmltools]{HTML}}
}
\examples{
util_html('abd')
}
\concept{util}
