% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_importwqwa.R
\name{util_importwqwa}
\alias{util_importwqwa}
\title{Retrieve metadata from the Water Atlas API}
\usage{
util_importwqwa(endpoint, pageSize = 10000, waterbodyId = NULL)
}
\arguments{
\item{endpoint}{Character indicating which API endpoint to retrieve, see details}

\item{pageSize}{Numeric indicating the number of records to retrieve}

\item{waterbodyId}{Optional numeric indicating a specific waterbody ID if requesting sampling locations}
}
\value{
A data frame of results for the given endpoint
}
\description{
Retrieve metadata from the Water Atlas API
}
\details{
Endpoints include \code{"dataSources"}, \code{"parameters"}, \code{"sampling-locations"}, or \code{"waterbodies"}.

Data are retrieved in pages if the total number of records exceeds the page size.
}
\examples{
# get available parameters
util_importwqwa('parameters')

# get sampling locations for a specific waterbody
waterbodies <- util_importwqwa('waterbodies')
waterbodyid <- waterbodies |> 
  dplyr::filter(grepl('Hillsborough Bay', name)) |>
  dplyr::pull(id)
util_importwqwa('sampling-locations', waterbodyId = waterbodyid)
}
