% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_map.R
\name{util_map}
\alias{util_map}
\title{Create an empty leaflet map from sf input}
\usage{
util_map(tomap, minimap = "bottomleft")
}
\arguments{
\item{tomap}{\code{sf} input object}

\item{minimap}{character string indicating location of minimap, use \code{minimap = NULL} to suppress}
}
\value{
A \code{leaflet} object with optional minimap and ESRI provider tiles
}
\description{
Create an empty leaflet map from sf input
}
\examples{
tomap <- tibble::tibble(
  lon = -82.6365,
  lat = 27.75822
  )
tomap <- sf::st_as_sf(tomap, coords = c('lon', 'lat'), crs = 4326)
util_map(tomap)
}
\concept{util}
