% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_orgin.R
\name{util_orgin}
\alias{util_orgin}
\title{Get county name from organization identifier}
\usage{
util_orgin(org, stanm = FALSE)
}
\arguments{
\item{org}{chr string indicating the organization identifier, see \code{\link{read_importwqp}} for valid entries}

\item{stanm}{logical indicating if a character string for a column name specific to the county is returned}
}
\value{
A character string of the county name that correspond to the organization identifier or a column name for the station identfier specific to the county if \code{stanm = TRUE}
}
\description{
Get county name from organization identifier
}
\examples{
util_orgin('21FLHILL_WQX')
util_orgin('21FLPASC_WQX', stanm = TRUE)
}
\concept{util}
