% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_path_ext_remove.R
\name{util_path_ext_remove}
\alias{util_path_ext_remove}
\title{Remove file extension from path TODO not exported yet b/c internal fs commands also need replacing}
\usage{
util_path_ext_remove(path)
}
\arguments{
\item{path}{}
}
\value{
A character vector of paths with the extension removed
}
\description{
Copied exactly from \code{fs::path_ext_remove()} to avoid dependency on \code{fs}.
}
\examples{
util_path_ext_remove("foo/bar.txt")
}
