% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_rain.R
\name{util_rain}
\alias{util_rain}
\title{Get rainfall data at NOAA NCDC sites}
\usage{
util_rain(station = NULL, start, end, token, ntry = 5, quiet = FALSE)
}
\arguments{
\item{station}{character for the station id to retrieve}

\item{start}{character for start of time period as YYYY-MM-DD}

\item{end}{character for end of time period as YYYY-MM-DD}

\item{token}{character for the NOAA API key}

\item{ntry}{numeric for the number of times to try to download the data}

\item{quiet}{logical to print progress in the console}
}
\value{
a single value data frame with the rainfall sum in inches
}
\description{
Get rainfall data at NOAA NCDC sites
}
\details{
This function is used to retrieve a long-term record of rainfall for the requested station.  A NOAA API key is required to use the function.
}
\examples{
\dontrun{
noaa_key <- Sys.getenv('NOAA_KEY')
util_rain('GHCND:USW00092806', start = '2021-01-01', end = '2021-12-31', noaa_key)
}
}
\seealso{
\code{\link{anlz_hydroload}}
}
