% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_attain.R
\name{anlz_attain}
\alias{anlz_attain}
\title{Get attainment categories}
\usage{
anlz_attain(avedat, magdurout = FALSE, trgs = NULL)
}
\arguments{
\item{avedat}{result returned from \code{\link{anlz_avedat}}}

\item{magdurout}{logical indicating if the separate magnitude and duration estimates are returned}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}
}
\value{
a \code{data.frame} for each year and bay segment showing the attainment category
}
\description{
Get attainment categories for each year and bay segment using chlorophyll and light attenuation
}
\examples{
\dontrun{
avedat <- anlz_avedat(epcdata)
anlz_attain(avedat)
}
}
