% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_attain.R
\name{anlz_attain}
\alias{anlz_attain}
\title{Get attainment categories}
\usage{
anlz_attain(avedat, magdurout = FALSE, trgs = NULL)
}
\arguments{
\item{avedat}{result returned from \code{\link{anlz_avedat}}}

\item{magdurout}{logical indicating if the separate magnitude and duration estimates are returned}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}
}
\value{
A \code{data.frame} for each year and bay segment showing the attainment category
}
\description{
Get attainment categories for each year and bay segment using chlorophyll and light attenuation
}
\examples{
avedat <- anlz_avedat(epcdata)
anlz_attain(avedat)
}
\seealso{
Other analyze: 
\code{\link{anlz_avedatsite}()},
\code{\link{anlz_avedat}()},
\code{\link{anlz_iwrraw}()},
\code{\link{anlz_refs}()},
\code{\link{anlz_tbbimed}()},
\code{\link{anlz_tdlcrkindic}()},
\code{\link{anlz_tdlcrk}()},
\code{\link{anlz_transectavespp}()},
\code{\link{anlz_transectave}()},
\code{\link{anlz_transectocc}()},
\code{\link{anlz_yrattain}()}
}
\concept{analyze}
