% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_avedat.R
\name{anlz_avedat}
\alias{anlz_avedat}
\title{Estimate annual means}
\usage{
anlz_avedat(datin)
}
\arguments{
\item{datin}{\code{data.frame} formatted from \code{read_importwq}}
}
\value{
Mean estimates for chlorophyll and secchi
}
\description{
Estimate annual means for chlorophyll and secchi data
}
\examples{
\dontrun{
# file path
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'

# load and assign to object
epcdata <- read_importwq(xlsx)

# view average estimates
anlz_avedat(epcdata)
}
}
\concept{analyze}
