% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_avedat.R
\name{anlz_avedat}
\alias{anlz_avedat}
\title{Estimate annual means}
\usage{
anlz_avedat(epcdata)
}
\arguments{
\item{epcdata}{\code{data.frame} formatted from \code{read_importwq}}
}
\value{
Mean estimates for chlorophyll and secchi
}
\description{
Estimate annual means for chlorophyll and secchi data
}
\examples{
# view average estimates
anlz_avedat(epcdata)
}
\seealso{
Other analyze: \code{\link{anlz_avedatsite}}
}
\concept{analyze}
