% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_fibmatrix.R
\name{anlz_fibmatrix}
\alias{anlz_fibmatrix}
\title{Analyze Fecal Indicator Bacteria categories over time by station}
\usage{
anlz_fibmatrix(
  fibdata,
  yrrng = NULL,
  stas = c(143, 108, 107, 135, 118, 148, 105, 152, 137)
)
}
\arguments{
\item{fibdata}{input data frame as returned by \code{\link{read_importfib}}}

\item{yrrng}{numeric vector indicating min, max years to include, defaults to range of years in \code{epcdata}}

\item{stas}{numeric vector of stations to include, default as those relevant for the Hillsborough River Basin Management Action Plan, see details}
}
\value{
A \code{\link[dplyr]{tibble}} object with FIB summaries by year and station, including columns for the estimated geometric mean of fecal coliform concentration (\code{gmean}), the proportions of samples exceeding 400 CFU / 100mL (\code{exced}), the count of samples (\code{cnt}), and a category indicating a letter outcome based on the proportion of exceedences (\code{cat}).
}
\description{
Analyze Fecal Indicator Bacteria categories over time by station
}
\details{
This function is used to create output for plotting a matrix stoplight graphic for FIB categories by station and year.  See \code{\link{show_fibmatrix}} for additional details.
}
\examples{
anlz_fibmatrix(fibdata)
}
\seealso{
\code{\link{show_fibmatrix}}
}
\concept{show}
