% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_fibmatrix.R
\name{anlz_fibmatrix}
\alias{anlz_fibmatrix}
\title{Analyze Fecal Indicator Bacteria categories over time by station}
\usage{
anlz_fibmatrix(
  fibdata,
  yrrng = NULL,
  stas = NULL,
  indic = c("fcolif", "ecocci"),
  threshold = NULL,
  lagyr = 3,
  subset_wetdry = c("all", "wet", "dry"),
  precipdata = NULL,
  temporal_window = NULL,
  wet_threshold = NULL
)
}
\arguments{
\item{fibdata}{input data frame as returned by \code{\link{read_importfib}} or \code{\link{read_importentero}}}

\item{yrrng}{numeric vector indicating min, max years to include, defaults to range of years in data}

\item{stas}{optional vector of stations to include, see details}

\item{indic}{character for choice of fecal indicator. Allowable options are \code{fcolif} (default) for fecal coliform, or \code{ecocci} for Enterococcus. A numeric column in the data frame must have this name.}

\item{threshold}{optional numeric for threshold against which to calculate exceedances for the indicator bacteria of choice. If not provided, defaults to 400 for \code{fcolif} and 130 for \code{ecocci}.}

\item{lagyr}{numeric for year lag to calculate categories, see details}

\item{subset_wetdry}{character, subset data frame to only wet or dry samples as defined by \code{wet_threshold} and \code{temporal_window}? Defaults to \code{"all"}, which will not subset. If \code{"wet"} or \code{"dry"} is specified, \code{\link{anlz_fibwetdry}} is called using the further specified parameters, and the data frame is subsetted accordingly.}

\item{precipdata}{input data frame as returned by \code{\link{read_importrain}}. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object \code{\link{catch_precip}} has this data from 1995-2023 for select Enterococcus stations.}

\item{temporal_window}{numeric; required if \code{subset_wetdry} is not \code{"all"}. number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)}

\item{wet_threshold}{numeric; required if \code{subset_wetdry} is not \code{"all"}. inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period}
}
\value{
A \code{\link[dplyr]{tibble}} object with FIB summaries by year and station, including columns for the estimated geometric mean of fecal coliform or Enterococcus concentrations (\code{gmean}), the proportion of samples exceeding 400 CFU / 100 mL (fecal coliform) or 130 CFU / 100 mL (Enterococcus) (\code{exced}), the count of samples (\code{cnt}), and a category indicating a letter outcome based on the proportion of exceedences (\code{cat}).
}
\description{
Analyze Fecal Indicator Bacteria categories over time by station
}
\details{
This function is used to create output for plotting a matrix stoplight graphic for FIB categories by station and year.  Each station and year combination is categorized based on the likelihood of fecal indicator bacteria concentrations exceeding some threshold in a given year.  For fecal coliform, the default threshold is 400 CFU / 100 mL in a given year (using Fecal Coliform, \code{fcolif} in \code{fibdata}).  For Enterococcus, the default threshold is 130 CFU / 100 mL.  The proportions are categorized as A, B, C, D, or E (Microbial Water Quality Assessment or MWQA categories) with corresponding colors, where the breakpoints for each category are <10\%, 10-30\%, 30-50\%, 50-75\%, and >75\% (right-closed).  By default, the results for each year are based on a right-centered window that uses the previous two years and the current year to calculate probabilities using the monthly samples (\code{lagyr = 3}). See \code{\link{show_fibmatrix}} for additional details.
}
\examples{
anlz_fibmatrix(fibdata)
# same result:
anlz_fibmatrix(fibdata, indic = 'fcolif')

# use different indicator:
anlz_fibmatrix(fibdata, indic = 'ecocci')

# use different dataset; does not contain an 'fcolif' column so we must specify indic:
anlz_fibmatrix(enterodata, indic = 'ecocci', lagyr = 1)
# same ecocci data; lower threshold - changes 'cat' scores
anlz_fibmatrix(enterodata, indic = 'ecocci', lagyr = 1, threshold = 30)

# subset to only wet samples
anlz_fibmatrix(enterodata, indic = 'ecocci', lagyr = 1, subset_wetdry = "wet",
               temporal_window = 2, wet_threshold = 0.5)

# subset to only dry samples
anlz_fibmatrix(enterodata, indic = 'ecocci', lagyr = 1, subset_wetdry = "dry",
               temporal_window = 2, wet_threshold = 0.5)
}
\seealso{
\code{\link{show_fibmatrix}}
}
\concept{show}
