% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_hmpreport.R
\name{anlz_hmpreport}
\alias{anlz_hmpreport}
\title{Evaluate Habitat Master Plan progress for report card}
\usage{
anlz_hmpreport(acres, subtacres, hmptrgs)
}
\arguments{
\item{acres}{\code{data.frame} for intertidal and supratidal land use and cover of habitat types for each year of data}

\item{subtacres}{\code{data.frame} for subtidal cover of habitat types for each year of data}

\item{hmptrgs}{\code{data.frame} of Habitat Master Plan targets and goals}
}
\value{
A summarized \code{data.frame} appropriate for creating a report card
}
\description{
Evaluate Habitat Master Plan progress for report card
}
\details{
The relevant output columns are \code{targeteval} and \code{goaleval} that indicate numeric values as -1 (target not met, trending below), 0 (target met, trending below), 0.5 (target not met, trending above), and 1 (target met, trending above).

Columns in the output are as follows:

\describe{
\item{year}{Year of the assessment}
\item{metric}{Habitat type assessed}
\item{Acres}{Coverage estimate for the year}
\item{lacres}{Coverage estimate for the previous set of available data}
\item{lyr}{Year for the previous set of available data}
\item{category}{Strata for the habitat type}
\item{Target}{2030 target for the habitat type from the Habitat Master Plan}
\item{Goal}{2050 goal for the habitat type from the Habitat Master Plan}
\item{acresdiff}{Difference in acres for the current year and the previous set of available data}
\item{yeardiff}{Difference in years for the current year and the previous set of available data}
\item{changerate}{Acreage change per year for the current year relative to the previous set of available data}
\item{targetrate}{Annual rate required to achieve the 2030 target}
\item{goalrate}{Annual rate required to achieve the 2050 goal}
\item{targetprop}{Proportion of target met for the current year}
\item{goalprop}{Proportion of goal met for the current year}
\item{targeteval}{A number indicating target status of the current year for the report card}
\item{goaleval}{A number indicating goal status of the current year for the report card}
}

The numbers in \code{targeteval} and \code{goaleval} are one of four values as -1, 0, 0.5, and 1. These numbers define the status for the assessment year:

\describe{
\item{-1}{target or goal not met, trending below}
\item{0}{target or goal met, trending below}
\item{0.5}{target or goal not met, trending above}
\item{1}{target or goal met, trending above}
}
}
\examples{
# view summarized data for report card
anlz_hmpreport(acres, subtacres, hmptrgs)
}
\concept{analyze}
