% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_iwrraw.R
\name{anlz_iwrraw}
\alias{anlz_iwrraw}
\title{Format raw IWR data}
\usage{
anlz_iwrraw(iwrraw, tidalcreeks, yr = 2018)
}
\arguments{
\item{iwrraw}{FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}}

\item{tidalcreeks}{\code{\link[sf]{sf}} object for population of tidal creeks}

\item{yr}{numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date}
}
\value{
A \code{\link{data.frame}} with the formatted data
}
\description{
Format raw IWR data
}
\details{
The function subsets the raw IWR data for the selected value in \code{yr} and the ten years prior to \code{yr} and subsets by the creek population in \code{\link{tidalcreeks}}. Select water quality parameters in \code{masterCode} are filtered and some of the names are combined for continuity.
}
\examples{
anlz_iwrraw(iwrraw, tidalcreeks, yr = 2018)
}
\concept{analyze}
