% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbbimed.R
\name{anlz_tbbimed}
\alias{anlz_tbbimed}
\title{Get annual medians of Tampa Bay Benthic Index scores by bay segment}
\usage{
anlz_tbbimed(
  tbbiscr,
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"),
  rev = FALSE,
  yrrng = c(1993, 2024),
  window = TRUE
)
}
\arguments{
\item{tbbiscr}{input data frame as returned by \code{\link{anlz_tbbiscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{window}{logical indicating whether to use a rolling 5-year window (default TRUE) or single year values (FALSE) for the bay segment categories, see details}
}
\value{
A data frame of annual medians by bay segment
}
\description{
Get annual medians of Tampa Bay Benthic Index scores by bay segment
}
\details{
Additional summaries are provided for the entire bay, as a summary across categories ("All") and a summary weighted across the relative sizes of each bay segment ("All (wt)").

Only sampling funded by TBEP and as part of the routine EPC benthic monitoring program are included in the final categories.

The default behavior is to use a rolling five-year window to calculate the percent of sites in each TBBI category by bay segment. This applies only to years 2005 and later, where the counts from the current year and the prior four years are summed to calculate the percentages. This is intended to help smooth out inter-annual variability due to reduced sampling effort from 2005 to present. If \code{window = FALSE}, then only single year values are used.
}
\examples{
tbbiscr <- anlz_tbbiscr(benthicdata)
anlz_tbbimed(tbbiscr)
}
\concept{analyze}
