% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbbimed.R
\name{anlz_tbbimed}
\alias{anlz_tbbimed}
\title{Get annual medians of Tampa Bay Benthic Index scores by bay segment}
\usage{
anlz_tbbimed(
  tbbiscr,
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"),
  rev = FALSE,
  yrrng = c(1993, 2023)
)
}
\arguments{
\item{tbbiscr}{input data frame as returned by \code{\link{anlz_tbbiscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{yrrng}{numeric indicating year ranges to evaluate}
}
\value{
A data frame of annual medians by bay segment
}
\description{
Get annual medians of Tampa Bay Benthic Index scores by bay segment
}
\details{
Additional summaries are provided for the entire bay, as a summary across categories ("All") and a summary weighted across the relative sizes of each bay segment ("All (wt)").

Only sampling funded by TBEP and as part of the routine EPC benthic monitoring program are included in the final categories.
}
\examples{
tbbiscr <- anlz_tbbiscr(benthicdata)
anlz_tbbimed(tbbiscr)
}
\concept{analyze}
