% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbbiscr.R
\name{anlz_tbbiscr}
\alias{anlz_tbbiscr}
\title{Get Tampa Bay Benthic Index scores}
\usage{
anlz_tbbiscr(benthicdata)
}
\arguments{
\item{benthicdata}{nested \code{\link[tibble]{tibble}} formatted from \code{\link{read_importbenthic}}}
}
\value{
A single data frame of TBBI scores for each site.
}
\description{
Get Tampa Bay Benthic Index scores
}
\details{
This function calculates scores for the TBBI based on station, taxa, and field sample data.  The total TBBI scores are returned as \code{TBBI} and \code{TBBICat}, where the latter is a categorical description of the scores.
}
\examples{
anlz_tbbiscr(benthicdata)
}
\concept{analyze}
