% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tbniscr.R
\name{anlz_tbniscr}
\alias{anlz_tbniscr}
\title{Get Tampa Bay Benthic Index scores}
\usage{
anlz_tbniscr(fimdata, raw = TRUE)
}
\arguments{
\item{fimdata}{\code{data.frame} formatted from \code{\link{read_importfim}}}

\item{raw}{logical indicating if raw metric values are also returned}
}
\value{
A data frame of metrics and TBNI scores in wide fomat.
}
\description{
Get Tampa Bay Benthic Index scores
}
\details{
This function calculates raw and scored metrics for the TBNI, including \code{NumTaxa}, \code{BenthicTaxa}, \code{TaxaSelect}, \code{NumGuilds}, and \code{Shannon}.  The total TBNI score is returned as \code{TBNI_Score}.
}
\examples{
anlz_tbniscr(fimdata)
}
\seealso{
\code{\link{anlz_tbnimet}}

Other anlz: 
\code{\link{anlz_tbniave}()},
\code{\link{anlz_tbnimet}()}
}
\concept{anlz}
