% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tdlcrk.R
\name{anlz_tdlcrk}
\alias{anlz_tdlcrk}
\title{Estimate tidal creek report card scores}
\usage{
anlz_tdlcrk(tidalcreeks, iwrraw, tidtrgs = NULL, yr = 2025)
}
\arguments{
\item{tidalcreeks}{\code{\link[sf]{sf}} object for population of tidal creeks}

\item{iwrraw}{FDEP impaired waters data base as \code{\link{data.frame}}}

\item{tidtrgs}{optional \code{data.frame} for tidal creek nitrogen targets, defaults to \code{\link{tidaltargets}}}

\item{yr}{numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date}
}
\value{
A \code{\link{data.frame}} with the report card scores for each creek, as prioritize, investigate, caution, monitor, or no data
}
\description{
Estimate tidal creek report card scores
}
\examples{
anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2025)
}
\concept{analyze}
