% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_tdlcrkindic.R
\name{anlz_tdlcrkindic}
\alias{anlz_tdlcrkindic}
\title{Analyze tidal creek water quality indicators}
\usage{
anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
}
\arguments{
\item{tidalcreeks}{\code{\link[sf]{sf}} object for population of tidal creeks}

\item{iwrraw}{FDEP impaired waters rule run 56 data base as \code{\link{data.frame}}}

\item{yr}{numeric for reference year to evaluate, scores are based on the planning period beginning ten years prior to this date}
}
\value{
A \code{\link{data.frame}} with the indicator values for each tidal creek
}
\description{
Estimate tidal creek water quality indicators to suppport report card scores
}
\details{
Annual geometric means for additional water quality data available at each wbid, JEI combination.  Florida trophic state index values are also estimated where data are available.  Nitrogen ratios are estimated for JEIs that cover source (upstream, freshwater) and tidal (downstream) WBIDs, defined as the ratio of concentrations between the two (i.e., ratios > 1 mean source has higher concentrations).  Nitrogen ratios for a given year reflect the median of all daily ratios when concentrations were measured in both a source and tidal segment during the same day.  Note that a ratio of one can be obtained if both the source and tidal segments are at minimum detection.
}
\examples{
dat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2018)
head(dat)
}
\seealso{
Other analyze: 
\code{\link{anlz_attain}()},
\code{\link{anlz_avedatsite}()},
\code{\link{anlz_avedat}()},
\code{\link{anlz_iwrraw}()},
\code{\link{anlz_refs}()},
\code{\link{anlz_tbbimed}()},
\code{\link{anlz_tdlcrk}()},
\code{\link{anlz_transectave}()},
\code{\link{anlz_transectocc}()},
\code{\link{anlz_yrattain}()}
}
\concept{analyze}
