% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_transectavespp.R
\name{anlz_transectavespp}
\alias{anlz_transectavespp}
\title{Get annual averages of seagrass frequency occurrence by bay segments, year, and species}
\usage{
anlz_transectavespp(
  transectocc,
  bay_segment = c("OTB", "HB", "MTB", "LTB", "BCB"),
  yrrng = c(1998, 2019),
  species = c("Halodule", "Syringodium", "Thalassia", "Ruppia", "Halophila spp.",
    "Caulerpa spp."),
  total = TRUE
)
}
\arguments{
\item{transectocc}{data frame returned by \code{\link{anlz_transectocc}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB", "BCB"}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{species}{chr string of species to summarize, one to many of "Halodule", "Syringodium", "Thalassia", "Ruppia", "Halophila spp.", "Caulerpa spp."}

\item{total}{logical indicating if total frequency occurrence for all species is also returned}
}
\value{
A data frame of annual averages by bay segment
}
\description{
Get annual averages of seagrass frequency occurrence by bay segments, year, and species
}
\details{
Frequency occurrence estimates are averaged across segments in \cod{bay_segment}, i.e., separate results by location are not returned.
}
\examples{
\dontrun{
transect <- read_transect()
}
transectocc <- anlz_transectocc(transect)
anlz_transectavespp(transectocc)
}
\seealso{
Other analyze: 
\code{\link{anlz_attain}()},
\code{\link{anlz_avedatsite}()},
\code{\link{anlz_avedat}()},
\code{\link{anlz_iwrraw}()},
\code{\link{anlz_refs}()},
\code{\link{anlz_tbbimed}()},
\code{\link{anlz_tdlcrkindic}()},
\code{\link{anlz_tdlcrk}()},
\code{\link{anlz_transectave}()},
\code{\link{anlz_transectocc}()},
\code{\link{anlz_yrattain}()}
}
\concept{analyze}
