% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_transectavespp.R
\name{anlz_transectavespp}
\alias{anlz_transectavespp}
\title{Get annual averages of seagrass frequency occurrence by bay segments, year, and species}
\usage{
anlz_transectavespp(
  transectocc,
  bay_segment = c("OTB", "HB", "MTB", "LTB", "BCB"),
  yrrng = c(1998, 2022),
  species = c("Halodule", "Syringodium", "Thalassia", "Ruppia", "Halophila",
    "Caulerpa"),
  total = TRUE,
  by_seg = FALSE
)
}
\arguments{
\item{transectocc}{data frame returned by \code{\link{anlz_transectocc}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB", "BCB"}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{species}{chr string of species to summarize, one to many of "Halodule", "Syringodium", "Thalassia", "Ruppia", "Halophila", "Caulerpa"}

\item{total}{logical indicating if total frequency occurrence for all species is also returned}

\item{by_seg}{logical indicating if separate results by bay segments are retained}
}
\value{
A data frame of annual averages by bay segment
}
\description{
Get annual averages of seagrass frequency occurrence by bay segments, year, and species
}
\details{
Frequency occurrence estimates are averaged across segments in \code{bay_segment} if \code{by_seg = F}, i.e., separate results by location are not returned.  Results are retained by bay segment if \code{by_seg = T}.  Also note that totals across species (\code{total = T}) are not returned if \code{by_seg = T}.
}
\examples{
\dontrun{
transect <- read_transect()
}
transectocc <- anlz_transectocc(transect)
anlz_transectavespp(transectocc)
}
\concept{analyze}
