% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anlz_transectocc.R
\name{anlz_transectocc}
\alias{anlz_transectocc}
\title{Get seagrass average abundance and occurrence across transects}
\usage{
anlz_transectocc(transect)
}
\arguments{
\item{transect}{data frame returned by \code{\link{read_transect}}}
}
\value{
A data frame with abundance and frequency occurrence estimates aggregated by species, transect, and date.  The nsites column is the total number of placements that were sampled along a transect for a particular date.
}
\description{
Get seagrass average abundance and occurrence across transects
}
\details{
Abundance and frequency occurrence are estimated as in Sherwood et al. 2017, equations 1 and 2.  In short, frequency occurrence is estimated as the number of instances a species was observed along a transect divided by the number of placements along a transect and average abundance was estimated as the sum of species-specific Braun-Blanquet scores divided by the number of placements along a transect.  The estimates are obtained for all seagrass species including Caulerpa (attached macroalgae), Dapis (cyanobacteria), and Chaetomorpha (drift green algea), whereas all other attached and drift algae species are aggregated. Drift or attached macroalgae and cyanobacteria (Dapis) estimates may not be accurate prior to 2021.
}
\examples{
\dontrun{
transect <- read_transect()
}
anlz_transectocc(transect)
}
\references{
Sherwood, E.T., Greening, H.S., Johansson, J.O.R., Kaufman, K., Raulerson, G.E. 2017. Tampa Bay (Florida, USA): Documenting seagrass recovery since the 1980's and reviewing the benefits. Southeastern Geographer. 57(3):294-319.
}
\concept{analyze}
