% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthicdata.R
\docType{data}
\name{benthicdata}
\alias{benthicdata}
\title{Benthic data for the Tampa Bay Benthic Index current as of 09132020}
\format{
A data frame with 4516 rows and 12 variables:
\describe{
  \item{StationID}{int}
  \item{ProgramID}{int}
  \item{ProgramName}{chr}
  \item{Latitude}{num}
  \item{Longitude}{num}
  \item{date}{Date}
  \item{yr}{num}
  \item{TotalAbundance}{num}
  \item{SpecesRichness}{num}
  \item{TBBI}{num}
  \item{TBBICat}{chr}
  \item{Salinity}{num}
  }
}
\usage{
benthicdata
}
\description{
Benthic data for the Tampa Bay Benthic Index current as of 09132020
}
\examples{
\dontrun{
# location to download data
path <- 'C:/Users/mbeck/Desktop/benthic.zip'

# load and assign to object
benthicdata <- read_importbenthic(path, download_latest = TRUE, remove = TRUE)

save(benthicdata, file = 'data/benthicdata.RData', compress = 'xz')

}
}
\seealso{
Other utilities: 
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{fimstations}},
\code{\link{iwrraw}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
