% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benthicdata.R
\docType{data}
\name{benthicdata}
\alias{benthicdata}
\title{Benthic data for the Tampa Bay Benthic Index current as of 10222020}
\format{
A nested \code{\link[tibble]{tibble}} with 3 rows and 2 variables:
\describe{
  \item{name}{chr}
  \item{value}{list}
  }
}
\usage{
benthicdata
}
\description{
Benthic data for the Tampa Bay Benthic Index current as of 10222020
}
\examples{
\dontrun{
# location to download data
path <- 'C:/Users/mbeck/Desktop/benthic.zip'

# load and assign to object
benthicdata <- read_importbenthic(path, download_latest = TRUE, remove = TRUE)

save(benthicdata, file = 'data/benthicdata.RData', compress = 'xz')

}
}
\seealso{
Other utilities: 
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{iwrraw}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}},
\code{\link{transect}}
}
\concept{utilities}
\keyword{datasets}
