% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsmap.R
\docType{data}
\name{bsmap}
\alias{bsmap}
\title{Terrain basemap}
\format{
A \code{\link[ggmap]{ggmap}} object
}
\usage{
bsmap
}
\description{
Terrain basemap
}
\seealso{
Other utilities: 
\code{\link{benthicdata}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{fimstations}},
\code{\link{iwrraw}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
