% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enterodata.R
\docType{data}
\name{enterodata}
\alias{enterodata}
\title{Enterococcus data from 53 key Enterococcus stations since 1995}
\format{
A data frame with 6207 rows and 12 columns:
\describe{
 \item{date}{data, sample date}
 \item{station}{character, sample station, as named in the Water Quality Portal}
 \item{ecocci}{numeric, sample's Enterococcus concentration}
 \item{ecocci_censored}{logical, whether sample concentration was censored (below detection limit)}
 \item{Latitude}{numeric, latitude in decimal degrees}
 \item{Longitude}{numeric, longitude in decimal degrees}
 \item{ecocci_units}{character, units of measurement of Enterococcus}
 \item{time}{character, time sample was obtained, in 24hr format}
 \item{time_zone}{character, time zone of sample time}
 \item{MDL}{numeric, minimum detection limit of laboratory for Enterococcus}
 \item{yr}{numeric year of sample date}
 \item{mo}{numeric month of sample date}
}
}
\source{
Water Quality Portal, \url{https://waterqualitydata.us}
}
\usage{
enterodata
}
\description{
Enterococcus data from 53 key Enterococcus stations since 1995
}
\details{
A sample dataset containing Enterococcus from 53 stations in the TBEP watershed from 1995-2023. Generated by \code{data-raw/enterodata-raw.R} (view on github: \url{https://github.com/tbep-tech/tbeptools/blob/master/data-raw/enterodata-raw.R})
}
\keyword{datasets}
