% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epcdata.R
\docType{data}
\name{epcdata}
\alias{epcdata}
\title{All bay data as of 02042020}
\format{A data frame with 25669 rows and 14 variables:
\describe{
  \item{bay_segment}{chr}
  \item{epchc_station}{num}
  \item{SampleTime}{POSIXct}
  \item{yr}{num}
  \item{mo}{num}
  \item{Latitude}{num}
  \item{Longitude}{num}
  \item{Total_Depth_m}{num}
  \item{Sample_Depth_m}{num}
  \item{sd_m}{num}
  \item{chla}{num}
  \item{Sal_Top_ppth}{num}
  \item{Sal_Mid_ppth}{num}
  \item{Sal_Bottom_ppth}{num}
  }}
\usage{
epcdata
}
\description{
All bay data as of 02042020
}
\seealso{
Other utilities: \code{\link{bsmap}}, \code{\link{iwrraw}},
  \code{\link{stations}}, \code{\link{targets}},
  \code{\link{tbseglines}}, \code{\link{tbseg}},
  \code{\link{tbshed}}, \code{\link{tidalcreeks}},
  \code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
