% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epcdata.R
\docType{data}
\name{epcdata}
\alias{epcdata}
\title{All bay data as of 11242019 (Nov 24 2019)}
\format{A data frame with 25534 rows and 11 variables:
\describe{
  \item{bay_segment}{chr}
  \item{epchc_station}{num}
  \item{SampleTime}{POSIXct}
  \item{yr}{num}
  \item{mo}{num}
  \item{Latitude}{num}
  \item{Longitude}{num}
  \item{Total_Depth_m}{num}
  \item{Sample_Depth_m}{num}
  \item{sd_m}{num}
  \item{chla}{num}
}}
\usage{
epcdata
}
\description{
All bay data as of 11242019 (Nov 24 2019)
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{stations}}, \code{\link{targets}},
  \code{\link{tbseg}}, \code{\link{tidalcreeks}}
}
\concept{utilities}
\keyword{datasets}
