% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimdata.R
\docType{data}
\name{fimdata}
\alias{fimdata}
\title{FIM data for Tampa Bay Nekton Index current as of 04242020}
\format{A data frame with 41653 rows and 17 variables:
\describe{
  \item{Reference}{chr}
  \item{Sampling_Date}{Date}
  \item{Zone}{chr}
  \item{Grid}{int}
  \item{NODCCode}{chr}
  \item{bay_segment}{chr}
  \item{Year}{num}
  \item{Month}{num}
  \item{Total_N}{num}
  \item{ScientificName}{chr}
  \item{Include_TB_Index}{chr}
  \item{Hab_Cat}{chr}
  \item{Est_Cat}{chr}
  \item{Est_Use}{chr}
  \item{Feeding_Cat}{chr}
  \item{Feeding_Guild}{chr}
  \item{Selected_Taxa}{chr}
  }}
\usage{
fimdata
}
\description{
FIM data for Tampa Bay Nekton Index current as of 04242020
}
\examples{
\dontrun{
csv <- '~/Desktop/fimdata.csv'

load and assign to object
fimdata <- read_importfim(csv, download_latest = T)

save(fimdata, file = 'data/fimdata.RData')
}
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{fimstations}},
  \code{\link{iwrraw}}, \code{\link{stations}},
  \code{\link{targets}}, \code{\link{tbseglines}},
  \code{\link{tbseg}}, \code{\link{tbshed}},
  \code{\link{tidalcreeks}}, \code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
