% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fimstations.R
\docType{data}
\name{fimstations}
\alias{fimstations}
\title{Spatial data object of FIM stations including Tampa Bay segments}
\format{
A simple features \code{\link[sf]{sf}} object (POINT) with 6186 features and 4 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
\describe{
  \item{Reference}{num}
  \item{bay_segment}{chr}
}
}
\usage{
fimstations
}
\description{
Spatial data object of FIM stations including Tampa Bay segments
}
\examples{
\dontrun{
# file path
csv <- '~/Desktop/fimraw.csv'

# load and assign to object
fimstations <- read_importfim(csv, download_latest = TRUE, locs = TRUE)
save(fimstations, file = 'data/fimstations.RData', compress = 'xz')
}
}
\seealso{
Other utilities: 
\code{\link{benthicdata}},
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{iwrraw}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}},
\code{\link{transect}}
}
\concept{utilities}
\keyword{datasets}
