% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwrraw.R
\docType{data}
\name{iwrraw}
\alias{iwrraw}
\title{FDEP IWR run 61}
\format{
A data frame 405682 rows and 15 variables
}
\usage{
iwrraw
}
\description{
Florida Department of Environmental Protection, Impaired Waters Rule, Run 61
}
\examples{
\dontrun{
library(dplyr)

load(file = '../../02_DOCUMENTS/tidal_creeks/iwrraw_run61.RData')
iwrraw <- sf::st_set_geometry(iwrraw, NULL) \%>\%
  rename(JEI = jei)
save(iwrraw, file = 'data/iwrraw.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
