% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwrraw.R
\docType{data}
\name{iwrraw}
\alias{iwrraw}
\title{FDEP IWR run 56}
\format{A data frame 1224558 rows and 52 variables}
\usage{
iwrraw
}
\description{
Florida Department of Environmental Protection, Impaired Waters Rule, Run 56
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{stations}},
  \code{\link{targets}}, \code{\link{tbseg}},
  \code{\link{tidalcreeks}}
}
\concept{utilities}
\keyword{datasets}
