% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwrraw.R
\docType{data}
\name{iwrraw}
\alias{iwrraw}
\title{FDEP IWR run 56}
\format{A data frame 1224558 rows and 52 variables}
\usage{
iwrraw
}
\description{
Florida Department of Environmental Protection, Impaired Waters Rule, Run 56
}
\details{
This file was created by importing a SAS file directly into R using the \code{\link[haven]{read_sas}} function.  The SAS file is prepared annually by FDEP for Janicki Environmental, Inc.
}
\examples{
\dontrun{
library(haven)

iwrraw <- read_sas('../../02_DOCUMENTS/tidal_creeks/iwr56_tidalcreeks.sas7bdat')
save(iwrraw, file = here('data', 'iwrraw.RData'), compress = 'xz')

}
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{stations}},
  \code{\link{targets}}, \code{\link{tbseg}},
  \code{\link{tbshed}}, \code{\link{tidalcreeks}},
  \code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
