% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwrraw.R
\docType{data}
\name{iwrraw}
\alias{iwrraw}
\title{FDEP IWR run 65}
\format{
A data frame 195826 rows and 11 variables
}
\usage{
iwrraw
}
\description{
Florida Department of Environmental Protection, Impaired Waters Rule, Run 65
}
\details{
File was created using workflow at <https://tbep-tech.github.io/tidalcreek-stats/Creek_select_tbeptools>, example below is old and for Run 61.
}
\examples{
\dontrun{
library(dplyr)

load(file = '../../02_DOCUMENTS/tidal_creeks/iwrraw_run61.RData')
iwrraw <- sf::st_set_geometry(iwrraw, NULL) \%>\%
  rename(JEI = jei)
save(iwrraw, file = 'data/iwrraw.RData', compress = 'xz')
}
}
\concept{data}
\keyword{datasets}
