% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_epchc_wq.R
\name{load_epchc_wq}
\alias{load_epchc_wq}
\title{Load local water quality file}
\usage{
load_epchc_wq(xlsx, na = "", download_latest_epchc = FALSE, ...)
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite it not current}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}

\item{download_latest_epchc}{logical passed to \link{get_epchc_wq} to download raw data and compare with existing in \code{xlsx} if available}

\item{...}{additional arguments passed to \code{\link[readxl]{read_xlsx}}}
}
\value{
A formatted \code{data.frame} of water quality data
}
\description{
Load local water quality file
}
\details{
Loads the "RWMDataSpreadsheet" worksheet from the file located at \code{xlsx}
}
\examples{
\dontrun{
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'
epcdata <- load_epchc_wq(xlsx)
}
}
