% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phytodata.R
\docType{data}
\name{phytodata}
\alias{phytodata}
\title{Phytoplankton data current as of 20260204}
\format{
A nested \code{\link[tibble]{tibble}} with 41390 rows and 8 variables:
\describe{
\item{epchc_station}{chr}
\item{Date}{Date}
\item{name}{chr}
\item{units}{chr}
\item{count}{num}
\item{yrqrt}{Date}
\item{yr}{num}
\item{mo}{Ord.factor}
}
}
\usage{
phytodata
}
\description{
Phytoplankton data current as of 20260204
}
\examples{
\dontrun{
# location to download data
path <- tempfile(fileext = '.xlsx')

# load and assign to object
phytodata <- read_importphyto(path, download_latest = TRUE)

nrow(phytodata)
ncol(phytodata)

save(phytodata, file = 'data/phytodata.RData', compress = 'xz')

file.remove(path)
}
}
\concept{data}
\keyword{datasets}
