% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chkdate.R
\name{read_chkdate}
\alias{read_chkdate}
\title{Compare date of local xlsx file with the same file on the server}
\usage{
read_chkdate(epchc_url, xlsx, connecttimeout = 20, tryurl = FALSE)
}
\arguments{
\item{epchc_url}{chr string of full path to file on the server}

\item{xlsx}{chr string of full path to local file}

\item{connecttimeout}{numeric for maximum number of seconds to wait until connection timeout for \code{\link[RCurl]{getURL}}}

\item{tryurl}{logical indicating if \code{\link[RCurl]{getURL}} is repeatedly called in a \code{while} loop if first connection is unsuccessful}
}
\value{
A logical vector indicating if the local file is current
}
\description{
Compare date of local xlsx file with the same file on the server
}
\examples{
\dontrun{
epchc_url <- "ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx"
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'
read_chkdate(epchc_url, xlsx)
}
}
\seealso{
Other read: \code{\link{read_dlcurrent}},
  \code{\link{read_formwq}}, \code{\link{read_importwq}}
}
\concept{read}
