% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dlcurrent.R
\name{read_dlcurrent}
\alias{read_dlcurrent}
\title{Download latest file from epchc.org}
\usage{
read_dlcurrent(xlsx, download_latest = TRUE, urlin)
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite it not current}

\item{download_latest}{logical to download latest file regardless of local copy}

\item{urlin}{url for file location}
}
\value{
The local copy specified in the path by \code{xlsx} is overwritten by the new file is not current or \code{download_latest = TRUE}.  The function does nothing if \code{download_latest = FALSE}.
}
\description{
Download latest file from epchc.org
}
\details{
The local copy is checked against a temporary file downloaded from the location specified by \code{urlin}.  The local file is replaced with the downloaded file if the MD5 hashes are different.
}
\examples{
\dontrun{
xlsx <- '~/Desktop/2018_Results_Updated.xls'
urlin <- 'ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/'
urlin <- paste0(urlin, 'RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx')
read_dlcurrent(xlsx, urlin = urlin)
}
}
\seealso{
Other read: 
\code{\link{read_chkdate}()},
\code{\link{read_formfim}()},
\code{\link{read_formphyto}()},
\code{\link{read_formwq}()},
\code{\link{read_importfim}()},
\code{\link{read_importphyto}()},
\code{\link{read_importwq}()}
}
\concept{read}
