% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formbenthic.R
\name{read_formbenthic}
\alias{read_formbenthic}
\title{Format benthic data for the Tampa Bay Benthic Index}
\usage{
read_formbenthic(channel)
}
\arguments{
\item{channel}{A \code{\link[RODBC]{RODBC}} class connection object to the .mdb benthic database}
}
\value{
A nested \code{\link[tibble]{tibble}} of station, field sample, and taxa data
}
\description{
Format benthic data for the Tampa Bay Benthic Index
}
\details{
Function is used internally within \code{\link{read_importbenthic}}, see the help file for limitations on using the function outside of Windows.
}
\examples{
\dontrun{
library(RODBC)

# setup driver and path to .mdb
drvr <- 'Driver={Microsoft Access Driver (*.mdb, *.accdb)};'
path <- 'C:/Users/mbeck/Desktop/benthic/EPC DataSubmittals.mdb'
odbccall <- paste0(drvr, 'DBQ=', path)

# create channel and pass to read_formbenthic
channel <- RODBC::odbcDriverConnect(odbccall)
read_formbenthic(channel)

# close connection
odbcClose(channel)
}
}
\concept{read}
