% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formfib.R
\name{read_formfib}
\alias{read_formfib}
\title{Format Fecal Indicator Bacteria (FIB) data}
\usage{
read_formfib(datin, all = FALSE)
}
\arguments{
\item{datin}{input \code{data.frame} loaded from \code{\link{read_importepc}}}

\item{all}{logical indicating if all stations with FIB data are returned, default is \code{FALSE}}
}
\value{
A lightly formatted \code{data.frame} with FIB data
}
\description{
Format Fecal Indicator Bacteria (FIB) data
}
\details{
Formats input data from \code{\link{read_importepc}} appropriate for FIB results, see the details in \code{\link{read_importfib}} for more more information.
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'

# load raw data and assign to object
epcall <- read_importepc(xlsx, download_latest = T)

# final formatting
fibdata <- read_formfib(epcall)
}
}
\seealso{
\code{\link{read_importfib}}, \code{\link{read_importepc}}
}
\concept{read}
