% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formfim.R
\name{read_formfim}
\alias{read_formfim}
\title{Format FIM data for the Tampa Bay Nekton Index}
\usage{
read_formfim(datin, locs = FALSE)
}
\arguments{
\item{datin}{input \code{data.frame} loaded from \code{\link{read_importfim}}}

\item{locs}{logical indicating if a spatial features object is returned with locations of each FIM sampling station}
}
\value{
A formatted \code{data.frame} with FIM data if \code{locs = FALSE}, otherwise a simple features object if \code{locs = TRUE}
}
\description{
Format FIM data for the Tampa Bay Nekton Index
}
\details{
Function is used internally within \code{\link{read_importfim}}
}
\examples{
\dontrun{
# file path
csv <- '~/Desktop/fimraw.csv'

# load and assign to object
fimdata <- read_importfim(csv)
}
}
\seealso{
\code{\link{read_importfim}}
}
\concept{read}
