% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formsediment.R
\name{read_formsediment}
\alias{read_formsediment}
\title{Format sediment data}
\usage{
read_formsediment(pathin)
}
\arguments{
\item{pathin}{A path to unzipped csv files with sediment result tables}
}
\value{
A \code{data.frame} of sediment data for Tampa Bay
}
\description{
Format sediment data
}
\details{
Function is used internally within \code{\link{read_importsediment}}
}
\examples{
\dontrun{

# location to download data
path <- '~/Desktop/sediment.zip'

# download
urlin1 <- 'https://epcbocc.sharepoint.com/:u:/s/PublishedShare/'
urlin2 <- 'IQAr2aDfEyegTrJLR61qt1CxAQYxfnVYYWAwkbtArczrJS4?e=FE6hyM'
urlin <- paste0(urlin1, urlin2, '&download=1')
read_dlcurrent(path, download_latest = TRUE, urlin = urlin)

# unzip
tmppth <- tempfile()
utils::unzip(path, exdir = tmppth, overwrite = TRUE)

# format sediment data
read_formsediment(pathin = tmppth)

# remove temporary path
unlink(tmppth, recursive = TRUE)

}
}
\concept{read}
