% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formwq.R
\name{read_formwq}
\alias{read_formwq}
\title{Format water quality data}
\usage{
read_formwq(datin, all = FALSE)
}
\arguments{
\item{datin}{input \code{data.frame} loaded from \code{\link{read_importwq}}}

\item{all}{logical indicating if all water quality parameters are returned, see details}
}
\value{
A lightly formatted \code{data.frame} with chloropyll and secchi observations
}
\description{
Format water quality data
}
\details{
Secchi data VOB depths or secchis < 0.5 ft from bottom are assigned \code{NA}, function is used internally within \code{\link{read_importwq}}
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'

# load raw data and assign to object
epcall <- read_importepc(xlsx, download_latest = T)

# final formatting
epcdata <- read_formwq(epcall)
}
}
\seealso{
\code{\link{read_importwq}}
}
\concept{read}
