% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_formwqp.R
\name{read_formwqp}
\alias{read_formwqp}
\title{Format water quality and station metadata from the Water Quality Portal}
\usage{
read_formwqp(res, sta, org, orgin, trace = F)
}
\arguments{
\item{res}{A data frame containing water quality results obtained from the API}

\item{sta}{A data frame containing station metadata obtained from the API}

\item{org}{chr string indicating either \code{"Manatee"} or \code{"Pinellas"}}

\item{orgin}{chr string indicating either \code{"21FLMANA_WQX"} or \code{"21FLPDEM_WQX"}}

\item{trace}{Logical indicating whether to display progress messages, default is \code{FALSE}}
}
\value{
A data frame containing formatted water quality and station metadata
}
\description{
Format water quality and station metadata from the Water Quality Portal
}
\details{
This function is used by \code{\link{read_importwqp}} to combine, format, and process water quality data (\code{res}) and station metadata (\code{sta}) obtained from the Water Quality Data Exchange API for the selected county. The resulting data frame includes the date, time, station identifier, latitude, longitude, variable name, value, unit, and quality flag.
}
\examples{
\dontrun{
url <- list(
  Result = "https://www.waterqualitydata.us/data/Result/search?mimeType=csv&zip=no",
  Station = "https://www.waterqualitydata.us/data/Station/search?mimeType=csv&zip=no"
)

headers <- c(
  "Content-Type" = "application/json",
  "Accept" = "application/zip"
)

body <- list(
  organization = c("21FLMANA_WQX"),
  sampleMedia = c("Water"),
  characteristicType = c("Information", "Nutrient", "Biological, Algae, Phytoplankton,
    Photosynthetic Pigments"),
  providers = c("STORET"),
  siteType = c("Estuary")
)

res <- url[['Result']] \%>\%
  httr::POST(httr::add_headers(headers), body = jsonlite::toJSON(body)) \%>\%
  httr::content('text') \%>\%
  read.csv(text = .)

sta <- url[['Station']] \%>\%
  httr::POST(httr::add_headers(headers), body = jsonlite::toJSON(body)) \%>\%
  httr::content('text') \%>\%
  read.csv(text = .)

# combine and format
read_formwqp(res, sta, 'Manatee', '21FLMANA_WQX')
}
}
\seealso{
\code{\link{read_importwqp}}
}
\concept{read}
