% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importbenthic.R
\name{read_importbenthic}
\alias{read_importbenthic}
\title{Download and import benthic data for Tampa Bay}
\usage{
read_importbenthic(path, download_latest = FALSE, remove = TRUE)
}
\arguments{
\item{path}{chr string for local path where the mdb database is located or where it will be downloaded, must include file extension either \code{.mdb} for former or \code{.zip} for latter}

\item{download_latest}{logical to download latest if a more recent dataset is available}

\item{remove}{logical if the downloaded zip file is removed after unzipping}
}
\value{
A \code{data.frame} of formatted benthic data.
}
\description{
Download and import benthic data for Tampa Bay
}
\details{
This function downloads and unzips a Microsoft Access Database from \url{ftp://ftp.epchc.org/EPC_ERM_FTP/Benthic_Monitoring/}.  The download location is temporary unless specified with the \code{path} argument.

The database tables are accessed using a connection through ODBC with the RODBC package.  This requires an installation of the Microsoft 64 bit driver for Access databases, available at <https://www.microsoft.com/en-us/download/details.aspx?id=54920>.  The function has not been tested for 32-bit versions of R or outside of Windows.

For the \code{path} argument, you must specify the full path and not the home expansion, i.e., \code{'C:/Users/yourname/Desktop/benthic.zip'} and not \code{'~/Desktop/benthic.zip'}.

i
}
\examples{
\dontrun{
# location to download data
path <- 'C:/path/to/benthic.zip'

# load and assign to object
benthicdata <- read_importbenthic(path, download_latest = TRUE)

# or use existing mdb
path <- 'C:/path/to/benthic/EPC DataSubmittals.mdb'
benthicdata <- read_importbenthic(path, download_latest = FALSE)
}
}
\seealso{
Other read: 
\code{\link{read_chkdate}()},
\code{\link{read_dlcurrent}()},
\code{\link{read_formfim}()},
\code{\link{read_formphyto}()},
\code{\link{read_formwq}()},
\code{\link{read_importfim}()},
\code{\link{read_importphyto}()},
\code{\link{read_importwq}()}
}
\concept{read}
