% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importentero.R
\name{read_importentero}
\alias{read_importentero}
\title{Download Enterococcus data from the Water Quality Portal}
\usage{
read_importentero(args)
}
\arguments{
\item{args}{a list with the following fields:}
}
\value{
a data frame
}
\description{
Download Enterococcus data from the Water Quality Portal
}
