% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importentero.R
\name{read_importentero}
\alias{read_importentero}
\title{Download Enterococcus data from the Water Quality Portal}
\usage{
read_importentero(args)
}
\arguments{
\item{args}{a list of arguments to pass to Water Quality Portal. See \code{examples} for detail on how to use. Must contain the following fields:
\describe{
 \item{\code{siteid}}{character, vector of stations}
 \item{\code{characteristicName}}{character, vector to select 'Characteristic Names' from WQP}
 \item{\code{startDateLo}}{date, start date for data to download}
 \item{\code{startDateHi}}{date, end date for data to download}
}}
}
\value{
a data frame containing one row for each sample. Columns returned are:
\describe{
 \item{\code{date}}{date, sample date}
 \item{\code{station}}{character, sample station}
 \item{\code{ecocci}}{numeric, Enterococcus concentration}
 \item{\code{ecocci_censored}}{logical, whether \code{ecocci} value was below the laboratory \code{MDL}, minimum detection limit}
 \item{\code{ecocci_units}}{character, units of measurement for \code{ecocci}}
 \item{\code{qualifier}}{qualifier codes associated with sample}
 \item{\code{LabComments}}{lab comments on sample}
 \item{\code{Latitude}}{numeric, latitude in decimal degrees}
 \item{\code{Longitude}}{numeric, longitude in decimal degrees}
 \item{\code{time}}{character, sample time}
 \item{\code{time_zone}}{character, sample time zone}
 \item{\code{MDL}}{numeric, minimum detection limit for the lab and time the sample was analyzed}
 \item{\code{yr}}{numeric, year of sample date}
 \item{\code{mo}}{numeric, month of sample date}
 }
}
\description{
Download Enterococcus data from the Water Quality Portal
}
\details{
Retrieves Enterococcus sample data from selected stations and date range from the Water Quality Portal, \url{https://www.waterqualitydata.us}
}
\examples{
\dontrun{
# set up list of args
stations <- c('21FLHILL_WQX-101',
'21FLHILL_WQX-102',
'21FLHILL_WQX-103')
entero_names <- c('Enterococci',
                  'Enterococcus')
startDate <- as.Date('2023-01-01')
endDate <- as.Date('2023-02-01')

args <- list(
  siteid = stations,
  characteristicName = entero_names,
  startDateLo = format(startDate, '\%m-\%d-\%Y'),
  startDateHi = format(endDate, '\%m-\%d-\%Y')
)

# download and read the data
entero_in <- read_importentero(args = args)
head(entero_in)

}
}
