% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importepc.R
\name{read_importepc}
\alias{read_importepc}
\title{Download and/or import local water quality file for internal use}
\usage{
read_importepc(xlsx, download_latest = FALSE, na = c("", "NULL"))
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{xlsx} if available}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}
}
\value{
An unformatted \code{data.frame} from EPC
}
\description{
Download and/or import local water quality file for internal use
}
\details{
Loads the "RWMDataSpreadsheet" worksheet from the file located at \code{xlsx}.  The file is downloaded from \url{https://epcbocc.sharepoint.com/:x:/s/Share/IQCF2ebdelJRRpNa8yFPdVaIAWvFN-5YfM_x37X9iiSEkh4?e=5eti14&download=1}.  The files can be viewed at \url{https://epcbocc.sharepoint.com/:f:/s/Share/EiypSSYdsEFCi84Sv_6-t7kBUYaXiIqN0B1n2w57Z_V3kQ?e=NdZQcU}.

This function is used internally by \code{\link{read_importwq}} and \code{\link{read_importfib}} because both use the same source data from the Environmental Protection Commission of Hillsborough County.
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'

# load and assign to object
epcall <- read_importepc(xlsx, download_latest = T)
}
}
\seealso{
\code{\link{read_importwq}}, \code{\link{read_importfib}}
}
\concept{read}
