% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importfim.R
\name{read_importfim}
\alias{read_importfim}
\title{Load local FIM data for the Tampa Bay Nekton Index}
\usage{
read_importfim(csv, download_latest = FALSE, locs = FALSE)
}
\arguments{
\item{csv}{chr string path for local csv file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{csv} if available}

\item{locs}{logical indicating if a spatial features object is returned with locations of each FIM sampling station}
}
\value{
A formatted \code{data.frame} with FIM data if \code{locs = FALSE}, otherwise a simple features object if \code{locs = TRUE}
}
\description{
Load local FIM data for the Tampa Bay Nekton Index
}
\details{
Data downloaded from \url{'ftp://ftp.floridamarine.org/users/fim/tmac/NektonIndex/TampaBay_NektonIndexData.csv'}.
}
\examples{
\dontrun{
# file path
csv <- '~/Desktop/fimraw.csv'

# load and assign to object
fimdata <- read_importfim(csv, download_latest = TRUE)
}
}
\seealso{
\code{\link{read_formwq}}, \code{\link{read_importphyto}}

Other read: 
\code{\link{read_chkdate}()},
\code{\link{read_dlcurrent}()},
\code{\link{read_formbenthic}()},
\code{\link{read_formfim}()},
\code{\link{read_formphyto}()},
\code{\link{read_formwq}()},
\code{\link{read_importbenthic}()},
\code{\link{read_importphyto}()},
\code{\link{read_importwq}()}
}
\concept{read}
