% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importphyto.R
\name{read_importphyto}
\alias{read_importphyto}
\title{Load local phytoplankton cell count file}
\usage{
read_importphyto(xlsx, download_latest = FALSE, na = "")
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{xlsx} if available}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}
}
\value{
A \code{data.frame} of formatted water quality data.
}
\description{
Load local phytoplankton cell count file
}
\details{
Pytoplankton cell count data downloaded from ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/, file PlanktonDataList_ThroughCurrentReportMonth.xlsx
}
\examples{
\dontrun{
# file path
xlsx <- 'C:/Users/Owner/Desktop/phyto_data.xlsx'

# load and assign to object
phytodata <- read_importphyto(xlsx)
}
}
\seealso{
\code{\link{read_importwq}}

Other read: \code{\link{read_chkdate}},
  \code{\link{read_dlcurrent}}, \code{\link{read_formfim}},
  \code{\link{read_formphyto}}, \code{\link{read_formwq}},
  \code{\link{read_importfim}}, \code{\link{read_importwq}}
}
\concept{read}
