% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importsealevels.R
\name{read_importsealevels}
\alias{read_importsealevels}
\title{Import monthly sea levels by station from NOAA Tides and Currents}
\usage{
read_importsealevels(
  path_csv,
  download_latest = TRUE,
  df_stations = sealevelstations[, c("station_id", "station_name")],
  api_url = "https://api.tidesandcurrents.noaa.gov/api/prod/datagetter",
  beg_int = 19010101,
  end_int = as.integer(format(lubridate::today(), "\%Y\%m\%d")),
  product = "monthly_mean",
  datum = "stnd",
  time_zone = "lst",
  units = "metric"
)
}
\arguments{
\item{path_csv}{chr string path of CSV file to store tabular output.
(Overwrites existing file.)}

\item{download_latest}{logical to download latest. (Overwrites existing
file.)}

\item{df_stations}{data frame of stations with column \code{station_id} (integer).
Defaults to \link{sealevelstations}, subset to columns \code{station_id} and
\code{station_name}.}

\item{api_url}{chr string URL for NOAA Center for Operational Oceanographic
Products and Services (CO-OPS) API. Defaults to the CO-OPS API for data
retrieval: https://api.tidesandcurrents.noaa.gov/api/prod/datagetter.}

\item{beg_int}{int integer of beginning date in YYYYMMDD format. Defaults to
\code{19010101}.}

\item{end_int}{int integer of ending date in YYYYMMDD format. Defaults to
\code{lubridate::today()}.}

\item{product}{chr string of product type. For options, see \href{https://api.tidesandcurrents.noaa.gov/api/prod/#products}{Data Products | CO-OPS API}.
Defaults to \code{"monthly_mean"}: "verified monthly mean water level data for
the station."}

\item{datum}{chr string of datum. Defaults to \code{"stnd"}: "station datum -
original reference that all data is collected to, uniquely defined for each
station." For options, see \href{https://api.tidesandcurrents.noaa.gov/api/prod/#datum}{Datum | CO-OPS API}}

\item{time_zone}{Time zone. Defaults to \code{"lst"}: "local standard time."
However, this does not get used with the \code{monthly_mean} product. For
options, see \href{https://api.tidesandcurrents.noaa.gov/api/prod/#timezone}{Time Zone | CO-OPS API}.}

\item{units}{chr string of units. Defaults to \code{"metric"} (i.e., meters). For
other options (i.e., \code{"english"}), see \href{https://api.tidesandcurrents.noaa.gov/api/prod/#units}{Units | CO-OPS API}.}
}
\value{
Given the default arguments in (and especially \code{product = "monthly_mean"}), this function returns a data frame from reading
\code{path_csv} (updated if \code{download_latest = TRUE} or newly written if
\code{path_csv} does not exist) having the following fields:
\itemize{
\item \code{station_id}: integer column from input argument \code{df_stations}
\item \code{station_name}: character column from input argument \code{df_stations}
\item \code{date}: first of the month given by \code{year} and \code{month} from API output
\item \code{year}: year of the data
\item \code{month}: month of the data
\item \code{mhhw}: Mean Higher-High Water
\item \code{mhw}: Mean High Water
\item \code{msl}: Mean Sea Level
\item \code{mtl}: Mean Tide Level
\item \code{mlw}: Mean Low Water
\item \code{mllw}: Mean Lower-Low Water
\item \code{dtl}: Mean Diurnal Tide Level
\item \code{gt}: Great Diurnal Range
\item \code{mn}: Mean Range of Tide
\item \code{dhq}: Mean Diurnal High Water Inequality
\item \code{dlq}: Mean Diurnal Low Water Inequality
\item \code{hwi}: Greenwich High Water Interval (in Hours)
\item \code{lwi}: Greenwich Low Water Interval (in Hours)
\item \code{highest}: Highest Tide
\item \code{lowest}: Lowest Tide
\item \code{inferred}: A flag that when set to 1 indicates that the water level value has been inferred
}

For more details on these output data columns, see \href{https://tidesandcurrents.noaa.gov/datum_options.html}{About Tidal Datums | NOAA Tides & Currents}.
}
\description{
Under \href{https://tidesandcurrents.noaa.gov/}{NOAA Tides and Currents}, there is
the NOAA Center for Operational Oceanographic Products and Services (CO-OPS).
This function uses the \href{https://api.tidesandcurrents.noaa.gov/api/prod/}{CO-OPS Data Retrieval API} to extract see level
data by station.
}
\examples{
read_importsealevels(tempfile("sealevels", fileext=".csv"))
}
\concept{read}
