% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwq.R
\name{read_importwq}
\alias{read_importwq}
\title{Load local water quality file}
\usage{
read_importwq(xlsx, download_latest = FALSE, na = c("", "NULL"), all = FALSE)
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{xlsx} if available}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}

\item{all}{logical indicating if all water quality parameters are returned, see details}
}
\value{
A \code{data.frame} of formatted water quality data.
}
\description{
Load local water quality file
}
\details{
Loads the "RWMDataSpreadsheet" worksheet from the file located at \code{xlsx}.  The file is downloaded from \url{https://epcbocc.sharepoint.com/:x:/s/PublishedShare/IQAtFpJjaVHzS6ENmWdNf4csAXO5fC5QFwI8U4b69Yyc8FA?e=tMe7Yy&download=1}.  The files can be viewed at \url{https://epcbocc.sharepoint.com/:f:/s/Share/EiypSSYdsEFCi84Sv_6-t7kBUYaXiIqN0B1n2w57Z_V3kQ?e=NdZQcU}.

Water quality parameters returned by default are total nitrogen (\code{tn}), Secchi depth (\code{sd}), chlorophyll-a (\code{chla}), salinity (top, mid, and bottom depths, \code{Sal_} prefix), water temperature (top, mid, and bottom depths, \code{Temp_Water_} prefix), turbidity (\code{Turbidity_JTU-NTU}), and water color (\code{Color_345_F45}).  Additional qualifier columns for each that include the \code{_Q} suffix are also returned, excluding salinity and water temperature.  All other water quality parameters and qualifiers can be returned by setting \code{all = T}.
}
\examples{
\dontrun{
# file path
xlsx <- '~/Desktop/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx'

# load and assign to object
epcdata <- read_importwq(xlsx, download_latest = T)

# get all water quality parameters
epcdataall <- read_importwq(xlsx, download_latest = T, all = T)
}
}
\seealso{
\code{\link{read_formwq}}, \code{\link{read_importphyto}}
}
\concept{read}
