% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwq.R
\name{read_importwq}
\alias{read_importwq}
\title{Load local water quality file}
\usage{
read_importwq(xlsx, na = "", download_latest_epchc = FALSE,
  connecttimeout = 10, tryurl = FALSE, ...)
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite it not current}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}

\item{download_latest_epchc}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{xlsx} if available}

\item{connecttimeout}{numeric for maximum number of seconds to wait until connection timeout for \code{\link[RCurl]{getURL}}}

\item{tryurl}{logical indicating if \code{\link[RCurl]{getURL}} is repeatedly called in a \code{while} loop if first connection is unsuccessful}

\item{...}{additional arguments passed to \code{\link[readxl]{read_xlsx}}}
}
\value{
A \code{data.frame} of formatted water quality data.
}
\description{
Load local water quality file
}
\details{
Loads the "RWMDataSpreadsheet" worksheet from the file located at \code{xlsx}
}
\examples{
\dontrun{
# file path
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'

# load and assign to object
epcdata <- read_importwq(xlsx)
}
}
\seealso{
\code{\link{read_formwq}}

Other read: \code{\link{read_chkdate}},
  \code{\link{read_dlcurrent}}, \code{\link{read_formwq}}
}
\concept{read}
