% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwq.R
\name{read_importwq}
\alias{read_importwq}
\title{Load local water quality file}
\usage{
read_importwq(xlsx, download_latest = FALSE, na = "")
}
\arguments{
\item{xlsx}{chr string path for local excel file, to overwrite if not current}

\item{download_latest}{logical passed to \code{\link{read_dlcurrent}} to download raw data and compare with existing in \code{xlsx} if available}

\item{na}{chr vector of strings to interpret as \code{NA}, passed to \code{\link[readxl]{read_xlsx}}}
}
\value{
A \code{data.frame} of formatted water quality data.
}
\description{
Load local water quality file
}
\details{
Loads the "RWMDataSpreadsheet" worksheet from the file located at \code{xlsx}.  The file is downloaded from \url{ftp://ftp.epchc.org/EPC_ERM_FTP/WQM_Reports/RWMDataSpreadsheet_ThroughCurrentReportMonth.xlsx}.
}
\examples{
\dontrun{
# file path
xlsx <- 'C:/Users/Owner/Desktop/2018_Results_Updated.xls'

# load and assign to object
epcdata <- read_importwq(xlsx)
}
}
\seealso{
\code{\link{read_formwq}}, \code{\link{read_importphyto}}

Other read: \code{\link{read_chkdate}},
  \code{\link{read_dlcurrent}},
  \code{\link{read_formphyto}}, \code{\link{read_formwq}},
  \code{\link{read_importfim}},
  \code{\link{read_importphyto}}
}
\concept{read}
