% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwqp.R
\name{read_importwqp}
\alias{read_importwqp}
\title{Import water quality data from the Water Quality Portal}
\usage{
read_importwqp(org, trace = F)
}
\arguments{
\item{org}{chr string indicating either \code{"Manatee"} or \code{"Pinellas"}}

\item{trace}{logical indicating whether to display progress messages, default \code{FALSE}}
}
\value{
A data frame containing the imported water quality data for Manatee or Pinellas County.
}
\description{
Import water quality data from the Water Quality Portal
}
\details{
This function retrieves water quality data from the Water Quality Data Exchange API for Manatee or Pinellas County. Based on the input to the \code{org} argument, data for the organization defined as \code{"21FLMANA_WQX"} or \code{"21FLPDEM_WQX"} for Manatee or Pinellas County are returned. It fetches results and station metadata, combines and formats them using the \code{read_formwqmp} function, and returns the processed data as a data frame.  Parameters included are nutrients, chlorophyll, and Secchi depth.
}
\examples{
\dontrun{
# get Manatee County data
mancodata <- read_importwqp(org = 'Manatee', trace = T)

# get Pinellas County data
pincodata <- read_importwqp(org = 'Pinellas', trace = T)
}
}
\seealso{
\code{\link{read_formwqp}}
}
\concept{read}
