% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_importwqp.R
\name{read_importwqp}
\alias{read_importwqp}
\title{Import data from the Water Quality Portal}
\usage{
read_importwqp(org, type, trace = F)
}
\arguments{
\item{org}{chr string indicating the organization identifier, see details}

\item{type}{chr string indicating data type to download, one of \code{"wq"} or \code{"fib"}}

\item{trace}{logical indicating whether to display progress messages, default \code{FALSE}}
}
\value{
A data frame containing the imported data for the selected county
}
\description{
Import data from the Water Quality Portal
}
\details{
This function retrieves data from the Water Quality Portal API (\url{https://www.waterqualitydata.us/}) for selected counties in or around the Tampa Bay watershed. The type of data returned are defined by the \code{type} argument as either \code{"wq"} or \code{"fib"} for water quality of Fecal Indicator Bacteria data, respectively.

The \code{org} argument retrieves data for the specific organization. Valid entries for \code{org} include \code{"21FLDOH_WQX"} (Florida Department of Health), \code{"21FLHILL_WQX"} (Hillsborough County), \code{"21FLMANA_WQX"} (Manatee County), \code{"21FLPASC_WQX"} (Pasco County), \code{"21FLPDEM_WQX"} (Pinellas County), or \code{"21FLPOLK_WQX"} (Polk County).  The naming convention follows the Organization ID in the Water Quality Portal.

The function fetches results and station metadata, combines and formats them using the \code{read_formwqp} function, and returns the processed data as a data frame.  Parameters are specific to the \code{type} argument.
}
\examples{
\dontrun{
# get Manatee County water quality data
mancodata <- read_importwqp(org = '21FLMANA_WQX', type = 'wq', trace = T)

# get Pinellas County FIB data
pincodata <- read_importwqp(org = '21FLPDEM_WQX', type = 'fib', trace = T)
}
}
\seealso{
\code{\link{read_formwqp}}
}
\concept{read}
