% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sealevelstations.R
\docType{data}
\name{sealevelstations}
\alias{sealevelstations}
\title{Sea level stations in Tampa Bay}
\format{
A data frame with 4 rows and 2 variables:
- `station_id` integer station identifier
- `station_name` character station name
}
\usage{
sealevelstations
}
\description{
Stations in Tampa Bay with sea level data through  [CO-OPS Data Retrieval
API](https://api.tidesandcurrents.noaa.gov/api/prod/) for use with
[read_importsealevels()].
}
\concept{data}
\keyword{datasets}
