% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_boxplot.R
\name{show_boxplot}
\alias{show_boxplot}
\title{Plot monthly chlorophyll values and thresholds for a segment}
\usage{
show_boxplot(epcdata, yrcur = NULL, yrrng = c(1975, 2018),
  ptsz = 0.5, bay_segment = c("OTB", "HB", "MTB", "LTB"),
  trgs = NULL)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrcur}{numeric for current year to emphasize, shown as separate red points on the plot}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{ptsz}{numeric indicating point size of observations not in \code{yrcur}}

\item{bay_segment}{chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot monthly chlorophyll values and thresholds for a bay segment
}
\details{
Points not included in \code{yrcur} are plotted over the box plots using \code{\link{ggbeeswarm}{geom_beeswarm}}. Use \code{ptsz = -1} to suppress.
}
\examples{
show_boxplot(epcdata, bay_segment = 'OTB')
}
\seealso{
Other visualize: \code{\link{show_chlmatrix}},
  \code{\link{show_matrix}}, \code{\link{show_thrplot}}
}
\concept{visualize}
