% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_boxplot.R
\name{show_boxplot}
\alias{show_boxplot}
\title{Plot monthly chlorophyll or light attenuation values for a segment}
\usage{
show_boxplot(epcdata, param = c("chla", "la"), yrsel = NULL,
  yrrng = c(1975, 2019), ptsz = 0.5, bay_segment = c("OTB", "HB",
  "MTB", "LTB"), trgs = NULL, family = NA, labelexp = TRUE,
  txtlab = TRUE)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{param}{chr string for which parameter to plot, one of \code{"chla"} for chlorophyll or \code{"la"} for light attenuation}

\item{yrsel}{numeric for year to emphasize, shown as separate red points on the plot}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{ptsz}{numeric indicating point size of observations not in \code{yrsel}}

\item{bay_segment}{chr string for the bay segment, one of "OTB", "HB", "MTB", "LTB"}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{family}{optional chr string indicating font family for text labels}

\item{labelexp}{logical indicating if y axis and target labels are plotted as expressions, default \code{TRUE}}

\item{txtlab}{logical indicating if a text label for the target value is shown in the plot}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot monthly chlorophyll or light attenuation values for a bay segment
}
\details{
Points not included in \code{yrsel} are plotted over the box plots using \code{\link[ggplot2]{position_jitter}}. Use \code{ptsz = -1} to suppress.  The dotted line in the plot shows the large exceedance value.
}
\examples{
show_boxplot(epcdata, bay_segment = 'OTB')
}
\seealso{
Other visualize: \code{\link{show_matrix}},
  \code{\link{show_reactable}},
  \code{\link{show_segmatrix}},
  \code{\link{show_segplotly}}, \code{\link{show_sitemap}},
  \code{\link{show_tbnimatrix}},
  \code{\link{show_tbniscr}}, \code{\link{show_tdlcrk}},
  \code{\link{show_thrplot}}, \code{\link{show_wqmatrix}}
}
\concept{visualize}
