% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_compplot.R
\name{show_compplot}
\alias{show_compplot}
\title{Make a bar plot for transect training group comparisons}
\usage{
show_compplot(
  transect,
  site,
  species = c("Halodule", "Halophila", "Ruppia", "Syringodium", "Thalassia"),
  varplo = c("Abundance", "Blade Length", "Short Shoot Density"),
  base_size = 18,
  xtxt = 10,
  size = 1
)
}
\arguments{
\item{transect}{data frame returned by \code{\link{read_transect}}}

\item{site}{chr string indicating site results to plot}

\item{species}{chr string indicating which species to plot}

\item{varplo}{chr string indicating which variable to plot}

\item{base_size}{numeric indicating text scaling size for plot}

\item{xtxt}{numeric indicating text size for x-axis labels}

\item{size}{numeric indicating line size}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Make a bar plot for transect training group comparisons
}
\examples{
transect <- read_transect(training = TRUE)
show_compplot(transect, site = '1', species = 'Halodule', varplo = 'Abundance')
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectmatrix}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
