% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_compplot.R
\name{show_compplot}
\alias{show_compplot}
\title{Make a bar plot for transect training group comparisons}
\usage{
show_compplot(
  transect,
  yr,
  site,
  species = c("Halodule", "Syringodium", "Thalassia", "Halophila", "Ruppia"),
  varplo = c("Abundance", "Blade Length", "Short Shoot Density"),
  base_size = 18,
  xtxt = 10,
  size = 1
)
}
\arguments{
\item{transect}{data frame returned by \code{\link{read_transect}}}

\item{yr}{numeric for year of training data to plot}

\item{site}{chr string indicating site results to plot}

\item{species}{chr string indicating which species to plot}

\item{varplo}{chr string indicating which variable to plot}

\item{base_size}{numeric indicating text scaling size for plot}

\item{xtxt}{numeric indicating text size for x-axis labels}

\item{size}{numeric indicating line size}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Make a bar plot for transect training group comparisons
}
\examples{
transect <- read_transect(training = TRUE)
show_compplot(transect, yr = 2023, site = '2', species = 'Halodule', varplo = 'Abundance')
}
\concept{show}
