% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_enteromap.R
\name{show_enteromap}
\alias{show_enteromap}
\title{Map Enterococcus results by month, year, and location}
\usage{
show_enteromap(
  fibdata,
  yrsel,
  mosel,
  areasel = NULL,
  wetdry = FALSE,
  precipdata = NULL,
  temporal_window = NULL,
  wet_threshold = NULL
)
}
\arguments{
\item{fibdata}{data frame of Enterococcus sample data as returned by \code{\link{enterodata}} or \code{\link{anlz_fibwetdry}}}

\item{yrsel}{optional numeric to filter data by year}

\item{mosel}{optional numeric to filter data by month}

\item{areasel}{optional character string to filter output by stations in the \code{long_name} column of \code{enterodata}, see details}

\item{wetdry}{logical; if \code{TRUE}, incorporate wet/dry differences (this will result in a call to \code{\link{anlz_fibwetdry}}, in which case \code{temporal_window} and \code{wet_threshold} are required). If \code{FALSE} (default), do not differentiate between wet and dry samples.}

\item{precipdata}{input data frame as returned by \code{\link{read_importrain}}. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object \code{\link{catchprecip}} has this data from 1995-2023 for select Enterococcus stations. If \code{NULL}, defaults to \code{\link{catchprecip}}.}

\item{temporal_window}{numeric; required if \code{wetdry} is \code{TRUE}. number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)}

\item{wet_threshold}{numeric; required if \code{wetdry} is \code{TRUE}. inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period}
}
\value{
A \code{leaflet} map for the selected year, month, and area showing stations and FIB concentration category
}
\description{
Map Enterococcus results by month, year, and location
}
\details{
Placing the mouse cursor over an item on the map will reveal additional information about a station.
}
\examples{
show_enteromap(enterodata, yrsel = 2020, mosel = 9)

# wet/dry samples
show_enteromap(enterodata, yrsel = 2020, mosel = 9, wetdry = TRUE,
               temporal_window = 2, wet_threshold = 0.5)

# Old Tampa Bay only
show_enteromap(enterodata, yrsel = 2020, mosel = 9, areasel = "Old Tampa Bay")
}
\seealso{
\code{\link{anlz_enteromap}} for details on the categories
}
\concept{anlz}
