% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fibmap.R
\name{show_fibmap}
\alias{show_fibmap}
\title{Map Fecal Indicator Bacteria (FIB) results by month, year, and location}
\usage{
show_fibmap(fibdata, yrsel, mosel, areasel)
}
\arguments{
\item{fibdata}{input FIB \code{data.frame} as returned by \code{\link{read_importfib}}}

\item{yrsel}{optional numeric value to filter output by years in \code{fibdata}}

\item{mosel}{optional numeric value to filter output by month in \code{fibdata}}

\item{areasel}{optional character string to filter output by stations in the \code{"Alafia"} or \code{"Hillsborough"} River basins, see details}
}
\value{
A \code{leaflet} map for the selected year, month, and area showing stations and FIB concentration category
}
\description{
Map Fecal Indicator Bacteria (FIB) results by month, year, and location
}
\details{
Placing the mouse cursor over an item on the map will reveal additional information about a station.
}
\examples{
show_fibmap(fibdata, yrsel = 2020, mosel = 7, areasel = 'Hillsborough')
}
\seealso{
\code{\link{anlz_fibmap}} for details on the categories
}
\concept{anlz}
