% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_fibmatmap.R
\name{show_fibmatmap}
\alias{show_fibmatmap}
\title{Map Fecal Indicator Bacteria matrix results by year}
\usage{
show_fibmatmap(
  fibdata,
  yrsel,
  areasel,
  indic,
  threshold = NULL,
  lagyr = 3,
  subset_wetdry = c("all", "wet", "dry"),
  precipdata = NULL,
  temporal_window = NULL,
  wet_threshold = NULL
)
}
\arguments{
\item{fibdata}{input data frame as returned by \code{\link{read_importfib}} or \code{\link{read_importentero}}}

\item{yrsel}{numeric value indicating the year to map}

\item{areasel}{vector of bay segment or area names to include, see details}

\item{indic}{character for choice of fecal indicator. Allowable options are \code{fcolif} for fecal coliform, or \code{entero} for Enterococcus. A numeric column in the data frame must have this name.}

\item{threshold}{optional numeric for threshold against which to calculate exceedances for the indicator bacteria of choice. If not provided, defaults to 400 for \code{fcolif} and 130 for \code{entero}.}

\item{lagyr}{numeric for year lag to calculate categories, see details}

\item{subset_wetdry}{character, subset data frame to only wet or dry samples as defined by \code{wet_threshold} and \code{temporal_window}? Defaults to \code{"all"}, which will not subset. If \code{"wet"} or \code{"dry"} is specified, \code{\link{anlz_fibwetdry}} is called using the further specified parameters, and the data frame is subsetted accordingly.}

\item{precipdata}{input data frame as returned by \code{\link{read_importrain}}. columns should be: station, date (yyyy-mm-dd), rain (in inches). The object \code{\link{catchprecip}} has this data from 1995-2023 for select Enterococcus stations. If \code{NULL}, defaults to \code{\link{catchprecip}}.}

\item{temporal_window}{numeric; required if \code{subset_wetdry} is not \code{"all"}. number of days precipitation should be summed over (1 = day of sample only; 2 = day of sample + day before; etc.)}

\item{wet_threshold}{numeric; required if \code{subset_wetdry} is not \code{"all"}. inches accumulated through the defined temporal window, above which a sample should be defined as being from a 'wet' time period}
}
\value{
A \code{leaflet} map for the selected year and area showing station matrix scores.  Bay segment scores are also shown if the input data are not from EPCHC.
}
\description{
Map Fecal Indicator Bacteria matrix results by year
}
\details{
Placing the mouse cursor over an item on the map will reveal additional information about a segment or station.

If the input to \code{fibdata} is from EPCHC (from \code{\link{read_importfib}}), valid entries for \code{areasel} include 'Alafia River', 'Hillsborough River', 'Big Bend', 'Cockroach Bay', 'East Lake Outfall', 'Hillsborough Bay', 'Little Manatee River', 'Lower Tampa Bay', 'McKay Bay', 'Middle Tampa Bay', 'Old Tampa Bay', 'Palm River', 'Tampa Bypass Canal', and 'Valrico Lake'.  If the input data is not from EPCHC (from \code{\link{read_importentero}}), valid entries for \code{areasel} include 'OTB', 'HB', 'MTB', 'LTB', 'BCB', and 'MR'.

Bay segment matrix categories are shown in addition to stations if \code{fibsel} input is not from EPCHC (from \code{\link{read_importentero}}).  Stations for these data were chosen specifically as downstream endpoints for each bay segment, whereas the EPCHC data are not appropriate for estimating bay segment outcomes.
}
\examples{
show_fibmatmap(enterodata, yrsel = 2020, indic = 'entero')

# wet/dry samples
show_fibmatmap(enterodata, yrsel = 2020, wetdry = TRUE, indic = 'entero',
               temporal_window = 2, wet_threshold = 0.5)

# Old Tampa Bay only
show_fibmatmap(enterodata, yrsel = 2020, areasel = "OTB", indic = 'entero')

# EPCHC data
show_fibmatmap(fibdata, yrsel = 2016, indic = 'fcolif',
   areasel = c("Hillsborough River", "Alafia River"))
}
\seealso{
\code{\link{anlz_fibmatrix}} for details on the categories
}
\concept{show}
