% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_hmpreport.R
\name{show_hmpreport}
\alias{show_hmpreport}
\title{Show Habitat Master Plan progress report card}
\usage{
show_hmpreport(acres, subtacres, hmptrgs, typ, text = TRUE)
}
\arguments{
\item{acres}{\code{data.frame} for intertidal and supratidal land use and cover of habitat types for each year of data}

\item{subtacres}{\code{data.frame} for subtidal cover of habitat types for each year of data}

\item{hmptrgs}{\code{data.frame} of Habitat Master Plan targets and goals}

\item{typ}{character string indicating \code{"targets"} or \code{"goals"}}

\item{text}{logical indicating if proportion of target or goal met for habitat types is shown in each cell types}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object showing overall progress in attaining Habitat Master Plan targets or goals.
}
\description{
Show Habitat Master Plan progress report card
}
\details{
Colors indicate target not met, trending below (red), target met, trending below (yellow), target not met, trending above (light green), and target met, trending above (green).  Numbers in cell show the proportion of the target or goal met at each year where data are available.
}
\examples{
# view summarized data for report card, targets
show_hmpreport(acres, subtacres, hmptrgs, typ = "targets")

# view summarized data for report card, goals
show_hmpreport(acres, subtacres, hmptrgs, typ = "goals")
}
\concept{show}
