% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_hmpreport.R
\name{show_hmpreport}
\alias{show_hmpreport}
\title{Show Habitat Master Plan progress report card}
\usage{
show_hmpreport(
  acres,
  subtacres,
  hmptrgs,
  typ,
  text = TRUE,
  family = NA,
  plotly = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{acres}{\code{data.frame} for intertidal and supratidal land use and cover of habitat types for each year of data}

\item{subtacres}{\code{data.frame} for subtidal cover of habitat types for each year of data}

\item{hmptrgs}{\code{data.frame} of Habitat Master Plan targets and goals}

\item{typ}{character string indicating \code{"targets"} or \code{"goals"}}

\item{text}{logical indicating if proportion of target or goal met for habitat types is shown in each cell types}

\item{family}{optional chr string indicating font family for text labels}

\item{plotly}{logical if matrix is created using plotly}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object showing overall progress in attaining Habitat Master Plan targets or goals.
}
\description{
Show Habitat Master Plan progress report card
}
\details{
Colors indicate target or goal not met, trending below (red), target or goal met, trending below (yellow), target or goal not met, trending above (light green), and target or goal  met, trending above (green).  Numbers in cell show the proportion of the target or goal met at each year where data are available.

The report card provides no information on artificial reefs, living shorelines, and hard bottom habitats.  These habitats are not assessed in routine data products from the Southwest Florida Water Management District, although targets and goals are provided in the Habitat Master Plan.
}
\examples{
# view summarized data for report card, targets
show_hmpreport(acres, subtacres, hmptrgs, typ = "targets")

# view summarized data for report card, goals
show_hmpreport(acres, subtacres, hmptrgs, typ = "goals")
}
\concept{show}
