% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_hmpreport.R
\name{show_hmpreport}
\alias{show_hmpreport}
\title{Show Habitat Master Plan progress report card}
\usage{
show_hmpreport(
  acres,
  subtacres,
  hmptrgs,
  typ,
  strata = c("Subtidal", "Intertidal", "Supratidal"),
  ycollapse = FALSE,
  text = TRUE,
  family = NA,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{acres}{\code{data.frame} for intertidal and supratidal land use and cover of habitat types for each year of data}

\item{subtacres}{\code{data.frame} for subtidal cover of habitat types for each year of data}

\item{hmptrgs}{\code{data.frame} of Habitat Master Plan targets and goals}

\item{typ}{character string indicating \code{"targets"} or \code{"goals"}}

\item{strata}{character string indicating with strata to plot, one to many of \code{"Subtidal"}, \code{"Intertidal"}, and \code{"Supratidal"}}

\item{ycollapse}{logical indicating if the y-axis is collapsed to year with data, see details}

\item{text}{logical indicating if proportion of target or goal met for habitat types is shown in each cell types}

\item{family}{optional chr string indicating font family for text labels}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object showing overall progress in attaining Habitat Master Plan targets or goals.
}
\description{
Show Habitat Master Plan progress report card
}
\details{
Colors indicate target or goal not met, trending below (red), target or goal met, trending below (yellow), target or goal not met, trending above (light green), and target or goal  met, trending above (green).  Numbers in cell show the proportion of the target or goal met at each year where data are available.

The report card provides no information on artificial reefs, living shorelines, and hard bottom habitats.  These habitats are not assessed in routine data products from the Southwest Florida Water Management District, although targets and goals are provided in the Habitat Master Plan.

The subtidal data in \code{subtacres} and the inter/supratidal data in \code{acres} are provided as different datasets by the Southwest Florida Water Management District.  The years in each dataset typically do not match and each dataset is collected at approximate 2 to 3 year intervals.  By default, year on the y-axis is shown as a continuous variable, where gaps are shown in years when each dataset was unavailable.  Use \code{ycollapse = TRUE} to remove years without data.
}
\examples{
# view summarized data for report card, targets
show_hmpreport(acres, subtacres, hmptrgs, typ = "targets")

# view summarized data for report card, goals
show_hmpreport(acres, subtacres, hmptrgs, typ = "goals")

# remove empty years
show_hmpreport(acres, subtacres, hmptrgs, typ = "targets", ycollapse = TRUE)

# select only subtidal
show_hmpreport(acres, subtacres, hmptrgs, typ = "targets", ycollapse = TRUE, strata = 'Subtidal')
}
\concept{show}
