% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_matrix.R
\name{show_matrix}
\alias{show_matrix}
\title{Create a colorized table for indicator reporting}
\usage{
show_matrix(epcdata)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}
}
\value{
An object of class \code{gt_tbl}
}
\description{
Create a colorized table for indicator reporting
}
\examples{
\dontrun{
show_matrix(epcdata)
}
}
\seealso{
\code{\link[gt]{gt}}

Other visualize: \code{\link{show_thrplot}}
}
\concept{visualize}
