% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_matrix.R
\name{show_matrix}
\alias{show_matrix}
\title{Create a colorized table for indicator reporting}
\usage{
show_matrix(epcdata, txtsz = 3, trgs = NULL, yrrng = c(1975, 2018),
  bay_segment = c("OTB", "HB", "MTB", "LTB"), asreact = FALSE,
  nrows = 10, abbrev = FALSE, family = NA, historic = FALSE)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{txtsz}{numeric for size of text in the plot, applies only if \code{tab = FALSE}}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{yrrng}{numeric vector indicating min, max years to include}

\item{bay_segment}{chr string for bay segments to include, one to all of "OTB", "HB", "MTB", "LTB"}

\item{asreact}{logical indicating if a \code{\link[reactable]{reactable}} object is returned}

\item{nrows}{if \code{asreact = TRUE}, a numeric specifying number of rows in the table}

\item{abbrev}{logical indicating if text labels in the plot are abbreviated as the first letter}

\item{family}{optional chr string indicating font family for text labels}

\item{historic}{logical if historic data are used from 2005 and earlier}
}
\value{
A static \code{\link[ggplot2]{ggplot}} object is returned if \code{asreact = FALSE}, otherwise a \code{\link[reactable]{reactable}} table is returned
}
\description{
Create a colorized table for indicator reporting
}
\examples{
show_matrix(epcdata)
}
\seealso{
\code{\link{show_wqmatrix}}, \code{\link{show_segmatrix}}

Other visualize: \code{\link{show_boxplot}},
  \code{\link{show_reactable}},
  \code{\link{show_segmatrix}}, \code{\link{show_sitemap}},
  \code{\link{show_tdlcrk}}, \code{\link{show_thrplot}},
  \code{\link{show_wqmatrix}}
}
\concept{visualize}
