% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_matrixplotly.R
\name{show_matrixplotly}
\alias{show_matrixplotly}
\title{Creates a plotly matrix from any matrix function input}
\usage{
show_matrixplotly(mat, family = NA, tooltip = "Result")
}
\arguments{
\item{mat}{input matrix as output from \code{\link{show_matrix}}, \code{\link{show_segmatrix}}, \code{\link{show_wqmatrix}}, or \code{\link{show_tbnimatrix}}}

\item{family}{optional chr string indicating font family for text labels}

\item{tooltip}{chr string indicating the column name for tooltip}
}
\value{
A plotly data object
}
\description{
Creates a plotly matrix from any matrix function input
}
\examples{
mat <- show_wqmatrix(epcdata)
show_matrixplotly(mat)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
