% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sedimentpelave.R
\name{show_sedimentpelave}
\alias{show_sedimentpelave}
\title{Plot summary of PEL averages by bay segment}
\usage{
show_sedimentpelave(
  sedimentdata,
  yrrng = c(1993, 2023),
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"),
  funding_proj = "TBEP",
  alph = 1,
  ylim = c(0, 0.4),
  lnsz = 1,
  base_size = 12,
  plotly = FALSE,
  family = NA,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{sedimentdata}{input sediment \code{data.frame} as returned by \code{\link{read_importsediment}}}

\item{yrrng}{numeric vector indicating min, max years to include, use single year for one year of data}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{funding_proj}{chr string for the funding project, one to many of "TBEP" (default), "TBEP-Special", "Apollo Beach", "Janicki Contract", "Rivers", "Tidal Streams"}

\item{alph}{numeric indicating alpha value for score category colors}

\item{ylim}{numeric for y axis limits}

\item{lnsz}{numeric for line size}

\item{base_size}{numeric indicating text scaling size for plot}

\item{plotly}{logical if matrix is created using plotly}

\item{family}{optional chr string indicating font family for text labels}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object or a \code{\link[plotly]{plotly}} object if \code{plotly = TRUE} showing PEL averages (average of the averages) and 95\% confidence intervals for each bay segment
}
\description{
Plot summary of PEL averages by bay segment
}
\details{
Lines for the grades are shown on the plot. Confidence intervals may not be shown for segments with insufficient data.
}
\examples{
show_sedimentpelave(sedimentdata)
}
\concept{show}
