% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sitemap.R
\name{show_sitemap}
\alias{show_sitemap}
\title{Map site attainment categories for a selected year}
\usage{
show_sitemap(
  epcdata,
  yrsel,
  mosel = c(1, 12),
  param = c("chla", "la"),
  trgs = NULL,
  thrs = FALSE,
  showseg = TRUE,
  partialyr = FALSE
)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrsel}{numeric for year to plot}

\item{mosel}{optional numeric of length one or two for mapping results for a specific month or month range in a given year, default full year}

\item{param}{chr string for which parameter to plot, one of \code{"chla"} for chlorophyll or \code{"la"} for light attenuation}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}, only applies if \code{mosel = c(1, 12)}}

\item{thrs}{logical indicating if attainment category is relative to targets (default) or thresholds, passed to \code{\link{anlz_attainsite}}, only applies if \code{mosel = c(1, 12)}}

\item{showseg}{logical indicating of bay segment labels are included}

\item{partialyr}{logical indicating if incomplete annual data for the most recent year are approximated by five year monthly averages for each parameter, only applies if \code{mosel = c(1, 12)}}
}
\value{
A static \code{ggplot} object is returned
}
\description{
Map site attainment categories for a selected year
}
\examples{
show_sitemap(epcdata, yrsel = 2025)
}
\concept{show}
