% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_sitemap.R
\name{show_sitemap}
\alias{show_sitemap}
\title{Map site attainment categories for a selected year}
\usage{
show_sitemap(epcdata, yrsel, trgs = NULL, thrs = FALSE)
}
\arguments{
\item{epcdata}{data frame of epc data returned by \code{\link{read_importwq}}}

\item{yrsel}{numeric for year to plot}

\item{trgs}{optional \code{data.frame} for annual bay segment water quality targets, defaults to \code{\link{targets}}}

\item{thrs}{logical indicating if attainment category is relative to targets (default) or thresholds, passed to \code{\link{anlz_attainsite}}}
}
\value{
A static \code{ggplot} object is returned
}
\description{
Map site attainment categories for a selected year
}
\examples{
show_sitemap(epcdata, yrsel = 2018)
}
\seealso{
Other visualize: \code{\link{show_boxplot}},
  \code{\link{show_matrix}}, \code{\link{show_reactable}},
  \code{\link{show_segmatrix}}, \code{\link{show_tdlcrk}},
  \code{\link{show_thrplot}}, \code{\link{show_wqmatrix}}
}
\concept{visualize}
