% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbbimatrix.R
\name{show_tbbimatrix}
\alias{show_tbbimatrix}
\title{Plot a matrix of Tampa Bay Benthic Index scores over time by bay segment}
\usage{
show_tbbimatrix(
  tbbiscr,
  bay_segment = c("HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"),
  yrrng = c(1993, 2018),
  alph = 1,
  txtsz = 3,
  family = NA,
  rev = FALSE,
  position = "top",
  plotly = FALSE
)
}
\arguments{
\item{tbbiscr}{input data frame as returned by \code{\link{anlz_tbbiscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB", "All", "All (wt)"}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{alph}{numeric indicating alpha value for score category colors}

\item{txtsz}{numeric for size of text in the plot}

\item{family}{optional chr string indicating font family for text labels}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{position}{chr string of location for bay segment labels, default on top, passed to \code{\link[ggplot2]{scale_x_discrete}}}

\item{plotly}{logical if matrix is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBBI scores for each bay segment if \code{plotly = FALSE}, otherwise a \code{\link[plotly]{plotly}} object
}
\description{
Plot a matrix of Tampa Bay Benthic Index scores over time by bay segment
}
\details{
Additional summaries are provided for the entire bay, as a summary across categories ("All") and a summary weighted across the relative sizes of each bay segment ("All (wt)").
}
\examples{
tbbiscr <- anlz_tbbiscr(benthicdata)
show_tbbimatrix(tbbiscr)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectavespp}()},
\code{\link{show_transectmatrix}()},
\code{\link{show_transectsum}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
