% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbnimatrix.R
\name{show_tbnimatrix}
\alias{show_tbnimatrix}
\title{Plot a matrix of Tampa Bay Nekton Index scores over time by bay segment}
\usage{
show_tbnimatrix(tbniscr, bay_segment = c("OTB", "HB", "MTB", "LTB"),
  perc = c(32, 46), alph = 0.3, txtsz = 3, family = NA)
}
\arguments{
\item{txtsz}{numeric for size of text in the plot}

\item{family}{optional chr string indicating font family for text labels}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBNI scores for each bay segment
}
\description{
Plot a matrix of Tampa Bay Nekton Index scores over time by bay segment
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
show_tbnimatrix(tbniscr)
}
\seealso{
Other visualize: \code{\link{show_boxplot}},
  \code{\link{show_matrix}}, \code{\link{show_reactable}},
  \code{\link{show_segmatrix}},
  \code{\link{show_segplotly}}, \code{\link{show_sitemap}},
  \code{\link{show_tbniscr}}, \code{\link{show_tdlcrk}},
  \code{\link{show_thrplot}}, \code{\link{show_wqmatrix}}
}
\concept{visualize}
