% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbnimatrix.R
\name{show_tbnimatrix}
\alias{show_tbnimatrix}
\title{Plot a matrix of Tampa Bay Nekton Index scores over time by bay segment}
\usage{
show_tbnimatrix(
  tbniscr,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  perc = c(32, 46),
  alph = 0.3,
  txtsz = 3,
  family = NA,
  rev = FALSE,
  position = "top",
  plotly = FALSE
)
}
\arguments{
\item{tbniscr}{input data frame as returned by \code{\link{anlz_tbniscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB"}

\item{perc}{numeric values indicating break points for score categories}

\item{alph}{numeric indicating alpha value for score category colors}

\item{txtsz}{numeric for size of text in the plot}

\item{family}{optional chr string indicating font family for text labels}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{position}{chr string of location for bay segment labels, default on top, passed to \code{\link[ggplot2]{scale_x_discrete}}}

\item{plotly}{logical if matrix is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBNI scores for each bay segment
}
\description{
Plot a matrix of Tampa Bay Nekton Index scores over time by bay segment
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
show_tbnimatrix(tbniscr)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
