% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbniscr.R
\name{show_tbniscr}
\alias{show_tbniscr}
\title{Plot Tampa Bay Nekton Index scores over time by bay segment}
\usage{
show_tbniscr(
  tbniscr,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  perc = c(32, 46),
  alph = 0.3,
  ylim = c(22, 58),
  rev = FALSE,
  plotly = FALSE
)
}
\arguments{
\item{tbniscr}{input dat frame as returned by \code{\link{anlz_tbniscr}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB"}

\item{perc}{numeric values indicating break points for score categories}

\item{alph}{numeric indicating alpha value for score category colors}

\item{ylim}{numeric for y axis limits}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{plotly}{logical if matrix is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time in TBNI scores for each bay segment or a \code{\link[plotly]{plotly}} object if \code{plotly = TRUE}
}
\description{
Plot Tampa Bay Nekton Index scores over time by bay segment
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
show_tbniscr(tbniscr)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectavespp}()},
\code{\link{show_transectmatrix}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
