% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tbniscrseas.R
\name{show_tbniscrseas}
\alias{show_tbniscrseas}
\title{Plot seasonal Tampa Bay Nekton Index results by month and bay segment}
\usage{
show_tbniscrseas(
  tbniscr,
  bay_segment = c("OTB", "HB", "MTB", "LTB"),
  yr,
  metric = NULL,
  perc = c(32, 46),
  alph = 1,
  plotly = FALSE,
  family = "sans",
  width = NULL,
  height = NULL
)
}
\arguments{
\item{tbniscr}{input dat frame as returned by \code{\link{anlz_tbniscr}}}

\item{bay_segment}{chr string for the bay segment to plot, one of "OTB", "HB", "MTB", "LTB"}

\item{yr}{numeric for the year to plot}

\item{metric}{chr string for the column in \code{tbniscr} to plot on the y-axis, defaults to \code{NULL} to plot \code{"TBNI_Score"}, otherwise an individual TBNI metric can be specified, one of "TBNI_Score", "NumTaxa", "BenthicTaxa", "TaxaSelect", "NumGuilds", or "Shannon"}

\item{perc}{numeric values indicating break points for score categories, only used if \code{metric} is \code{NULL}}

\item{alph}{numeric indicating alpha value for the score category background colors and the boxplot fill, only used if \code{metric} is \code{NULL}}

\item{plotly}{logical if matrix is created using plotly}

\item{family}{optional chr string indicating font family for text labels}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing monthly results for \code{yr} in \code{bay_segment} for the selected metric, or a \code{\link[plotly]{plotly}} object if \code{plotly = TRUE}
}
\description{
Plot seasonal Tampa Bay Nekton Index results by month and bay segment
}
\details{
Boxplots show the distribution of site-level results by month for \code{yr} in \code{bay_segment}, with individual site values overlaid as jittered points.  If \code{metric} is \code{NULL}, the TBNI score is plotted with the same red/yellow/green score category background and break lines from \code{perc} as in \code{\link{show_tbniscr}}.  If \code{metric} is specified, the selected metric is plotted as raw values (note that scored metrics cannot be shown).  Metric options include \code{"TBNI_Score"} or \code{NULL} (default), \code{"NumTaxa"}, \code{"BenthicTaxa"}, \code{"TaxaSelect"}, \code{"NumGuilds"}, and \code{"Shannon"}.
}
\examples{
tbniscr <- anlz_tbniscr(fimdata)
show_tbniscrseas(tbniscr, yr = 2018)
}
\concept{show}
