% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrk.R
\name{show_tdlcrk}
\alias{show_tdlcrk}
\title{Make a map for tidal creek report card}
\usage{
show_tdlcrk(dat, weight = 1.5)
}
\arguments{
\item{dat}{input creek score data returned from \code{\link{anlz_tdlcrk}}}

\item{weight}{numeric for weight of polylines, passed to \code{\link[leaflet]{addPolylines}}}
}
\value{
A \code{\link[leaflet]{leaflet}} object
}
\description{
Make a map for tidal creek report card
}
\examples{
dat <- anlz_tdlcrk(tidalcreeks, iwrraw, yr = 2018)
show_tdlcrk(dat)
}
\concept{show}
