% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkindic.R
\name{show_tdlcrkindic}
\alias{show_tdlcrkindic}
\title{Plotly barplots of tidal creek context indicators}
\usage{
show_tdlcrkindic(
  id,
  cntdat,
  yr = 2025,
  thrsel = FALSE,
  pal = c("#5C4A42", "#427355", "#004F7E")
)
}
\arguments{
\item{id}{numeric indicating the \code{id} number of the tidal creek to plot}

\item{cntdat}{output from \code{\link{anlz_tdlcrkindic}}}

\item{yr}{numeric indicating reference year}

\item{thrsel}{logical if threshold lines and annotations are shown on the plots}

\item{pal}{vector of colors for the palette}
}
\value{
A plotly object
}
\description{
Plotly barplots of tidal creek context indicators
}
\examples{
cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2025)

set.seed(123)
id <- sample(unique(cntdat$id), 1)
show_tdlcrkindic(id, cntdat, thrsel = TRUE)
}
\concept{show}
