% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tdlcrkradar.R
\name{show_tdlcrkradar}
\alias{show_tdlcrkradar}
\title{Radar plots for tidal creek indicators}
\usage{
show_tdlcrkradar(
  id,
  cntdat,
  col = "#338080E6",
  ptsz = 1,
  lbsz = 0.8,
  valsz = 1,
  brdwd = 5
)
}
\arguments{
\item{id}{numeric indicating the \code{id} number of the tidal creek to plot}

\item{cntdat}{output from \code{\link{anlz_tdlcrkindic}}}

\item{col}{color input for polygon and line portions}

\item{ptsz}{numeric size of points}

\item{lbsz}{numeric for size of text labels}

\item{valsz}{numeric for size of numeric value labels}

\item{brdwd}{numeric for polygon border width}
}
\value{
A radar plot
}
\description{
Radar plots for tidal creek indicators
}
\details{
See details in \code{\link{anlz_tdlcrkindic}} for an explanation of the indicators

Internal code borrowed heavily from the radarchart function in the fmsb package.
}
\examples{
cntdat <- anlz_tdlcrkindic(tidalcreeks, iwrraw, yr = 2022, radar = TRUE)

set.seed(123)
id <- sample(unique(cntdat$id), 1)
show_tdlcrkradar(id, cntdat)
}
\concept{show}
