% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_transect.R
\name{show_transect}
\alias{show_transect}
\title{Plot results for a seagrass transect by time and location}
\usage{
show_transect(
  transect,
  site,
  species = c("Halodule", "Syringodium", "Thalassia", "Halophila", "Ruppia",
    "Caulerpa"),
  yrrng = c(1998, 2021),
  varplo = c("Abundance", "Blade Length", "Short Shoot Density"),
  base_size = 12,
  facet = FALSE,
  ncol = NULL,
  plotly = FALSE,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{transect}{data frame returned by \code{\link{read_transect}}}

\item{site}{chr string indicating site results to plot}

\item{species}{chr string indicating one to many of which species to plot}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{varplo}{chr string indicating which variable to plot}

\item{base_size}{numeric indicating text scaling size for plot}

\item{facet}{logical indicating if plots are separated into facets by species}

\item{ncol}{numeric indicating number of columns if \code{facet = TRUE}}

\item{plotly}{logical if plot is created using plotly}

\item{width}{numeric for width of the plot in pixels, only applies of \code{plotly = TRUE}}

\item{height}{numeric for height of the plot in pixels, only applies of \code{plotly = TRUE}}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot results for a seagrass transect by time and location
}
\details{
All sites along a transect that were surveyed are shown in the plot, including those where the selected species were not found.  The latter is colored in grey hollow points.  Species options include Halodule, Syringodium, Thalassia, Halophila, Ruppia, and/or Caulerpa.

Note that if \code{plotly = TRUE}, the size legend is not shown.
}
\examples{
\dontrun{
transect <- read_transect()
}

# one species
show_transect(transect, site = 'S3T10', species = 'Halodule', varplo = 'Abundance')

# multiple species, one plot
show_transect(transect, site = 'S3T10',
  species = c('Halodule', 'Syringodium', 'Thalassia', 'Halophila', 'Ruppia', 'Caulerpa'),
  varplo = 'Abundance')

# multiple species, multiple plots
show_transect(transect, site = 'S3T10',
  species = c('Halodule', 'Syringodium', 'Thalassia', 'Halophila', 'Ruppia', 'Caulerpa'),
  varplo = 'Abundance', facet = TRUE)
}
\concept{show}
