% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_transect.R
\name{show_transect}
\alias{show_transect}
\title{Plot results for a seagrass transect by time and location}
\usage{
show_transect(
  transect,
  site,
  species = c("Halodule", "Halophila", "Ruppia", "Syringodium", "Thalassia"),
  varplo = c("Abundance", "Blade Length", "Short Shoot Density"),
  base_size = 12,
  plotly = FALSE
)
}
\arguments{
\item{transect}{data frame returned by \code{\link{read_transect}}}

\item{site}{chr string indicating site results to plot}

\item{species}{chr string indicating which species to plot}

\item{varplo}{chr string indicating which variable to plot}

\item{base_size}{numeric indicating text scaling size for plot}

\item{plotly}{logical if plot is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object
}
\description{
Plot results for a seagrass transect by time and location
}
\details{
All sites along at transect that were surveyed are shown in the plot, including those where the selected species was not found.  The latter is colored in red.
}
\examples{
\dontrun{
transect <- read_transect()
}
show_transect(transect, site = 'S3T10', species = 'Halodule', varplo = 'Abundance')
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectavespp}()},
\code{\link{show_transectmatrix}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
