% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_transectavespp.R
\name{show_transectavespp}
\alias{show_transectavespp}
\title{Show annual averages of seagrass frequency occurrence by bay segments, year, and species}
\usage{
show_transectavespp(
  transectocc,
  bay_segment = c("OTB", "HB", "MTB", "LTB", "BCB"),
  yrrng = c(1998, 2019),
  species = c("Halodule", "Syringodium", "Thalassia", "Ruppia", "Halophila spp.",
    "Caulerpa spp."),
  alph = 1,
  family = NA,
  plotly = FALSE
)
}
\arguments{
\item{transectocc}{data frame returned by \code{\link{anlz_transectocc}}}

\item{bay_segment}{chr string for the bay segment, one to many of "OTB", "HB", "MTB", "LTB", "BCB"}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{species}{chr string of species to summarize, one to many of "Halodule", "Syringodium", "Thalassia", "Ruppia", "Halophila spp.", "Caulerpa spp."}

\item{alph}{numeric indicating alpha value for score category colors}

\item{family}{optional chr string indicating font family for text labels}

\item{plotly}{logical if matrix is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time by speies for selected bay segments if \code{plotly = FALSE}, otherwise a \code{\link[plotly]{plotly}} object
}
\description{
Show annual averages of seagrass frequency occurrence by bay segments, year, and species
}
\details{
Results are based on averages across species by date and transect in each bay segment
}
\examples{
\dontrun{
transect <- read_transect()
}
transectocc <- anlz_transectocc(transect)
show_transectavespp(transectocc)
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectmatrix}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
