% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_transectmatrix.R
\name{show_transectmatrix}
\alias{show_transectmatrix}
\title{Show matrix of seagrass frequency occurrence by bay segments and year}
\usage{
show_transectmatrix(
  transectocc,
  bay_segment = c("OTB", "HB", "MTB", "LTB", "BCB"),
  yrrng = c(1998, 2019),
  alph = 1,
  txtsz = 3,
  family = NA,
  rev = FALSE,
  position = "top",
  plotly = FALSE
)
}
\arguments{
\item{transectocc}{data frame returned by \code{\link{anlz_transectocc}}}

\item{bay_segment}{chr string for the bay segment, one to many of "HB", "OTB", "MTB", "LTB", "TCB", "MR", "BCB"}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{alph}{numeric indicating alpha value for score category colors}

\item{txtsz}{numeric for size of text in the plot}

\item{family}{optional chr string indicating font family for text labels}

\item{rev}{logical if factor levels for bay segments are reversed}

\item{position}{chr string of location for bay segment labels, default on top, passed to \code{\link[ggplot2]{scale_x_discrete}}}

\item{plotly}{logical if matrix is created using plotly}
}
\value{
A \code{\link[ggplot2]{ggplot}} object showing trends over time for each bay segment if \code{plotly = FALSE}, otherwise a \code{\link[plotly]{plotly}} object
}
\description{
Show matrix of seagrass frequency occurrence by bay segments and year
}
\details{
Results are based on averages across species by date and transect in each bay segment
}
\examples{
\dontrun{
transect <- read_transect()
}
transectocc <- anlz_transectocc(transect)
show_transectmatrix(transectocc)
}
\references{
#' @references
This plot is a representation of Table 1 in R. Johansson (2016) Seagrass Transect Monitoring in Tampa Bay: A Summary of Findings from 1997 through 2015, Technical report #08-16, Tampa Bay Estuary Program, St. Petersburg, Florida.
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectavespp}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
