% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_transectsum.R
\name{show_transectsum}
\alias{show_transectsum}
\title{Plot frequency occurrence for a seagrass transect by time for all species}
\usage{
show_transectsum(
  transectocc,
  site,
  species = c("Halodule", "Syringodium", "Thalassia", "Halophila", "Ruppia",
    "Caulerpa"),
  yrrng = c(1998, 2020),
  abund = FALSE
)
}
\arguments{
\item{transectocc}{data frame returned by \code{\link{anlz_transectocc}}}

\item{site}{chr string indicating site results to plot}

\item{species}{chr string indicating which species to plot}

\item{yrrng}{numeric indicating year ranges to evaluate}

\item{abund}{logical indicating if abundance averages are plotted instead of frequency occurrence}
}
\value{
A \code{\link[plotly]{plotly}} object
}
\description{
Plot frequency occurrence for a seagrass transect by time for all species
}
\details{
This plot provides a quick visual assessment of how frequency occurrence or abundance for multiple species has changed over time at a selected transect.
}
\examples{
\dontrun{
transect <- read_transect()
}
transectocc <- anlz_transectocc(transect)
show_transectsum(transectocc, site = 'S3T10')
}
\seealso{
Other visualize: 
\code{\link{show_boxplot}()},
\code{\link{show_compplot}()},
\code{\link{show_matrixplotly}()},
\code{\link{show_matrix}()},
\code{\link{show_reactable}()},
\code{\link{show_segmatrix}()},
\code{\link{show_segplotly}()},
\code{\link{show_sitemap}()},
\code{\link{show_tbbimatrix}()},
\code{\link{show_tbnimatrix}()},
\code{\link{show_tbniscrall}()},
\code{\link{show_tbniscrplotly}()},
\code{\link{show_tbniscr}()},
\code{\link{show_tdlcrk}()},
\code{\link{show_thrplot}()},
\code{\link{show_transectavespp}()},
\code{\link{show_transectmatrix}()},
\code{\link{show_transect}()},
\code{\link{show_wqmatrix}()}
}
\concept{visualize}
