% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\docType{data}
\name{targets}
\alias{targets}
\title{Bay segment targets}
\format{A data frame with 4 rows and 8 variables:
\describe{
  \item{bay_segment}{chr}
  \item{name}{chr}
  \item{chla_target}{num}
  \item{chla_smallex}{num}
  \item{chla_thresh}{num}
  \item{la_target}{num}
  \item{la_smallex}{num}
  \item{la_thresh}{num}
}}
\usage{
targets
}
\description{
Bay segment specific management targets including low and high magnitude exceedance thresholds
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{stations}},
  \code{\link{tbseg}}, \code{\link{tidalcreeks}}
}
\concept{utilities}
\keyword{datasets}
