% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbnispp.R
\docType{data}
\name{tbnispp}
\alias{tbnispp}
\title{Reference table for Tampa Bay Nekton Index species classifications}
\format{
A data frame with 196 rows and 9 variables:
\describe{
  \item{NODCCODE}{chr}
  \item{ScientificName}{chr}
  \item{Include_TB_Index}{chr}
  \item{Hab_Cat}{chr}
  \item{Est_Cat}{chr}
  \item{Est_Use}{chr}
  \item{Feeding_Cat}{chr}
  \item{Feeding_Guild}{chr}
  \item{Selected_Taxa}{chr}
}
}
\usage{
tbnispp
}
\description{
Reference table for Tampa Bay Nekton Index species classifications
}
\examples{
\dontrun{
library(dplyr)

# import and clean
tbnispp <- read.csv('../tbni-proc/data/TBIndex_spp_codes.csv',
    header = TRUE, stringsAsFactors = FALSE) \%>\%
  mutate(
    NODCCODE = as.character(NODCCODE),
    NODCCODE = case_when(NODCCODE == "9.998e+09" ~ "9998000000",
                             TRUE ~ NODCCODE)
  )

save(tbnispp, file = 'data/tbnispp.RData', compress = 'xz')
}
}
\seealso{
Other utilities: 
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{fimstations}},
\code{\link{iwrraw}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbseglines}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}}
}
\concept{utilities}
\keyword{datasets}
