% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbshed.R
\docType{data}
\name{tbshed}
\alias{tbshed}
\title{Spatial data object of Tampa Bay watershed}
\format{
A simple features \code{\link[sf]{sf}} object (POLYGON) with 1 feature and 0 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
}
\usage{
tbshed
}
\description{
Spatial data object of Tampa Bay watershed, includes the bay proper
}
\examples{
library(sf)
plot(st_geometry(tbshed))
}
\seealso{
Other utilities: 
\code{\link{benthicdata}},
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{fimstations}},
\code{\link{iwrraw}},
\code{\link{seagrass}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tidalcreeks}},
\code{\link{tidaltargets}},
\code{\link{transect}}
}
\concept{utilities}
\keyword{datasets}
