% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidalcreeks.R
\docType{data}
\name{tidalcreeks}
\alias{tidalcreeks}
\title{Spatial data object of tidal creeks}
\format{A simple features \code{\link[sf]{sf}} object (MULTILINESTRING) with 581 feaures and 4 fields, +proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs
\describe{
  \item{wbid}{chr}
  \item{JEI}{chr}
  \item{class}{chr}
  \item{Creek_Length_m}{num}
}}
\usage{
tidalcreeks
}
\description{
Spatial data object of tidal creeks
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{stations}},
  \code{\link{targets}}, \code{\link{tbseg}}
}
\concept{utilities}
\keyword{datasets}
