% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidaltargets.R
\docType{data}
\name{tidaltargets}
\alias{tidaltargets}
\title{Tidal creek nitrogen targets}
\format{A data frame with 2 rows and 4 variables:
\describe{
  \item{region}{chr}
  \item{threshold}{num}
  \item{action}{num}
  \item{caution}{num}
}}
\usage{
tidaltargets
}
\description{
Tidal creek nitrogen targets
}
\seealso{
Other utilities: \code{\link{bsmap}},
  \code{\link{epcdata}}, \code{\link{iwrraw}},
  \code{\link{stations}}, \code{\link{targets}},
  \code{\link{tbseg}}, \code{\link{tidalcreeks}}
}
\concept{utilities}
\keyword{datasets}
