% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidaltargets.R
\docType{data}
\name{tidaltargets}
\alias{tidaltargets}
\title{Tidal creek nitrogen targets}
\format{
A data frame with 2 rows and 4 variables:
\describe{
  \item{region}{chr}
  \item{prioritize}{num}
  \item{investigate}{num}
  \item{caution}{num}
}
}
\usage{
tidaltargets
}
\description{
Tidal creek nitrogen targets
}
\seealso{
Other utilities: 
\code{\link{benthicdata}},
\code{\link{bsmap}},
\code{\link{epcdata}},
\code{\link{fimdata}},
\code{\link{fimstations}},
\code{\link{iwrraw}},
\code{\link{seagrass}},
\code{\link{stations}},
\code{\link{targets}},
\code{\link{tbniref}},
\code{\link{tbnispp}},
\code{\link{tbseglines}},
\code{\link{tbsegshed}},
\code{\link{tbseg}},
\code{\link{tbshed}},
\code{\link{tidalcreeks}},
\code{\link{transect}}
}
\concept{utilities}
\keyword{datasets}
