% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_fibicons.R
\name{util_fibicons}
\alias{util_fibicons}
\title{Return leaflet icon set for FIB maps}
\usage{
util_fibicons(indic)
}
\arguments{
\item{indic}{character indicating \code{"entero"}, \code{"entero&ecoli"}, or \code{"fibmat"} for \emph{Enterococcus}, Fecal \emph{Enterococcus} and \emph{E. coli}, or FIB matrix maps, respectively}
}
\value{
A leaflet icon set as returned by \code{\link[leaflet]{iconList}}.
}
\description{
Return leaflet icon set for FIB maps
}
\details{
Used internally with \code{\link{show_enteromap}} for wet/dry icons for \emph{Enterococcus}, with \code{\link{show_fibmap}} for \emph{E. Coli}/\emph{Enterococcus} icons (EPCHC data), and with \code{\link{show_fibmatmap}} for matrix annual score category icons for EPCHC and non-EPCHC data.
}
\examples{
util_fibicons(indic = 'entero')
util_fibicons(indic = 'entero&ecoli')
util_fibicons(indic = 'fibmat')
}
