% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_importwqwin.R
\name{util_importwqwin}
\alias{util_importwqwin}
\title{Utility function to retrieve water quality data from the Florida Department of Environmental Protection's Watershed Information Network (WIN)}
\usage{
util_importwqwin(start_date, org_id, page_num)
}
\arguments{
\item{start_date}{character string for the start date in the format "YYYY-MM-DD"}

\item{org_id}{character string for the organization ID}

\item{page_num}{integer for the page number to retrieve}

\item{verbose}{logical indicating whether to print verbose output}
}
\value{
A list containing the results from the API request
}
\description{
Utility function to retrieve water quality data from the Florida Department of Environmental Protection's Watershed Information Network (WIN)
}
\details{
Used internally with \code{\link{read_importwqin}}.
}
\examples{
\dontrun{
onepg <- util_importwqwin("2025-01-01", "21FLMANA", 1)
head(onepg)
}
}
\concept{util}
