% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_orgin.R
\name{util_orgin}
\alias{util_orgin}
\title{Get organization name from organization identifier in USEPA Water Quality Portal}
\usage{
util_orgin(org, stanm = FALSE)
}
\arguments{
\item{org}{chr string indicating the organization identifier, see \code{\link{read_importwqp}} for valid entries}

\item{stanm}{logical indicating if a character string for a column name specific to the organization is returned}
}
\value{
A character string of the organization name that corresponds to the organization identifier or a column name for the station identifier specific to the organization if \code{stanm = TRUE}
}
\description{
Get organization name from organization identifier in USEPA Water Quality Portal
}
\examples{
util_orgin('21FLHILL_WQX')
util_orgin('21FLPASC_WQX', stanm = TRUE)
}
\concept{util}
